replaced previous version of abstract factory demo with new implementation
