--- docs/abstract-factory-2/blackpieces.dart
@@ -1,25 +1,0 @@
-import "abstractfactory.dart";
-
-class BlackPawn extends Pawn {
-  BlackPawn() : super("b") {}
-}
-
-class BlackKnight extends Knight {
-  BlackKnight() : super("b") {}
-}
-
-class BlackBishop extends Bishop {
-  BlackBishop() : super("b") {}
-}
-
-class BlackRook extends Rook {
-  BlackRook() : super("b") {}
-}
-
-class BlackKing extends King {
-  BlackKing() : super("b") {}
-}
-
-class BlackQueen extends Queen {
-  BlackQueen() : super("b") {}
-}
