--- docs/abstract-factory-2/script.js
@@ -1,5855 +1,0 @@
-// Generated by dart2js (NullSafetyMode.sound, csp, deferred-serialization, intern-composite-values), the Dart to JavaScript compiler version: 3.0.7.
-// The code supports the following hooks:
-// dartPrint(message):
-//    if this function is defined it is called instead of the Dart [print]
-//    method.
-//
-// dartMainRunner(main, args):
-//    if this function is defined, the Dart [main] method will not be invoked
-//    directly. Instead, a closure that will invoke [main], and its arguments
-//    [args] is passed to [dartMainRunner].
-//
-// dartDeferredLibraryLoader(uri, successCallback, errorCallback, loadId, loadPriority):
-//    if this function is defined, it will be called when a deferred library
-//    is loaded. It should load and eval the javascript of `uri`, and call
-//    successCallback. If it fails to do so, it should call errorCallback with
-//    an error. The loadId argument is the deferred import that resulted in
-//    this uri being loaded. The loadPriority argument is the priority the
-//    library should be loaded with as specified in the code via the
-//    load-priority annotation (0: normal, 1: high).
-//
-// dartCallInstrumentation(id, qualifiedName):
-//    if this function is defined, it will be called at each entry of a
-//    method or constructor. Used only when compiling programs with
-//    --experiment-call-instrumentation.
-(function dartProgram() {
-  function copyProperties(from, to) {
-    var keys = Object.keys(from);
-    for (var i = 0; i < keys.length; i++) {
-      var key = keys[i];
-      to[key] = from[key];
-    }
-  }
-  function mixinPropertiesHard(from, to) {
-    var keys = Object.keys(from);
-    for (var i = 0; i < keys.length; i++) {
-      var key = keys[i];
-      if (!to.hasOwnProperty(key))
-        to[key] = from[key];
-    }
-  }
-  function mixinPropertiesEasy(from, to) {
-    Object.assign(to, from);
-  }
-  var supportsDirectProtoAccess = function() {
-    var cls = function() {
-    };
-    cls.prototype = {p: {}};
-    var object = new cls();
-    if (!(Object.getPrototypeOf(object) && Object.getPrototypeOf(object).p === cls.prototype.p))
-      return false;
-    try {
-      if (typeof navigator != "undefined" && typeof navigator.userAgent == "string" && navigator.userAgent.indexOf("Chrome/") >= 0)
-        return true;
-      if (typeof version == "function" && version.length == 0) {
-        var v = version();
-        if (/^\d+\.\d+\.\d+\.\d+$/.test(v))
-          return true;
-      }
-    } catch (_) {
-    }
-    return false;
-  }();
-  function inherit(cls, sup) {
-    cls.prototype.constructor = cls;
-    cls.prototype["$is" + cls.name] = cls;
-    if (sup != null) {
-      if (supportsDirectProtoAccess) {
-        Object.setPrototypeOf(cls.prototype, sup.prototype);
-        return;
-      }
-      var clsPrototype = Object.create(sup.prototype);
-      copyProperties(cls.prototype, clsPrototype);
-      cls.prototype = clsPrototype;
-    }
-  }
-  function inheritMany(sup, classes) {
-    for (var i = 0; i < classes.length; i++)
-      inherit(classes[i], sup);
-  }
-  function mixinEasy(cls, mixin) {
-    mixinPropertiesEasy(mixin.prototype, cls.prototype);
-    cls.prototype.constructor = cls;
-  }
-  function mixinHard(cls, mixin) {
-    mixinPropertiesHard(mixin.prototype, cls.prototype);
-    cls.prototype.constructor = cls;
-  }
-  function lazyOld(holder, name, getterName, initializer) {
-    var uninitializedSentinel = holder;
-    holder[name] = uninitializedSentinel;
-    holder[getterName] = function() {
-      holder[getterName] = function() {
-        A.throwCyclicInit(name);
-      };
-      var result;
-      var sentinelInProgress = initializer;
-      try {
-        if (holder[name] === uninitializedSentinel) {
-          result = holder[name] = sentinelInProgress;
-          result = holder[name] = initializer();
-        } else
-          result = holder[name];
-      } finally {
-        if (result === sentinelInProgress)
-          holder[name] = null;
-        holder[getterName] = function() {
-          return this[name];
-        };
-      }
-      return result;
-    };
-  }
-  function lazy(holder, name, getterName, initializer) {
-    var uninitializedSentinel = holder;
-    holder[name] = uninitializedSentinel;
-    holder[getterName] = function() {
-      if (holder[name] === uninitializedSentinel)
-        holder[name] = initializer();
-      holder[getterName] = function() {
-        return this[name];
-      };
-      return holder[name];
-    };
-  }
-  function lazyFinal(holder, name, getterName, initializer) {
-    var uninitializedSentinel = holder;
-    holder[name] = uninitializedSentinel;
-    holder[getterName] = function() {
-      if (holder[name] === uninitializedSentinel) {
-        var value = initializer();
-        if (holder[name] !== uninitializedSentinel)
-          A.throwLateFieldADI(name);
-        holder[name] = value;
-      }
-      var finalValue = holder[name];
-      holder[getterName] = function() {
-        return finalValue;
-      };
-      return finalValue;
-    };
-  }
-  function makeConstList(list) {
-    list.immutable$list = Array;
-    list.fixed$length = Array;
-    return list;
-  }
-  function convertToFastObject(properties) {
-    function t() {
-    }
-    t.prototype = properties;
-    new t();
-    return properties;
-  }
-  function convertAllToFastObject(arrayOfObjects) {
-    for (var i = 0; i < arrayOfObjects.length; ++i)
-      convertToFastObject(arrayOfObjects[i]);
-  }
-  var functionCounter = 0;
-  function instanceTearOffGetter(isIntercepted, parameters) {
-    var cache = null;
-    return isIntercepted ? function(receiver) {
-      if (cache === null)
-        cache = A.closureFromTearOff(parameters);
-      return new cache(receiver, this);
-    } : function() {
-      if (cache === null)
-        cache = A.closureFromTearOff(parameters);
-      return new cache(this, null);
-    };
-  }
-  function staticTearOffGetter(parameters) {
-    var cache = null;
-    return function() {
-      if (cache === null)
-        cache = A.closureFromTearOff(parameters).prototype;
-      return cache;
-    };
-  }
-  var typesOffset = 0;
-  function tearOffParameters(container, isStatic, isIntercepted, requiredParameterCount, optionalParameterDefaultValues, callNames, funsOrNames, funType, applyIndex, needsDirectAccess) {
-    if (typeof funType == "number")
-      funType += typesOffset;
-    return {co: container, iS: isStatic, iI: isIntercepted, rC: requiredParameterCount, dV: optionalParameterDefaultValues, cs: callNames, fs: funsOrNames, fT: funType, aI: applyIndex || 0, nDA: needsDirectAccess};
-  }
-  function installStaticTearOff(holder, getterName, requiredParameterCount, optionalParameterDefaultValues, callNames, funsOrNames, funType, applyIndex) {
-    var parameters = tearOffParameters(holder, true, false, requiredParameterCount, optionalParameterDefaultValues, callNames, funsOrNames, funType, applyIndex, false);
-    var getterFunction = staticTearOffGetter(parameters);
-    holder[getterName] = getterFunction;
-  }
-  function installInstanceTearOff(prototype, getterName, isIntercepted, requiredParameterCount, optionalParameterDefaultValues, callNames, funsOrNames, funType, applyIndex, needsDirectAccess) {
-    isIntercepted = !!isIntercepted;
-    var parameters = tearOffParameters(prototype, false, isIntercepted, requiredParameterCount, optionalParameterDefaultValues, callNames, funsOrNames, funType, applyIndex, !!needsDirectAccess);
-    var getterFunction = instanceTearOffGetter(isIntercepted, parameters);
-    prototype[getterName] = getterFunction;
-  }
-  function setOrUpdateInterceptorsByTag(newTags) {
-    var tags = init.interceptorsByTag;
-    if (!tags) {
-      init.interceptorsByTag = newTags;
-      return;
-    }
-    copyProperties(newTags, tags);
-  }
-  function setOrUpdateLeafTags(newTags) {
-    var tags = init.leafTags;
-    if (!tags) {
-      init.leafTags = newTags;
-      return;
-    }
-    copyProperties(newTags, tags);
-  }
-  function updateTypes(newTypes) {
-    var types = init.types;
-    var length = types.length;
-    types.push.apply(types, newTypes);
-    return length;
-  }
-  function updateHolder(holder, newHolder) {
-    copyProperties(newHolder, holder);
-    return holder;
-  }
-  var hunkHelpers = function() {
-    var mkInstance = function(isIntercepted, requiredParameterCount, optionalParameterDefaultValues, callNames, applyIndex) {
-        return function(container, getterName, name, funType) {
-          return installInstanceTearOff(container, getterName, isIntercepted, requiredParameterCount, optionalParameterDefaultValues, callNames, [name], funType, applyIndex, false);
-        };
-      },
-      mkStatic = function(requiredParameterCount, optionalParameterDefaultValues, callNames, applyIndex) {
-        return function(container, getterName, name, funType) {
-          return installStaticTearOff(container, getterName, requiredParameterCount, optionalParameterDefaultValues, callNames, [name], funType, applyIndex);
-        };
-      };
-    return {inherit: inherit, inheritMany: inheritMany, mixin: mixinEasy, mixinHard: mixinHard, installStaticTearOff: installStaticTearOff, installInstanceTearOff: installInstanceTearOff, _instance_0u: mkInstance(0, 0, null, ["call$0"], 0), _instance_1u: mkInstance(0, 1, null, ["call$1"], 0), _instance_2u: mkInstance(0, 2, null, ["call$2"], 0), _instance_0i: mkInstance(1, 0, null, ["call$0"], 0), _instance_1i: mkInstance(1, 1, null, ["call$1"], 0), _instance_2i: mkInstance(1, 2, null, ["call$2"], 0), _static_0: mkStatic(0, null, ["call$0"], 0), _static_1: mkStatic(1, null, ["call$1"], 0), _static_2: mkStatic(2, null, ["call$2"], 0), makeConstList: makeConstList, lazy: lazy, lazyFinal: lazyFinal, lazyOld: lazyOld, updateHolder: updateHolder, convertToFastObject: convertToFastObject, updateTypes: updateTypes, setOrUpdateInterceptorsByTag: setOrUpdateInterceptorsByTag, setOrUpdateLeafTags: setOrUpdateLeafTags};
-  }();
-  function initializeDeferredHunk(hunk) {
-    typesOffset = init.types.length;
-    hunk(hunkHelpers, init, holders, $);
-  }
-  var A = {JS_CONST: function JS_CONST() {
-    },
-    LateError$fieldNI(fieldName) {
-      return new A.LateError("Field '" + fieldName + "' has not been initialized.");
-    },
-    isToStringVisiting(object) {
-      var t1, i;
-      for (t1 = $.toStringVisiting.length, i = 0; i < t1; ++i)
-        if (object === $.toStringVisiting[i])
-          return true;
-      return false;
-    },
-    LateError: function LateError(t0) {
-      this._message = t0;
-    },
-    ListIterator: function ListIterator(t0, t1, t2) {
-      var _ = this;
-      _.__internal$_iterable = t0;
-      _.__internal$_length = t1;
-      _.__internal$_index = 0;
-      _.__internal$_current = null;
-      _.$ti = t2;
-    },
-    MappedIterable: function MappedIterable(t0, t1, t2) {
-      this.__internal$_iterable = t0;
-      this._f = t1;
-      this.$ti = t2;
-    },
-    MappedIterator: function MappedIterator(t0, t1, t2) {
-      var _ = this;
-      _.__internal$_current = null;
-      _._iterator = t0;
-      _._f = t1;
-      _.$ti = t2;
-    },
-    WhereIterable: function WhereIterable(t0, t1, t2) {
-      this.__internal$_iterable = t0;
-      this._f = t1;
-      this.$ti = t2;
-    },
-    WhereIterator: function WhereIterator(t0, t1, t2) {
-      this._iterator = t0;
-      this._f = t1;
-      this.$ti = t2;
-    },
-    unminifyOrTag(rawClassName) {
-      var preserved = init.mangledGlobalNames[rawClassName];
-      if (preserved != null)
-        return preserved;
-      return rawClassName;
-    },
-    isJsIndexable(object, record) {
-      var result;
-      if (record != null) {
-        result = record.x;
-        if (result != null)
-          return result;
-      }
-      return type$.JavaScriptIndexingBehavior_dynamic._is(object);
-    },
-    S(value) {
-      var result;
-      if (typeof value == "string")
-        return value;
-      if (typeof value == "number") {
-        if (value !== 0)
-          return "" + value;
-      } else if (true === value)
-        return "true";
-      else if (false === value)
-        return "false";
-      else if (value == null)
-        return "null";
-      result = J.toString$0$(value);
-      return result;
-    },
-    Primitives_objectHashCode(object) {
-      var hash,
-        property = $.Primitives__identityHashCodeProperty;
-      if (property == null)
-        property = $.Primitives__identityHashCodeProperty = Symbol("identityHashCode");
-      hash = object[property];
-      if (hash == null) {
-        hash = Math.random() * 0x3fffffff | 0;
-        object[property] = hash;
-      }
-      return hash;
-    },
-    Primitives_objectTypeName(object) {
-      return A.Primitives__objectTypeNameNewRti(object);
-    },
-    Primitives__objectTypeNameNewRti(object) {
-      var interceptor, dispatchName, $constructor, constructorName;
-      if (object instanceof A.Object)
-        return A._rtiToString(A.instanceType(object), null);
-      interceptor = J.getInterceptor$(object);
-      if (interceptor === B.Interceptor_methods || interceptor === B.JavaScriptObject_methods || type$.UnknownJavaScriptObject._is(object)) {
-        dispatchName = B.C_JS_CONST(object);
-        if (dispatchName !== "Object" && dispatchName !== "")
-          return dispatchName;
-        $constructor = object.constructor;
-        if (typeof $constructor == "function") {
-          constructorName = $constructor.name;
-          if (typeof constructorName == "string" && constructorName !== "Object" && constructorName !== "")
-            return constructorName;
-        }
-      }
-      return A._rtiToString(A.instanceType(object), null);
-    },
-    Primitives_safeToString(object) {
-      if (typeof object == "number" || A._isBool(object))
-        return J.toString$0$(object);
-      if (typeof object == "string")
-        return JSON.stringify(object);
-      if (object instanceof A.Closure)
-        return object.toString$0(0);
-      return "Instance of '" + A.Primitives_objectTypeName(object) + "'";
-    },
-    iae(argument) {
-      throw A.wrapException(A.argumentErrorValue(argument));
-    },
-    ioore(receiver, index) {
-      if (receiver == null)
-        J.get$length$asx(receiver);
-      throw A.wrapException(A.diagnoseIndexError(receiver, index));
-    },
-    diagnoseIndexError(indexable, index) {
-      var $length, _s5_ = "index", _null = null;
-      if (!A._isInt(index))
-        return new A.ArgumentError(true, index, _s5_, _null);
-      $length = A._asInt(J.get$length$asx(indexable));
-      if (index < 0 || index >= $length)
-        return A.IndexError$withLength(index, $length, indexable, _null, _s5_);
-      return new A.RangeError(_null, _null, true, index, _s5_, "Value not in range");
-    },
-    argumentErrorValue(object) {
-      return new A.ArgumentError(true, object, null, null);
-    },
-    wrapException(ex) {
-      var wrapper, t1;
-      if (ex == null)
-        ex = new A.TypeError();
-      wrapper = new Error();
-      wrapper.dartException = ex;
-      t1 = A.toStringWrapper;
-      if ("defineProperty" in Object) {
-        Object.defineProperty(wrapper, "message", {get: t1});
-        wrapper.name = "";
-      } else
-        wrapper.toString = t1;
-      return wrapper;
-    },
-    toStringWrapper() {
-      return J.toString$0$(this.dartException);
-    },
-    throwExpression(ex) {
-      throw A.wrapException(ex);
-    },
-    throwConcurrentModificationError(collection) {
-      throw A.wrapException(A.ConcurrentModificationError$(collection));
-    },
-    objectHashCode(object) {
-      if (object == null || typeof object != "object")
-        return J.get$hashCode$(object);
-      else
-        return A.Primitives_objectHashCode(object);
-    },
-    fillLiteralSet(values, result) {
-      var index,
-        $length = values.length;
-      for (index = 0; index < $length; ++index)
-        result.add$1(0, values[index]);
-      return result;
-    },
-    invokeClosure(closure, numberOfArguments, arg1, arg2, arg3, arg4) {
-      type$.Function._as(closure);
-      switch (A._asInt(numberOfArguments)) {
-        case 0:
-          return closure.call$0();
-        case 1:
-          return closure.call$1(arg1);
-        case 2:
-          return closure.call$2(arg1, arg2);
-        case 3:
-          return closure.call$3(arg1, arg2, arg3);
-        case 4:
-          return closure.call$4(arg1, arg2, arg3, arg4);
-      }
-      throw A.wrapException(new A._Exception("Unsupported number of arguments for wrapped closure"));
-    },
-    convertDartClosureToJS(closure, arity) {
-      var $function;
-      if (closure == null)
-        return null;
-      $function = closure.$identity;
-      if (!!$function)
-        return $function;
-      $function = function(closure, arity, invoke) {
-        return function(a1, a2, a3, a4) {
-          return invoke(closure, arity, a1, a2, a3, a4);
-        };
-      }(closure, arity, A.invokeClosure);
-      closure.$identity = $function;
-      return $function;
-    },
-    Closure_fromTearOff(parameters) {
-      var $prototype, $constructor, t2, trampoline, applyTrampoline, i, stub, stub0, stubName, stubCallName,
-        container = parameters.co,
-        isStatic = parameters.iS,
-        isIntercepted = parameters.iI,
-        needsDirectAccess = parameters.nDA,
-        applyTrampolineIndex = parameters.aI,
-        funsOrNames = parameters.fs,
-        callNames = parameters.cs,
-        $name = funsOrNames[0],
-        callName = callNames[0],
-        $function = container[$name],
-        t1 = parameters.fT;
-      t1.toString;
-      $prototype = isStatic ? Object.create(new A.StaticClosure().constructor.prototype) : Object.create(new A.BoundClosure(null, null).constructor.prototype);
-      $prototype.$initialize = $prototype.constructor;
-      if (isStatic)
-        $constructor = function static_tear_off() {
-          this.$initialize();
-        };
-      else
-        $constructor = function tear_off(a, b) {
-          this.$initialize(a, b);
-        };
-      $prototype.constructor = $constructor;
-      $constructor.prototype = $prototype;
-      $prototype.$_name = $name;
-      $prototype.$_target = $function;
-      t2 = !isStatic;
-      if (t2)
-        trampoline = A.Closure_forwardCallTo($name, $function, isIntercepted, needsDirectAccess);
-      else {
-        $prototype.$static_name = $name;
-        trampoline = $function;
-      }
-      $prototype.$signature = A.Closure__computeSignatureFunctionNewRti(t1, isStatic, isIntercepted);
-      $prototype[callName] = trampoline;
-      for (applyTrampoline = trampoline, i = 1; i < funsOrNames.length; ++i) {
-        stub = funsOrNames[i];
-        if (typeof stub == "string") {
-          stub0 = container[stub];
-          stubName = stub;
-          stub = stub0;
-        } else
-          stubName = "";
-        stubCallName = callNames[i];
-        if (stubCallName != null) {
-          if (t2)
-            stub = A.Closure_forwardCallTo(stubName, stub, isIntercepted, needsDirectAccess);
-          $prototype[stubCallName] = stub;
-        }
-        if (i === applyTrampolineIndex)
-          applyTrampoline = stub;
-      }
-      $prototype["call*"] = applyTrampoline;
-      $prototype.$requiredArgCount = parameters.rC;
-      $prototype.$defaultValues = parameters.dV;
-      return $constructor;
-    },
-    Closure__computeSignatureFunctionNewRti(functionType, isStatic, isIntercepted) {
-      if (typeof functionType == "number")
-        return functionType;
-      if (typeof functionType == "string") {
-        if (isStatic)
-          throw A.wrapException("Cannot compute signature for static tearoff.");
-        return function(recipe, evalOnReceiver) {
-          return function() {
-            return evalOnReceiver(this, recipe);
-          };
-        }(functionType, A.BoundClosure_evalRecipe);
-      }
-      throw A.wrapException("Error in functionType of tearoff");
-    },
-    Closure_cspForwardCall(arity, needsDirectAccess, stubName, $function) {
-      var getReceiver = A.BoundClosure_receiverOf;
-      switch (needsDirectAccess ? -1 : arity) {
-        case 0:
-          return function(entry, receiverOf) {
-            return function() {
-              return receiverOf(this)[entry]();
-            };
-          }(stubName, getReceiver);
-        case 1:
-          return function(entry, receiverOf) {
-            return function(a) {
-              return receiverOf(this)[entry](a);
-            };
-          }(stubName, getReceiver);
-        case 2:
-          return function(entry, receiverOf) {
-            return function(a, b) {
-              return receiverOf(this)[entry](a, b);
-            };
-          }(stubName, getReceiver);
-        case 3:
-          return function(entry, receiverOf) {
-            return function(a, b, c) {
-              return receiverOf(this)[entry](a, b, c);
-            };
-          }(stubName, getReceiver);
-        case 4:
-          return function(entry, receiverOf) {
-            return function(a, b, c, d) {
-              return receiverOf(this)[entry](a, b, c, d);
-            };
-          }(stubName, getReceiver);
-        case 5:
-          return function(entry, receiverOf) {
-            return function(a, b, c, d, e) {
-              return receiverOf(this)[entry](a, b, c, d, e);
-            };
-          }(stubName, getReceiver);
-        default:
-          return function(f, receiverOf) {
-            return function() {
-              return f.apply(receiverOf(this), arguments);
-            };
-          }($function, getReceiver);
-      }
-    },
-    Closure_forwardCallTo(stubName, $function, isIntercepted, needsDirectAccess) {
-      var arity, t1;
-      if (isIntercepted)
-        return A.Closure_forwardInterceptedCallTo(stubName, $function, needsDirectAccess);
-      arity = $function.length;
-      t1 = A.Closure_cspForwardCall(arity, needsDirectAccess, stubName, $function);
-      return t1;
-    },
-    Closure_cspForwardInterceptedCall(arity, needsDirectAccess, stubName, $function) {
-      var getReceiver = A.BoundClosure_receiverOf,
-        getInterceptor = A.BoundClosure_interceptorOf;
-      switch (needsDirectAccess ? -1 : arity) {
-        case 0:
-          throw A.wrapException(new A.RuntimeError("Intercepted function with no arguments."));
-        case 1:
-          return function(entry, interceptorOf, receiverOf) {
-            return function() {
-              return interceptorOf(this)[entry](receiverOf(this));
-            };
-          }(stubName, getInterceptor, getReceiver);
-        case 2:
-          return function(entry, interceptorOf, receiverOf) {
-            return function(a) {
-              return interceptorOf(this)[entry](receiverOf(this), a);
-            };
-          }(stubName, getInterceptor, getReceiver);
-        case 3:
-          return function(entry, interceptorOf, receiverOf) {
-            return function(a, b) {
-              return interceptorOf(this)[entry](receiverOf(this), a, b);
-            };
-          }(stubName, getInterceptor, getReceiver);
-        case 4:
-          return function(entry, interceptorOf, receiverOf) {
-            return function(a, b, c) {
-              return interceptorOf(this)[entry](receiverOf(this), a, b, c);
-            };
-          }(stubName, getInterceptor, getReceiver);
-        case 5:
-          return function(entry, interceptorOf, receiverOf) {
-            return function(a, b, c, d) {
-              return interceptorOf(this)[entry](receiverOf(this), a, b, c, d);
-            };
-          }(stubName, getInterceptor, getReceiver);
-        case 6:
-          return function(entry, interceptorOf, receiverOf) {
-            return function(a, b, c, d, e) {
-              return interceptorOf(this)[entry](receiverOf(this), a, b, c, d, e);
-            };
-          }(stubName, getInterceptor, getReceiver);
-        default:
-          return function(f, interceptorOf, receiverOf) {
-            return function() {
-              var a = [receiverOf(this)];
-              Array.prototype.push.apply(a, arguments);
-              return f.apply(interceptorOf(this), a);
-            };
-          }($function, getInterceptor, getReceiver);
-      }
-    },
-    Closure_forwardInterceptedCallTo(stubName, $function, needsDirectAccess) {
-      var arity, t1;
-      if ($.BoundClosure__interceptorFieldNameCache == null)
-        $.BoundClosure__interceptorFieldNameCache = A.BoundClosure__computeFieldNamed("interceptor");
-      if ($.BoundClosure__receiverFieldNameCache == null)
-        $.BoundClosure__receiverFieldNameCache = A.BoundClosure__computeFieldNamed("receiver");
-      arity = $function.length;
-      t1 = A.Closure_cspForwardInterceptedCall(arity, needsDirectAccess, stubName, $function);
-      return t1;
-    },
-    closureFromTearOff(parameters) {
-      return A.Closure_fromTearOff(parameters);
-    },
-    BoundClosure_evalRecipe(closure, recipe) {
-      return A._Universe_evalInEnvironment(init.typeUniverse, A.instanceType(closure._receiver), recipe);
-    },
-    BoundClosure_receiverOf(closure) {
-      return closure._receiver;
-    },
-    BoundClosure_interceptorOf(closure) {
-      return closure._interceptor;
-    },
-    BoundClosure__computeFieldNamed(fieldName) {
-      var t1, i, $name,
-        template = new A.BoundClosure("receiver", "interceptor"),
-        names = J.JSArray_markFixedList(Object.getOwnPropertyNames(template), type$.nullable_Object);
-      for (t1 = names.length, i = 0; i < t1; ++i) {
-        $name = names[i];
-        if (template[$name] === fieldName)
-          return $name;
-      }
-      throw A.wrapException(A.ArgumentError$("Field name " + fieldName + " not found."));
-    },
-    boolConversionCheck(value) {
-      if (value == null)
-        A.assertThrow("boolean expression must not be null");
-      return value;
-    },
-    assertThrow(message) {
-      throw A.wrapException(new A._AssertionError(message));
-    },
-    throwCyclicInit(staticName) {
-      throw A.wrapException(new A._CyclicInitializationError(staticName));
-    },
-    getIsolateAffinityTag($name) {
-      return init.getIsolateTag($name);
-    },
-    defineProperty(obj, property, value) {
-      Object.defineProperty(obj, property, {value: value, enumerable: false, writable: true, configurable: true});
-    },
-    lookupAndCacheInterceptor(obj) {
-      var interceptor, interceptorClass, altTag, mark, t1,
-        tag = A._asString($.getTagFunction.call$1(obj)),
-        record = $.dispatchRecordsForInstanceTags[tag];
-      if (record != null) {
-        Object.defineProperty(obj, init.dispatchPropertyName, {value: record, enumerable: false, writable: true, configurable: true});
-        return record.i;
-      }
-      interceptor = $.interceptorsForUncacheableTags[tag];
-      if (interceptor != null)
-        return interceptor;
-      interceptorClass = init.interceptorsByTag[tag];
-      if (interceptorClass == null) {
-        altTag = A._asStringQ($.alternateTagFunction.call$2(obj, tag));
-        if (altTag != null) {
-          record = $.dispatchRecordsForInstanceTags[altTag];
-          if (record != null) {
-            Object.defineProperty(obj, init.dispatchPropertyName, {value: record, enumerable: false, writable: true, configurable: true});
-            return record.i;
-          }
-          interceptor = $.interceptorsForUncacheableTags[altTag];
-          if (interceptor != null)
-            return interceptor;
-          interceptorClass = init.interceptorsByTag[altTag];
-          tag = altTag;
-        }
-      }
-      if (interceptorClass == null)
-        return null;
-      interceptor = interceptorClass.prototype;
-      mark = tag[0];
-      if (mark === "!") {
-        record = A.makeLeafDispatchRecord(interceptor);
-        $.dispatchRecordsForInstanceTags[tag] = record;
-        Object.defineProperty(obj, init.dispatchPropertyName, {value: record, enumerable: false, writable: true, configurable: true});
-        return record.i;
-      }
-      if (mark === "~") {
-        $.interceptorsForUncacheableTags[tag] = interceptor;
-        return interceptor;
-      }
-      if (mark === "-") {
-        t1 = A.makeLeafDispatchRecord(interceptor);
-        Object.defineProperty(Object.getPrototypeOf(obj), init.dispatchPropertyName, {value: t1, enumerable: false, writable: true, configurable: true});
-        return t1.i;
-      }
-      if (mark === "+")
-        return A.patchInteriorProto(obj, interceptor);
-      if (mark === "*")
-        throw A.wrapException(A.UnimplementedError$(tag));
-      if (init.leafTags[tag] === true) {
-        t1 = A.makeLeafDispatchRecord(interceptor);
-        Object.defineProperty(Object.getPrototypeOf(obj), init.dispatchPropertyName, {value: t1, enumerable: false, writable: true, configurable: true});
-        return t1.i;
-      } else
-        return A.patchInteriorProto(obj, interceptor);
-    },
-    patchInteriorProto(obj, interceptor) {
-      var proto = Object.getPrototypeOf(obj);
-      Object.defineProperty(proto, init.dispatchPropertyName, {value: J.makeDispatchRecord(interceptor, proto, null, null), enumerable: false, writable: true, configurable: true});
-      return interceptor;
-    },
-    makeLeafDispatchRecord(interceptor) {
-      return J.makeDispatchRecord(interceptor, false, null, !!interceptor.$isJavaScriptIndexingBehavior);
-    },
-    makeDefaultDispatchRecord(tag, interceptorClass, proto) {
-      var interceptor = interceptorClass.prototype;
-      if (init.leafTags[tag] === true)
-        return A.makeLeafDispatchRecord(interceptor);
-      else
-        return J.makeDispatchRecord(interceptor, proto, null, null);
-    },
-    initNativeDispatch() {
-      if (true === $.initNativeDispatchFlag)
-        return;
-      $.initNativeDispatchFlag = true;
-      A.initNativeDispatchContinue();
-    },
-    initNativeDispatchContinue() {
-      var map, tags, fun, i, tag, proto, record, interceptorClass;
-      $.dispatchRecordsForInstanceTags = Object.create(null);
-      $.interceptorsForUncacheableTags = Object.create(null);
-      A.initHooks();
-      map = init.interceptorsByTag;
-      tags = Object.getOwnPropertyNames(map);
-      if (typeof window != "undefined") {
-        window;
-        fun = function() {
-        };
-        for (i = 0; i < tags.length; ++i) {
-          tag = tags[i];
-          proto = $.prototypeForTagFunction.call$1(tag);
-          if (proto != null) {
-            record = A.makeDefaultDispatchRecord(tag, map[tag], proto);
-            if (record != null) {
-              Object.defineProperty(proto, init.dispatchPropertyName, {value: record, enumerable: false, writable: true, configurable: true});
-              fun.prototype = proto;
-            }
-          }
-        }
-      }
-      for (i = 0; i < tags.length; ++i) {
-        tag = tags[i];
-        if (/^[A-Za-z_]/.test(tag)) {
-          interceptorClass = map[tag];
-          map["!" + tag] = interceptorClass;
-          map["~" + tag] = interceptorClass;
-          map["-" + tag] = interceptorClass;
-          map["+" + tag] = interceptorClass;
-          map["*" + tag] = interceptorClass;
-        }
-      }
-    },
-    initHooks() {
-      var transformers, i, transformer, getTag, getUnknownTag, prototypeForTag,
-        hooks = B.C_JS_CONST0();
-      hooks = A.applyHooksTransformer(B.C_JS_CONST1, A.applyHooksTransformer(B.C_JS_CONST2, A.applyHooksTransformer(B.C_JS_CONST3, A.applyHooksTransformer(B.C_JS_CONST3, A.applyHooksTransformer(B.C_JS_CONST4, A.applyHooksTransformer(B.C_JS_CONST5, A.applyHooksTransformer(B.C_JS_CONST6(B.C_JS_CONST), hooks)))))));
-      if (typeof dartNativeDispatchHooksTransformer != "undefined") {
-        transformers = dartNativeDispatchHooksTransformer;
-        if (typeof transformers == "function")
-          transformers = [transformers];
-        if (transformers.constructor == Array)
-          for (i = 0; i < transformers.length; ++i) {
-            transformer = transformers[i];
-            if (typeof transformer == "function")
-              hooks = transformer(hooks) || hooks;
-          }
-      }
-      getTag = hooks.getTag;
-      getUnknownTag = hooks.getUnknownTag;
-      prototypeForTag = hooks.prototypeForTag;
-      $.getTagFunction = new A.initHooks_closure(getTag);
-      $.alternateTagFunction = new A.initHooks_closure0(getUnknownTag);
-      $.prototypeForTagFunction = new A.initHooks_closure1(prototypeForTag);
-    },
-    applyHooksTransformer(transformer, hooks) {
-      return transformer(hooks) || hooks;
-    },
-    createRecordTypePredicate(shape, fieldRtis) {
-      var $length = fieldRtis.length,
-        $function = init.rttc["" + $length + ";" + shape];
-      if ($function == null)
-        return null;
-      if ($length === 0)
-        return $function;
-      if ($length === $function.length)
-        return $function.apply(null, fieldRtis);
-      return $function(fieldRtis);
-    },
-    JSSyntaxRegExp_makeNative(source, multiLine, caseSensitive, unicode, dotAll, global) {
-      var m = multiLine ? "m" : "",
-        i = caseSensitive ? "" : "i",
-        u = unicode ? "u" : "",
-        s = dotAll ? "s" : "",
-        g = global ? "g" : "",
-        regexp = function(source, modifiers) {
-          try {
-            return new RegExp(source, modifiers);
-          } catch (e) {
-            return e;
-          }
-        }(source, m + i + u + s + g);
-      if (regexp instanceof RegExp)
-        return regexp;
-      throw A.wrapException(new A.FormatException("Illegal RegExp pattern (" + String(regexp) + ")", source));
-    },
-    Closure: function Closure() {
-    },
-    Closure2Args: function Closure2Args() {
-    },
-    TearOffClosure: function TearOffClosure() {
-    },
-    StaticClosure: function StaticClosure() {
-    },
-    BoundClosure: function BoundClosure(t0, t1) {
-      this._receiver = t0;
-      this._interceptor = t1;
-    },
-    _CyclicInitializationError: function _CyclicInitializationError(t0) {
-      this.variableName = t0;
-    },
-    RuntimeError: function RuntimeError(t0) {
-      this.message = t0;
-    },
-    _AssertionError: function _AssertionError(t0) {
-      this.message = t0;
-    },
-    initHooks_closure: function initHooks_closure(t0) {
-      this.getTag = t0;
-    },
-    initHooks_closure0: function initHooks_closure0(t0) {
-      this.getUnknownTag = t0;
-    },
-    initHooks_closure1: function initHooks_closure1(t0) {
-      this.prototypeForTag = t0;
-    },
-    JSSyntaxRegExp: function JSSyntaxRegExp(t0, t1) {
-      var _ = this;
-      _.pattern = t0;
-      _._nativeRegExp = t1;
-      _._nativeAnchoredRegExp = _._nativeGlobalRegExp = null;
-    },
-    Rti__getQuestionFromStar(universe, rti) {
-      var question = rti._precomputed1;
-      return question == null ? rti._precomputed1 = A._Universe__lookupQuestionRti(universe, rti._primary, true) : question;
-    },
-    Rti__getFutureFromFutureOr(universe, rti) {
-      var future = rti._precomputed1;
-      return future == null ? rti._precomputed1 = A._Universe__lookupInterfaceRti(universe, "Future", [rti._primary]) : future;
-    },
-    Rti__isUnionOfFunctionType(rti) {
-      var kind = rti._kind;
-      if (kind === 6 || kind === 7 || kind === 8)
-        return A.Rti__isUnionOfFunctionType(rti._primary);
-      return kind === 12 || kind === 13;
-    },
-    Rti__getCanonicalRecipe(rti) {
-      return rti._canonicalRecipe;
-    },
-    findType(recipe) {
-      return A._Universe_eval(init.typeUniverse, recipe, false);
-    },
-    _substitute(universe, rti, typeArguments, depth) {
-      var baseType, substitutedBaseType, interfaceTypeArguments, substitutedInterfaceTypeArguments, base, substitutedBase, $arguments, substitutedArguments, returnType, substitutedReturnType, functionParameters, substitutedFunctionParameters, bounds, substitutedBounds, index, argument,
-        kind = rti._kind;
-      switch (kind) {
-        case 5:
-        case 1:
-        case 2:
-        case 3:
-        case 4:
-          return rti;
-        case 6:
-          baseType = rti._primary;
-          substitutedBaseType = A._substitute(universe, baseType, typeArguments, depth);
-          if (substitutedBaseType === baseType)
-            return rti;
-          return A._Universe__lookupStarRti(universe, substitutedBaseType, true);
-        case 7:
-          baseType = rti._primary;
-          substitutedBaseType = A._substitute(universe, baseType, typeArguments, depth);
-          if (substitutedBaseType === baseType)
-            return rti;
-          return A._Universe__lookupQuestionRti(universe, substitutedBaseType, true);
-        case 8:
-          baseType = rti._primary;
-          substitutedBaseType = A._substitute(universe, baseType, typeArguments, depth);
-          if (substitutedBaseType === baseType)
-            return rti;
-          return A._Universe__lookupFutureOrRti(universe, substitutedBaseType, true);
-        case 9:
-          interfaceTypeArguments = rti._rest;
-          substitutedInterfaceTypeArguments = A._substituteArray(universe, interfaceTypeArguments, typeArguments, depth);
-          if (substitutedInterfaceTypeArguments === interfaceTypeArguments)
-            return rti;
-          return A._Universe__lookupInterfaceRti(universe, rti._primary, substitutedInterfaceTypeArguments);
-        case 10:
-          base = rti._primary;
-          substitutedBase = A._substitute(universe, base, typeArguments, depth);
-          $arguments = rti._rest;
-          substitutedArguments = A._substituteArray(universe, $arguments, typeArguments, depth);
-          if (substitutedBase === base && substitutedArguments === $arguments)
-            return rti;
-          return A._Universe__lookupBindingRti(universe, substitutedBase, substitutedArguments);
-        case 12:
-          returnType = rti._primary;
-          substitutedReturnType = A._substitute(universe, returnType, typeArguments, depth);
-          functionParameters = rti._rest;
-          substitutedFunctionParameters = A._substituteFunctionParameters(universe, functionParameters, typeArguments, depth);
-          if (substitutedReturnType === returnType && substitutedFunctionParameters === functionParameters)
-            return rti;
-          return A._Universe__lookupFunctionRti(universe, substitutedReturnType, substitutedFunctionParameters);
-        case 13:
-          bounds = rti._rest;
-          depth += bounds.length;
-          substitutedBounds = A._substituteArray(universe, bounds, typeArguments, depth);
-          base = rti._primary;
-          substitutedBase = A._substitute(universe, base, typeArguments, depth);
-          if (substitutedBounds === bounds && substitutedBase === base)
-            return rti;
-          return A._Universe__lookupGenericFunctionRti(universe, substitutedBase, substitutedBounds, true);
-        case 14:
-          index = rti._primary;
-          if (index < depth)
-            return rti;
-          argument = typeArguments[index - depth];
-          if (argument == null)
-            return rti;
-          return argument;
-        default:
-          throw A.wrapException(A.AssertionError$("Attempted to substitute unexpected RTI kind " + kind));
-      }
-    },
-    _substituteArray(universe, rtiArray, typeArguments, depth) {
-      var changed, i, rti, substitutedRti,
-        $length = rtiArray.length,
-        result = A._Utils_newArrayOrEmpty($length);
-      for (changed = false, i = 0; i < $length; ++i) {
-        rti = rtiArray[i];
-        substitutedRti = A._substitute(universe, rti, typeArguments, depth);
-        if (substitutedRti !== rti)
-          changed = true;
-        result[i] = substitutedRti;
-      }
-      return changed ? result : rtiArray;
-    },
-    _substituteNamed(universe, namedArray, typeArguments, depth) {
-      var changed, i, t1, t2, rti, substitutedRti,
-        $length = namedArray.length,
-        result = A._Utils_newArrayOrEmpty($length);
-      for (changed = false, i = 0; i < $length; i += 3) {
-        t1 = namedArray[i];
-        t2 = namedArray[i + 1];
-        rti = namedArray[i + 2];
-        substitutedRti = A._substitute(universe, rti, typeArguments, depth);
-        if (substitutedRti !== rti)
-          changed = true;
-        result.splice(i, 3, t1, t2, substitutedRti);
-      }
-      return changed ? result : namedArray;
-    },
-    _substituteFunctionParameters(universe, functionParameters, typeArguments, depth) {
-      var result,
-        requiredPositional = functionParameters._requiredPositional,
-        substitutedRequiredPositional = A._substituteArray(universe, requiredPositional, typeArguments, depth),
-        optionalPositional = functionParameters._optionalPositional,
-        substitutedOptionalPositional = A._substituteArray(universe, optionalPositional, typeArguments, depth),
-        named = functionParameters._named,
-        substitutedNamed = A._substituteNamed(universe, named, typeArguments, depth);
-      if (substitutedRequiredPositional === requiredPositional && substitutedOptionalPositional === optionalPositional && substitutedNamed === named)
-        return functionParameters;
-      result = new A._FunctionParameters();
-      result._requiredPositional = substitutedRequiredPositional;
-      result._optionalPositional = substitutedOptionalPositional;
-      result._named = substitutedNamed;
-      return result;
-    },
-    _setArrayType(target, rti) {
-      target[init.arrayRti] = rti;
-      return target;
-    },
-    closureFunctionType(closure) {
-      var t1,
-        signature = closure.$signature;
-      if (signature != null) {
-        if (typeof signature == "number")
-          return A.getTypeFromTypesTable(signature);
-        t1 = closure.$signature();
-        return t1;
-      }
-      return null;
-    },
-    instanceOrFunctionType(object, testRti) {
-      var rti;
-      if (A.Rti__isUnionOfFunctionType(testRti))
-        if (object instanceof A.Closure) {
-          rti = A.closureFunctionType(object);
-          if (rti != null)
-            return rti;
-        }
-      return A.instanceType(object);
-    },
-    instanceType(object) {
-      if (object instanceof A.Object)
-        return A._instanceType(object);
-      if (Array.isArray(object))
-        return A._arrayInstanceType(object);
-      return A._instanceTypeFromConstructor(J.getInterceptor$(object));
-    },
-    _arrayInstanceType(object) {
-      var rti = object[init.arrayRti],
-        defaultRti = type$.JSArray_dynamic;
-      if (rti == null)
-        return defaultRti;
-      if (rti.constructor !== defaultRti.constructor)
-        return defaultRti;
-      return rti;
-    },
-    _instanceType(object) {
-      var rti = object.$ti;
-      return rti != null ? rti : A._instanceTypeFromConstructor(object);
-    },
-    _instanceTypeFromConstructor(instance) {
-      var $constructor = instance.constructor,
-        probe = $constructor.$ccache;
-      if (probe != null)
-        return probe;
-      return A._instanceTypeFromConstructorMiss(instance, $constructor);
-    },
-    _instanceTypeFromConstructorMiss(instance, $constructor) {
-      var effectiveConstructor = instance instanceof A.Closure ? instance.__proto__.__proto__.constructor : $constructor,
-        rti = A._Universe_findErasedType(init.typeUniverse, effectiveConstructor.name);
-      $constructor.$ccache = rti;
-      return rti;
-    },
-    getTypeFromTypesTable(index) {
-      var rti,
-        table = init.types,
-        type = table[index];
-      if (typeof type == "string") {
-        rti = A._Universe_eval(init.typeUniverse, type, false);
-        table[index] = rti;
-        return rti;
-      }
-      return type;
-    },
-    getRuntimeTypeOfDartObject(object) {
-      return A.createRuntimeType(A._instanceType(object));
-    },
-    _structuralTypeOf(object) {
-      var functionRti = object instanceof A.Closure ? A.closureFunctionType(object) : null;
-      if (functionRti != null)
-        return functionRti;
-      if (type$.TrustedGetRuntimeType._is(object))
-        return J.get$runtimeType$(object)._rti;
-      if (Array.isArray(object))
-        return A._arrayInstanceType(object);
-      return A.instanceType(object);
-    },
-    createRuntimeType(rti) {
-      var t1 = rti._cachedRuntimeType;
-      return t1 == null ? rti._cachedRuntimeType = A._createRuntimeType(rti) : t1;
-    },
-    _createRuntimeType(rti) {
-      var starErasedRti, t1,
-        s = rti._canonicalRecipe,
-        starErasedRecipe = s.replace(/\*/g, "");
-      if (starErasedRecipe === s)
-        return rti._cachedRuntimeType = new A._Type(rti);
-      starErasedRti = A._Universe_eval(init.typeUniverse, starErasedRecipe, true);
-      t1 = starErasedRti._cachedRuntimeType;
-      return t1 == null ? starErasedRti._cachedRuntimeType = A._createRuntimeType(starErasedRti) : t1;
-    },
-    _installSpecializedIsTest(object) {
-      var t1, unstarred, isFn, $name, predicate, testRti = this;
-      if (testRti === type$.Object)
-        return A._finishIsFn(testRti, object, A._isObject);
-      if (!A.isStrongTopType(testRti))
-        if (!(testRti === type$.legacy_Object))
-          t1 = false;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      if (t1)
-        return A._finishIsFn(testRti, object, A._isTop);
-      t1 = testRti._kind;
-      if (t1 === 7)
-        return A._finishIsFn(testRti, object, A._generalNullableIsTestImplementation);
-      if (t1 === 1)
-        return A._finishIsFn(testRti, object, A._isNever);
-      unstarred = t1 === 6 ? testRti._primary : testRti;
-      t1 = unstarred._kind;
-      if (t1 === 8)
-        return A._finishIsFn(testRti, object, A._isFutureOr);
-      if (unstarred === type$.int)
-        isFn = A._isInt;
-      else if (unstarred === type$.double || unstarred === type$.num)
-        isFn = A._isNum;
-      else if (unstarred === type$.String)
-        isFn = A._isString;
-      else
-        isFn = unstarred === type$.bool ? A._isBool : null;
-      if (isFn != null)
-        return A._finishIsFn(testRti, object, isFn);
-      if (t1 === 9) {
-        $name = unstarred._primary;
-        if (unstarred._rest.every(A.isTopType)) {
-          testRti._specializedTestResource = "$is" + $name;
-          if ($name === "List")
-            return A._finishIsFn(testRti, object, A._isListTestViaProperty);
-          return A._finishIsFn(testRti, object, A._isTestViaProperty);
-        }
-      } else if (t1 === 11) {
-        predicate = A.createRecordTypePredicate(unstarred._primary, unstarred._rest);
-        return A._finishIsFn(testRti, object, predicate == null ? A._isNever : predicate);
-      }
-      return A._finishIsFn(testRti, object, A._generalIsTestImplementation);
-    },
-    _finishIsFn(testRti, object, isFn) {
-      testRti._is = isFn;
-      return testRti._is(object);
-    },
-    _installSpecializedAsCheck(object) {
-      var t1, testRti = this,
-        asFn = A._generalAsCheckImplementation;
-      if (!A.isStrongTopType(testRti))
-        if (!(testRti === type$.legacy_Object))
-          t1 = false;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      if (t1)
-        asFn = A._asTop;
-      else if (testRti === type$.Object)
-        asFn = A._asObject;
-      else {
-        t1 = A.isNullable(testRti);
-        if (t1)
-          asFn = A._generalNullableAsCheckImplementation;
-      }
-      testRti._as = asFn;
-      return testRti._as(object);
-    },
-    _nullIs(testRti) {
-      var t1,
-        kind = testRti._kind;
-      if (!A.isStrongTopType(testRti))
-        if (!(testRti === type$.legacy_Object))
-          if (!(testRti === type$.legacy_Never))
-            if (kind !== 7)
-              if (!(kind === 6 && A._nullIs(testRti._primary)))
-                t1 = kind === 8 && A._nullIs(testRti._primary) || testRti === type$.Null || testRti === type$.JSNull;
-              else
-                t1 = true;
-            else
-              t1 = true;
-          else
-            t1 = true;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      return t1;
-    },
-    _generalIsTestImplementation(object) {
-      var testRti = this;
-      if (object == null)
-        return A._nullIs(testRti);
-      return A._isSubtype(init.typeUniverse, A.instanceOrFunctionType(object, testRti), null, testRti, null);
-    },
-    _generalNullableIsTestImplementation(object) {
-      if (object == null)
-        return true;
-      return this._primary._is(object);
-    },
-    _isTestViaProperty(object) {
-      var tag, testRti = this;
-      if (object == null)
-        return A._nullIs(testRti);
-      tag = testRti._specializedTestResource;
-      if (object instanceof A.Object)
-        return !!object[tag];
-      return !!J.getInterceptor$(object)[tag];
-    },
-    _isListTestViaProperty(object) {
-      var tag, testRti = this;
-      if (object == null)
-        return A._nullIs(testRti);
-      if (typeof object != "object")
-        return false;
-      if (Array.isArray(object))
-        return true;
-      tag = testRti._specializedTestResource;
-      if (object instanceof A.Object)
-        return !!object[tag];
-      return !!J.getInterceptor$(object)[tag];
-    },
-    _generalAsCheckImplementation(object) {
-      var t1, testRti = this;
-      if (object == null) {
-        t1 = A.isNullable(testRti);
-        if (t1)
-          return object;
-      } else if (testRti._is(object))
-        return object;
-      A._failedAsCheck(object, testRti);
-    },
-    _generalNullableAsCheckImplementation(object) {
-      var testRti = this;
-      if (object == null)
-        return object;
-      else if (testRti._is(object))
-        return object;
-      A._failedAsCheck(object, testRti);
-    },
-    _failedAsCheck(object, testRti) {
-      throw A.wrapException(A._TypeError$fromMessage(A._Error_compose(object, A._rtiToString(testRti, null))));
-    },
-    _Error_compose(object, checkedTypeDescription) {
-      return A.Error_safeToString(object) + ": type '" + A._rtiToString(A._structuralTypeOf(object), null) + "' is not a subtype of type '" + checkedTypeDescription + "'";
-    },
-    _TypeError$fromMessage(message) {
-      return new A._TypeError("TypeError: " + message);
-    },
-    _TypeError__TypeError$forType(object, type) {
-      return new A._TypeError("TypeError: " + A._Error_compose(object, type));
-    },
-    _isFutureOr(object) {
-      var testRti = this;
-      return testRti._primary._is(object) || A.Rti__getFutureFromFutureOr(init.typeUniverse, testRti)._is(object);
-    },
-    _isObject(object) {
-      return object != null;
-    },
-    _asObject(object) {
-      if (object != null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "Object"));
-    },
-    _isTop(object) {
-      return true;
-    },
-    _asTop(object) {
-      return object;
-    },
-    _isNever(object) {
-      return false;
-    },
-    _isBool(object) {
-      return true === object || false === object;
-    },
-    _asBool(object) {
-      if (true === object)
-        return true;
-      if (false === object)
-        return false;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "bool"));
-    },
-    _asBoolS(object) {
-      if (true === object)
-        return true;
-      if (false === object)
-        return false;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "bool"));
-    },
-    _asBoolQ(object) {
-      if (true === object)
-        return true;
-      if (false === object)
-        return false;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "bool?"));
-    },
-    _asDouble(object) {
-      if (typeof object == "number")
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "double"));
-    },
-    _asDoubleS(object) {
-      if (typeof object == "number")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "double"));
-    },
-    _asDoubleQ(object) {
-      if (typeof object == "number")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "double?"));
-    },
-    _isInt(object) {
-      return typeof object == "number" && Math.floor(object) === object;
-    },
-    _asInt(object) {
-      if (typeof object == "number" && Math.floor(object) === object)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "int"));
-    },
-    _asIntS(object) {
-      if (typeof object == "number" && Math.floor(object) === object)
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "int"));
-    },
-    _asIntQ(object) {
-      if (typeof object == "number" && Math.floor(object) === object)
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "int?"));
-    },
-    _isNum(object) {
-      return typeof object == "number";
-    },
-    _asNum(object) {
-      if (typeof object == "number")
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "num"));
-    },
-    _asNumS(object) {
-      if (typeof object == "number")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "num"));
-    },
-    _asNumQ(object) {
-      if (typeof object == "number")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "num?"));
-    },
-    _isString(object) {
-      return typeof object == "string";
-    },
-    _asString(object) {
-      if (typeof object == "string")
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "String"));
-    },
-    _asStringS(object) {
-      if (typeof object == "string")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "String"));
-    },
-    _asStringQ(object) {
-      if (typeof object == "string")
-        return object;
-      if (object == null)
-        return object;
-      throw A.wrapException(A._TypeError__TypeError$forType(object, "String?"));
-    },
-    _rtiArrayToString(array, genericContext) {
-      var s, sep, i;
-      for (s = "", sep = "", i = 0; i < array.length; ++i, sep = ", ")
-        s += sep + A._rtiToString(array[i], genericContext);
-      return s;
-    },
-    _recordRtiToString(recordType, genericContext) {
-      var fieldCount, names, namesIndex, s, comma, i,
-        partialShape = recordType._primary,
-        fields = recordType._rest;
-      if ("" === partialShape)
-        return "(" + A._rtiArrayToString(fields, genericContext) + ")";
-      fieldCount = fields.length;
-      names = partialShape.split(",");
-      namesIndex = names.length - fieldCount;
-      for (s = "(", comma = "", i = 0; i < fieldCount; ++i, comma = ", ") {
-        s += comma;
-        if (namesIndex === 0)
-          s += "{";
-        s += A._rtiToString(fields[i], genericContext);
-        if (namesIndex >= 0)
-          s += " " + names[namesIndex];
-        ++namesIndex;
-      }
-      return s + "})";
-    },
-    _functionRtiToString(functionType, genericContext, bounds) {
-      var boundsLength, outerContextLength, offset, i, t1, t2, typeParametersText, typeSep, t3, t4, boundRti, kind, parameters, requiredPositional, requiredPositionalLength, optionalPositional, optionalPositionalLength, named, namedLength, returnTypeText, argumentsText, sep, _s2_ = ", ";
-      if (bounds != null) {
-        boundsLength = bounds.length;
-        if (genericContext == null) {
-          genericContext = A._setArrayType([], type$.JSArray_String);
-          outerContextLength = null;
-        } else
-          outerContextLength = genericContext.length;
-        offset = genericContext.length;
-        for (i = boundsLength; i > 0; --i)
-          B.JSArray_methods.add$1(genericContext, "T" + (offset + i));
-        for (t1 = type$.nullable_Object, t2 = type$.legacy_Object, typeParametersText = "<", typeSep = "", i = 0; i < boundsLength; ++i, typeSep = _s2_) {
-          t3 = genericContext.length;
-          t4 = t3 - 1 - i;
-          if (!(t4 >= 0))
-            return A.ioore(genericContext, t4);
-          typeParametersText = B.JSString_methods.$add(typeParametersText + typeSep, genericContext[t4]);
-          boundRti = bounds[i];
-          kind = boundRti._kind;
-          if (!(kind === 2 || kind === 3 || kind === 4 || kind === 5 || boundRti === t1))
-            if (!(boundRti === t2))
-              t3 = false;
-            else
-              t3 = true;
-          else
-            t3 = true;
-          if (!t3)
-            typeParametersText += " extends " + A._rtiToString(boundRti, genericContext);
-        }
-        typeParametersText += ">";
-      } else {
-        typeParametersText = "";
-        outerContextLength = null;
-      }
-      t1 = functionType._primary;
-      parameters = functionType._rest;
-      requiredPositional = parameters._requiredPositional;
-      requiredPositionalLength = requiredPositional.length;
-      optionalPositional = parameters._optionalPositional;
-      optionalPositionalLength = optionalPositional.length;
-      named = parameters._named;
-      namedLength = named.length;
-      returnTypeText = A._rtiToString(t1, genericContext);
-      for (argumentsText = "", sep = "", i = 0; i < requiredPositionalLength; ++i, sep = _s2_)
-        argumentsText += sep + A._rtiToString(requiredPositional[i], genericContext);
-      if (optionalPositionalLength > 0) {
-        argumentsText += sep + "[";
-        for (sep = "", i = 0; i < optionalPositionalLength; ++i, sep = _s2_)
-          argumentsText += sep + A._rtiToString(optionalPositional[i], genericContext);
-        argumentsText += "]";
-      }
-      if (namedLength > 0) {
-        argumentsText += sep + "{";
-        for (sep = "", i = 0; i < namedLength; i += 3, sep = _s2_) {
-          argumentsText += sep;
-          if (named[i + 1])
-            argumentsText += "required ";
-          argumentsText += A._rtiToString(named[i + 2], genericContext) + " " + named[i];
-        }
-        argumentsText += "}";
-      }
-      if (outerContextLength != null) {
-        genericContext.toString;
-        genericContext.length = outerContextLength;
-      }
-      return typeParametersText + "(" + argumentsText + ") => " + returnTypeText;
-    },
-    _rtiToString(rti, genericContext) {
-      var s, questionArgument, argumentKind, $name, $arguments, t1, t2,
-        kind = rti._kind;
-      if (kind === 5)
-        return "erased";
-      if (kind === 2)
-        return "dynamic";
-      if (kind === 3)
-        return "void";
-      if (kind === 1)
-        return "Never";
-      if (kind === 4)
-        return "any";
-      if (kind === 6) {
-        s = A._rtiToString(rti._primary, genericContext);
-        return s;
-      }
-      if (kind === 7) {
-        questionArgument = rti._primary;
-        s = A._rtiToString(questionArgument, genericContext);
-        argumentKind = questionArgument._kind;
-        return (argumentKind === 12 || argumentKind === 13 ? "(" + s + ")" : s) + "?";
-      }
-      if (kind === 8)
-        return "FutureOr<" + A._rtiToString(rti._primary, genericContext) + ">";
-      if (kind === 9) {
-        $name = A._unminifyOrTag(rti._primary);
-        $arguments = rti._rest;
-        return $arguments.length > 0 ? $name + ("<" + A._rtiArrayToString($arguments, genericContext) + ">") : $name;
-      }
-      if (kind === 11)
-        return A._recordRtiToString(rti, genericContext);
-      if (kind === 12)
-        return A._functionRtiToString(rti, genericContext, null);
-      if (kind === 13)
-        return A._functionRtiToString(rti._primary, genericContext, rti._rest);
-      if (kind === 14) {
-        t1 = rti._primary;
-        t2 = genericContext.length;
-        t1 = t2 - 1 - t1;
-        if (!(t1 >= 0 && t1 < t2))
-          return A.ioore(genericContext, t1);
-        return genericContext[t1];
-      }
-      return "?";
-    },
-    _unminifyOrTag(rawClassName) {
-      var preserved = init.mangledGlobalNames[rawClassName];
-      if (preserved != null)
-        return preserved;
-      return rawClassName;
-    },
-    _Universe_findRule(universe, targetType) {
-      var rule = universe.tR[targetType];
-      for (; typeof rule == "string";)
-        rule = universe.tR[rule];
-      return rule;
-    },
-    _Universe_findErasedType(universe, cls) {
-      var $length, erased, $arguments, i, $interface,
-        t1 = universe.eT,
-        probe = t1[cls];
-      if (probe == null)
-        return A._Universe_eval(universe, cls, false);
-      else if (typeof probe == "number") {
-        $length = probe;
-        erased = A._Universe__lookupTerminalRti(universe, 5, "#");
-        $arguments = A._Utils_newArrayOrEmpty($length);
-        for (i = 0; i < $length; ++i)
-          $arguments[i] = erased;
-        $interface = A._Universe__lookupInterfaceRti(universe, cls, $arguments);
-        t1[cls] = $interface;
-        return $interface;
-      } else
-        return probe;
-    },
-    _Universe_addRules(universe, rules) {
-      return A._Utils_objectAssign(universe.tR, rules);
-    },
-    _Universe_addErasedTypes(universe, types) {
-      return A._Utils_objectAssign(universe.eT, types);
-    },
-    _Universe_eval(universe, recipe, normalize) {
-      var rti,
-        t1 = universe.eC,
-        probe = t1.get(recipe);
-      if (probe != null)
-        return probe;
-      rti = A._Parser_parse(A._Parser_create(universe, null, recipe, normalize));
-      t1.set(recipe, rti);
-      return rti;
-    },
-    _Universe_evalInEnvironment(universe, environment, recipe) {
-      var probe, rti,
-        cache = environment._evalCache;
-      if (cache == null)
-        cache = environment._evalCache = new Map();
-      probe = cache.get(recipe);
-      if (probe != null)
-        return probe;
-      rti = A._Parser_parse(A._Parser_create(universe, environment, recipe, true));
-      cache.set(recipe, rti);
-      return rti;
-    },
-    _Universe_bind(universe, environment, argumentsRti) {
-      var argumentsRecipe, probe, rti,
-        cache = environment._bindCache;
-      if (cache == null)
-        cache = environment._bindCache = new Map();
-      argumentsRecipe = argumentsRti._canonicalRecipe;
-      probe = cache.get(argumentsRecipe);
-      if (probe != null)
-        return probe;
-      rti = A._Universe__lookupBindingRti(universe, environment, argumentsRti._kind === 10 ? argumentsRti._rest : [argumentsRti]);
-      cache.set(argumentsRecipe, rti);
-      return rti;
-    },
-    _Universe__installTypeTests(universe, rti) {
-      rti._as = A._installSpecializedAsCheck;
-      rti._is = A._installSpecializedIsTest;
-      return rti;
-    },
-    _Universe__lookupTerminalRti(universe, kind, key) {
-      var rti, t1,
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = kind;
-      rti._canonicalRecipe = key;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__lookupStarRti(universe, baseType, normalize) {
-      var t1,
-        key = baseType._canonicalRecipe + "*",
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      t1 = A._Universe__createStarRti(universe, baseType, key, normalize);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__createStarRti(universe, baseType, key, normalize) {
-      var baseKind, t1, rti;
-      if (normalize) {
-        baseKind = baseType._kind;
-        if (!A.isStrongTopType(baseType))
-          t1 = baseType === type$.Null || baseType === type$.JSNull || baseKind === 7 || baseKind === 6;
-        else
-          t1 = true;
-        if (t1)
-          return baseType;
-      }
-      rti = new A.Rti(null, null);
-      rti._kind = 6;
-      rti._primary = baseType;
-      rti._canonicalRecipe = key;
-      return A._Universe__installTypeTests(universe, rti);
-    },
-    _Universe__lookupQuestionRti(universe, baseType, normalize) {
-      var t1,
-        key = baseType._canonicalRecipe + "?",
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      t1 = A._Universe__createQuestionRti(universe, baseType, key, normalize);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__createQuestionRti(universe, baseType, key, normalize) {
-      var baseKind, t1, starArgument, rti;
-      if (normalize) {
-        baseKind = baseType._kind;
-        if (!A.isStrongTopType(baseType))
-          if (!(baseType === type$.Null || baseType === type$.JSNull))
-            if (baseKind !== 7)
-              t1 = baseKind === 8 && A.isNullable(baseType._primary);
-            else
-              t1 = true;
-          else
-            t1 = true;
-        else
-          t1 = true;
-        if (t1)
-          return baseType;
-        else if (baseKind === 1 || baseType === type$.legacy_Never)
-          return type$.Null;
-        else if (baseKind === 6) {
-          starArgument = baseType._primary;
-          if (starArgument._kind === 8 && A.isNullable(starArgument._primary))
-            return starArgument;
-          else
-            return A.Rti__getQuestionFromStar(universe, baseType);
-        }
-      }
-      rti = new A.Rti(null, null);
-      rti._kind = 7;
-      rti._primary = baseType;
-      rti._canonicalRecipe = key;
-      return A._Universe__installTypeTests(universe, rti);
-    },
-    _Universe__lookupFutureOrRti(universe, baseType, normalize) {
-      var t1,
-        key = baseType._canonicalRecipe + "/",
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      t1 = A._Universe__createFutureOrRti(universe, baseType, key, normalize);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__createFutureOrRti(universe, baseType, key, normalize) {
-      var t1, t2, rti;
-      if (normalize) {
-        t1 = baseType._kind;
-        if (!A.isStrongTopType(baseType))
-          if (!(baseType === type$.legacy_Object))
-            t2 = false;
-          else
-            t2 = true;
-        else
-          t2 = true;
-        if (t2 || baseType === type$.Object)
-          return baseType;
-        else if (t1 === 1)
-          return A._Universe__lookupInterfaceRti(universe, "Future", [baseType]);
-        else if (baseType === type$.Null || baseType === type$.JSNull)
-          return type$.nullable_Future_Null;
-      }
-      rti = new A.Rti(null, null);
-      rti._kind = 8;
-      rti._primary = baseType;
-      rti._canonicalRecipe = key;
-      return A._Universe__installTypeTests(universe, rti);
-    },
-    _Universe__lookupGenericFunctionParameterRti(universe, index) {
-      var rti, t1,
-        key = "" + index + "^",
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = 14;
-      rti._primary = index;
-      rti._canonicalRecipe = key;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__canonicalRecipeJoin($arguments) {
-      var s, sep, i,
-        $length = $arguments.length;
-      for (s = "", sep = "", i = 0; i < $length; ++i, sep = ",")
-        s += sep + $arguments[i]._canonicalRecipe;
-      return s;
-    },
-    _Universe__canonicalRecipeJoinNamed($arguments) {
-      var s, sep, i, t1, nameSep,
-        $length = $arguments.length;
-      for (s = "", sep = "", i = 0; i < $length; i += 3, sep = ",") {
-        t1 = $arguments[i];
-        nameSep = $arguments[i + 1] ? "!" : ":";
-        s += sep + t1 + nameSep + $arguments[i + 2]._canonicalRecipe;
-      }
-      return s;
-    },
-    _Universe__lookupInterfaceRti(universe, $name, $arguments) {
-      var probe, rti, t1,
-        s = $name;
-      if ($arguments.length > 0)
-        s += "<" + A._Universe__canonicalRecipeJoin($arguments) + ">";
-      probe = universe.eC.get(s);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = 9;
-      rti._primary = $name;
-      rti._rest = $arguments;
-      if ($arguments.length > 0)
-        rti._precomputed1 = $arguments[0];
-      rti._canonicalRecipe = s;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(s, t1);
-      return t1;
-    },
-    _Universe__lookupBindingRti(universe, base, $arguments) {
-      var newBase, newArguments, key, probe, rti, t1;
-      if (base._kind === 10) {
-        newBase = base._primary;
-        newArguments = base._rest.concat($arguments);
-      } else {
-        newArguments = $arguments;
-        newBase = base;
-      }
-      key = newBase._canonicalRecipe + (";<" + A._Universe__canonicalRecipeJoin(newArguments) + ">");
-      probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = 10;
-      rti._primary = newBase;
-      rti._rest = newArguments;
-      rti._canonicalRecipe = key;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__lookupRecordRti(universe, partialShapeTag, fields) {
-      var rti, t1,
-        key = "+" + (partialShapeTag + "(" + A._Universe__canonicalRecipeJoin(fields) + ")"),
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = 11;
-      rti._primary = partialShapeTag;
-      rti._rest = fields;
-      rti._canonicalRecipe = key;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__lookupFunctionRti(universe, returnType, parameters) {
-      var sep, key, probe, rti, t1,
-        s = returnType._canonicalRecipe,
-        requiredPositional = parameters._requiredPositional,
-        requiredPositionalLength = requiredPositional.length,
-        optionalPositional = parameters._optionalPositional,
-        optionalPositionalLength = optionalPositional.length,
-        named = parameters._named,
-        namedLength = named.length,
-        recipe = "(" + A._Universe__canonicalRecipeJoin(requiredPositional);
-      if (optionalPositionalLength > 0) {
-        sep = requiredPositionalLength > 0 ? "," : "";
-        recipe += sep + "[" + A._Universe__canonicalRecipeJoin(optionalPositional) + "]";
-      }
-      if (namedLength > 0) {
-        sep = requiredPositionalLength > 0 ? "," : "";
-        recipe += sep + "{" + A._Universe__canonicalRecipeJoinNamed(named) + "}";
-      }
-      key = s + (recipe + ")");
-      probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      rti = new A.Rti(null, null);
-      rti._kind = 12;
-      rti._primary = returnType;
-      rti._rest = parameters;
-      rti._canonicalRecipe = key;
-      t1 = A._Universe__installTypeTests(universe, rti);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__lookupGenericFunctionRti(universe, baseFunctionType, bounds, normalize) {
-      var t1,
-        key = baseFunctionType._canonicalRecipe + ("<" + A._Universe__canonicalRecipeJoin(bounds) + ">"),
-        probe = universe.eC.get(key);
-      if (probe != null)
-        return probe;
-      t1 = A._Universe__createGenericFunctionRti(universe, baseFunctionType, bounds, key, normalize);
-      universe.eC.set(key, t1);
-      return t1;
-    },
-    _Universe__createGenericFunctionRti(universe, baseFunctionType, bounds, key, normalize) {
-      var $length, typeArguments, count, i, bound, substitutedBase, substitutedBounds, rti;
-      if (normalize) {
-        $length = bounds.length;
-        typeArguments = A._Utils_newArrayOrEmpty($length);
-        for (count = 0, i = 0; i < $length; ++i) {
-          bound = bounds[i];
-          if (bound._kind === 1) {
-            typeArguments[i] = bound;
-            ++count;
-          }
-        }
-        if (count > 0) {
-          substitutedBase = A._substitute(universe, baseFunctionType, typeArguments, 0);
-          substitutedBounds = A._substituteArray(universe, bounds, typeArguments, 0);
-          return A._Universe__lookupGenericFunctionRti(universe, substitutedBase, substitutedBounds, bounds !== substitutedBounds);
-        }
-      }
-      rti = new A.Rti(null, null);
-      rti._kind = 13;
-      rti._primary = baseFunctionType;
-      rti._rest = bounds;
-      rti._canonicalRecipe = key;
-      return A._Universe__installTypeTests(universe, rti);
-    },
-    _Parser_create(universe, environment, recipe, normalize) {
-      return {u: universe, e: environment, r: recipe, s: [], p: 0, n: normalize};
-    },
-    _Parser_parse(parser) {
-      var t2, i, ch, t3, array, end, item,
-        source = parser.r,
-        t1 = parser.s;
-      for (t2 = source.length, i = 0; i < t2;) {
-        ch = source.charCodeAt(i);
-        if (ch >= 48 && ch <= 57)
-          i = A._Parser_handleDigit(i + 1, ch, source, t1);
-        else if ((((ch | 32) >>> 0) - 97 & 65535) < 26 || ch === 95 || ch === 36 || ch === 124)
-          i = A._Parser_handleIdentifier(parser, i, source, t1, false);
-        else if (ch === 46)
-          i = A._Parser_handleIdentifier(parser, i, source, t1, true);
-        else {
-          ++i;
-          switch (ch) {
-            case 44:
-              break;
-            case 58:
-              t1.push(false);
-              break;
-            case 33:
-              t1.push(true);
-              break;
-            case 59:
-              t1.push(A._Parser_toType(parser.u, parser.e, t1.pop()));
-              break;
-            case 94:
-              t1.push(A._Universe__lookupGenericFunctionParameterRti(parser.u, t1.pop()));
-              break;
-            case 35:
-              t1.push(A._Universe__lookupTerminalRti(parser.u, 5, "#"));
-              break;
-            case 64:
-              t1.push(A._Universe__lookupTerminalRti(parser.u, 2, "@"));
-              break;
-            case 126:
-              t1.push(A._Universe__lookupTerminalRti(parser.u, 3, "~"));
-              break;
-            case 60:
-              t1.push(parser.p);
-              parser.p = t1.length;
-              break;
-            case 62:
-              A._Parser_handleTypeArguments(parser, t1);
-              break;
-            case 38:
-              A._Parser_handleExtendedOperations(parser, t1);
-              break;
-            case 42:
-              t3 = parser.u;
-              t1.push(A._Universe__lookupStarRti(t3, A._Parser_toType(t3, parser.e, t1.pop()), parser.n));
-              break;
-            case 63:
-              t3 = parser.u;
-              t1.push(A._Universe__lookupQuestionRti(t3, A._Parser_toType(t3, parser.e, t1.pop()), parser.n));
-              break;
-            case 47:
-              t3 = parser.u;
-              t1.push(A._Universe__lookupFutureOrRti(t3, A._Parser_toType(t3, parser.e, t1.pop()), parser.n));
-              break;
-            case 40:
-              t1.push(-3);
-              t1.push(parser.p);
-              parser.p = t1.length;
-              break;
-            case 41:
-              A._Parser_handleArguments(parser, t1);
-              break;
-            case 91:
-              t1.push(parser.p);
-              parser.p = t1.length;
-              break;
-            case 93:
-              array = t1.splice(parser.p);
-              A._Parser_toTypes(parser.u, parser.e, array);
-              parser.p = t1.pop();
-              t1.push(array);
-              t1.push(-1);
-              break;
-            case 123:
-              t1.push(parser.p);
-              parser.p = t1.length;
-              break;
-            case 125:
-              array = t1.splice(parser.p);
-              A._Parser_toTypesNamed(parser.u, parser.e, array);
-              parser.p = t1.pop();
-              t1.push(array);
-              t1.push(-2);
-              break;
-            case 43:
-              end = source.indexOf("(", i);
-              t1.push(source.substring(i, end));
-              t1.push(-4);
-              t1.push(parser.p);
-              parser.p = t1.length;
-              i = end + 1;
-              break;
-            default:
-              throw "Bad character " + ch;
-          }
-        }
-      }
-      item = t1.pop();
-      return A._Parser_toType(parser.u, parser.e, item);
-    },
-    _Parser_handleDigit(i, digit, source, stack) {
-      var t1, ch,
-        value = digit - 48;
-      for (t1 = source.length; i < t1; ++i) {
-        ch = source.charCodeAt(i);
-        if (!(ch >= 48 && ch <= 57))
-          break;
-        value = value * 10 + (ch - 48);
-      }
-      stack.push(value);
-      return i;
-    },
-    _Parser_handleIdentifier(parser, start, source, stack, hasPeriod) {
-      var t1, ch, t2, string, environment, recipe,
-        i = start + 1;
-      for (t1 = source.length; i < t1; ++i) {
-        ch = source.charCodeAt(i);
-        if (ch === 46) {
-          if (hasPeriod)
-            break;
-          hasPeriod = true;
-        } else {
-          if (!((((ch | 32) >>> 0) - 97 & 65535) < 26 || ch === 95 || ch === 36 || ch === 124))
-            t2 = ch >= 48 && ch <= 57;
-          else
-            t2 = true;
-          if (!t2)
-            break;
-        }
-      }
-      string = source.substring(start, i);
-      if (hasPeriod) {
-        t1 = parser.u;
-        environment = parser.e;
-        if (environment._kind === 10)
-          environment = environment._primary;
-        recipe = A._Universe_findRule(t1, environment._primary)[string];
-        if (recipe == null)
-          A.throwExpression('No "' + string + '" in "' + A.Rti__getCanonicalRecipe(environment) + '"');
-        stack.push(A._Universe_evalInEnvironment(t1, environment, recipe));
-      } else
-        stack.push(string);
-      return i;
-    },
-    _Parser_handleTypeArguments(parser, stack) {
-      var base,
-        t1 = parser.u,
-        $arguments = A._Parser_collectArray(parser, stack),
-        head = stack.pop();
-      if (typeof head == "string")
-        stack.push(A._Universe__lookupInterfaceRti(t1, head, $arguments));
-      else {
-        base = A._Parser_toType(t1, parser.e, head);
-        switch (base._kind) {
-          case 12:
-            stack.push(A._Universe__lookupGenericFunctionRti(t1, base, $arguments, parser.n));
-            break;
-          default:
-            stack.push(A._Universe__lookupBindingRti(t1, base, $arguments));
-            break;
-        }
-      }
-    },
-    _Parser_handleArguments(parser, stack) {
-      var optionalPositional, named, requiredPositional, returnType, parameters, _null = null,
-        t1 = parser.u,
-        head = stack.pop();
-      if (typeof head == "number")
-        switch (head) {
-          case -1:
-            optionalPositional = stack.pop();
-            named = _null;
-            break;
-          case -2:
-            named = stack.pop();
-            optionalPositional = _null;
-            break;
-          default:
-            stack.push(head);
-            named = _null;
-            optionalPositional = named;
-            break;
-        }
-      else {
-        stack.push(head);
-        named = _null;
-        optionalPositional = named;
-      }
-      requiredPositional = A._Parser_collectArray(parser, stack);
-      head = stack.pop();
-      switch (head) {
-        case -3:
-          head = stack.pop();
-          if (optionalPositional == null)
-            optionalPositional = t1.sEA;
-          if (named == null)
-            named = t1.sEA;
-          returnType = A._Parser_toType(t1, parser.e, head);
-          parameters = new A._FunctionParameters();
-          parameters._requiredPositional = requiredPositional;
-          parameters._optionalPositional = optionalPositional;
-          parameters._named = named;
-          stack.push(A._Universe__lookupFunctionRti(t1, returnType, parameters));
-          return;
-        case -4:
-          stack.push(A._Universe__lookupRecordRti(t1, stack.pop(), requiredPositional));
-          return;
-        default:
-          throw A.wrapException(A.AssertionError$("Unexpected state under `()`: " + A.S(head)));
-      }
-    },
-    _Parser_handleExtendedOperations(parser, stack) {
-      var $top = stack.pop();
-      if (0 === $top) {
-        stack.push(A._Universe__lookupTerminalRti(parser.u, 1, "0&"));
-        return;
-      }
-      if (1 === $top) {
-        stack.push(A._Universe__lookupTerminalRti(parser.u, 4, "1&"));
-        return;
-      }
-      throw A.wrapException(A.AssertionError$("Unexpected extended operation " + A.S($top)));
-    },
-    _Parser_collectArray(parser, stack) {
-      var array = stack.splice(parser.p);
-      A._Parser_toTypes(parser.u, parser.e, array);
-      parser.p = stack.pop();
-      return array;
-    },
-    _Parser_toType(universe, environment, item) {
-      if (typeof item == "string")
-        return A._Universe__lookupInterfaceRti(universe, item, universe.sEA);
-      else if (typeof item == "number") {
-        environment.toString;
-        return A._Parser_indexToType(universe, environment, item);
-      } else
-        return item;
-    },
-    _Parser_toTypes(universe, environment, items) {
-      var i,
-        $length = items.length;
-      for (i = 0; i < $length; ++i)
-        items[i] = A._Parser_toType(universe, environment, items[i]);
-    },
-    _Parser_toTypesNamed(universe, environment, items) {
-      var i,
-        $length = items.length;
-      for (i = 2; i < $length; i += 3)
-        items[i] = A._Parser_toType(universe, environment, items[i]);
-    },
-    _Parser_indexToType(universe, environment, index) {
-      var typeArguments, len,
-        kind = environment._kind;
-      if (kind === 10) {
-        if (index === 0)
-          return environment._primary;
-        typeArguments = environment._rest;
-        len = typeArguments.length;
-        if (index <= len)
-          return typeArguments[index - 1];
-        index -= len;
-        environment = environment._primary;
-        kind = environment._kind;
-      } else if (index === 0)
-        return environment;
-      if (kind !== 9)
-        throw A.wrapException(A.AssertionError$("Indexed base must be an interface type"));
-      typeArguments = environment._rest;
-      if (index <= typeArguments.length)
-        return typeArguments[index - 1];
-      throw A.wrapException(A.AssertionError$("Bad index " + index + " for " + environment.toString$0(0)));
-    },
-    _isSubtype(universe, s, sEnv, t, tEnv) {
-      var t1, sKind, leftTypeVariable, tKind, t2, sBounds, tBounds, sLength, i, sBound, tBound;
-      if (s === t)
-        return true;
-      if (!A.isStrongTopType(t))
-        if (!(t === type$.legacy_Object))
-          t1 = false;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      if (t1)
-        return true;
-      sKind = s._kind;
-      if (sKind === 4)
-        return true;
-      if (A.isStrongTopType(s))
-        return false;
-      if (s._kind !== 1)
-        t1 = false;
-      else
-        t1 = true;
-      if (t1)
-        return true;
-      leftTypeVariable = sKind === 14;
-      if (leftTypeVariable)
-        if (A._isSubtype(universe, sEnv[s._primary], sEnv, t, tEnv))
-          return true;
-      tKind = t._kind;
-      t1 = s === type$.Null || s === type$.JSNull;
-      if (t1) {
-        if (tKind === 8)
-          return A._isSubtype(universe, s, sEnv, t._primary, tEnv);
-        return t === type$.Null || t === type$.JSNull || tKind === 7 || tKind === 6;
-      }
-      if (t === type$.Object) {
-        if (sKind === 8)
-          return A._isSubtype(universe, s._primary, sEnv, t, tEnv);
-        if (sKind === 6)
-          return A._isSubtype(universe, s._primary, sEnv, t, tEnv);
-        return sKind !== 7;
-      }
-      if (sKind === 6)
-        return A._isSubtype(universe, s._primary, sEnv, t, tEnv);
-      if (tKind === 6) {
-        t1 = A.Rti__getQuestionFromStar(universe, t);
-        return A._isSubtype(universe, s, sEnv, t1, tEnv);
-      }
-      if (sKind === 8) {
-        if (!A._isSubtype(universe, s._primary, sEnv, t, tEnv))
-          return false;
-        return A._isSubtype(universe, A.Rti__getFutureFromFutureOr(universe, s), sEnv, t, tEnv);
-      }
-      if (sKind === 7) {
-        t1 = A._isSubtype(universe, type$.Null, sEnv, t, tEnv);
-        return t1 && A._isSubtype(universe, s._primary, sEnv, t, tEnv);
-      }
-      if (tKind === 8) {
-        if (A._isSubtype(universe, s, sEnv, t._primary, tEnv))
-          return true;
-        return A._isSubtype(universe, s, sEnv, A.Rti__getFutureFromFutureOr(universe, t), tEnv);
-      }
-      if (tKind === 7) {
-        t1 = A._isSubtype(universe, s, sEnv, type$.Null, tEnv);
-        return t1 || A._isSubtype(universe, s, sEnv, t._primary, tEnv);
-      }
-      if (leftTypeVariable)
-        return false;
-      t1 = sKind !== 12;
-      if ((!t1 || sKind === 13) && t === type$.Function)
-        return true;
-      t2 = sKind === 11;
-      if (t2 && t === type$.Record)
-        return true;
-      if (tKind === 13) {
-        if (s === type$.JavaScriptFunction)
-          return true;
-        if (sKind !== 13)
-          return false;
-        sBounds = s._rest;
-        tBounds = t._rest;
-        sLength = sBounds.length;
-        if (sLength !== tBounds.length)
-          return false;
-        sEnv = sEnv == null ? sBounds : sBounds.concat(sEnv);
-        tEnv = tEnv == null ? tBounds : tBounds.concat(tEnv);
-        for (i = 0; i < sLength; ++i) {
-          sBound = sBounds[i];
-          tBound = tBounds[i];
-          if (!A._isSubtype(universe, sBound, sEnv, tBound, tEnv) || !A._isSubtype(universe, tBound, tEnv, sBound, sEnv))
-            return false;
-        }
-        return A._isFunctionSubtype(universe, s._primary, sEnv, t._primary, tEnv);
-      }
-      if (tKind === 12) {
-        if (s === type$.JavaScriptFunction)
-          return true;
-        if (t1)
-          return false;
-        return A._isFunctionSubtype(universe, s, sEnv, t, tEnv);
-      }
-      if (sKind === 9) {
-        if (tKind !== 9)
-          return false;
-        return A._isInterfaceSubtype(universe, s, sEnv, t, tEnv);
-      }
-      if (t2 && tKind === 11)
-        return A._isRecordSubtype(universe, s, sEnv, t, tEnv);
-      return false;
-    },
-    _isFunctionSubtype(universe, s, sEnv, t, tEnv) {
-      var sParameters, tParameters, sRequiredPositional, tRequiredPositional, sRequiredPositionalLength, tRequiredPositionalLength, requiredPositionalDelta, sOptionalPositional, tOptionalPositional, sOptionalPositionalLength, tOptionalPositionalLength, i, t1, sNamed, tNamed, sNamedLength, tNamedLength, sIndex, tIndex, tName, sName, sIsRequired;
-      if (!A._isSubtype(universe, s._primary, sEnv, t._primary, tEnv))
-        return false;
-      sParameters = s._rest;
-      tParameters = t._rest;
-      sRequiredPositional = sParameters._requiredPositional;
-      tRequiredPositional = tParameters._requiredPositional;
-      sRequiredPositionalLength = sRequiredPositional.length;
-      tRequiredPositionalLength = tRequiredPositional.length;
-      if (sRequiredPositionalLength > tRequiredPositionalLength)
-        return false;
-      requiredPositionalDelta = tRequiredPositionalLength - sRequiredPositionalLength;
-      sOptionalPositional = sParameters._optionalPositional;
-      tOptionalPositional = tParameters._optionalPositional;
-      sOptionalPositionalLength = sOptionalPositional.length;
-      tOptionalPositionalLength = tOptionalPositional.length;
-      if (sRequiredPositionalLength + sOptionalPositionalLength < tRequiredPositionalLength + tOptionalPositionalLength)
-        return false;
-      for (i = 0; i < sRequiredPositionalLength; ++i) {
-        t1 = sRequiredPositional[i];
-        if (!A._isSubtype(universe, tRequiredPositional[i], tEnv, t1, sEnv))
-          return false;
-      }
-      for (i = 0; i < requiredPositionalDelta; ++i) {
-        t1 = sOptionalPositional[i];
-        if (!A._isSubtype(universe, tRequiredPositional[sRequiredPositionalLength + i], tEnv, t1, sEnv))
-          return false;
-      }
-      for (i = 0; i < tOptionalPositionalLength; ++i) {
-        t1 = sOptionalPositional[requiredPositionalDelta + i];
-        if (!A._isSubtype(universe, tOptionalPositional[i], tEnv, t1, sEnv))
-          return false;
-      }
-      sNamed = sParameters._named;
-      tNamed = tParameters._named;
-      sNamedLength = sNamed.length;
-      tNamedLength = tNamed.length;
-      for (sIndex = 0, tIndex = 0; tIndex < tNamedLength; tIndex += 3) {
-        tName = tNamed[tIndex];
-        for (; true;) {
-          if (sIndex >= sNamedLength)
-            return false;
-          sName = sNamed[sIndex];
-          sIndex += 3;
-          if (tName < sName)
-            return false;
-          sIsRequired = sNamed[sIndex - 2];
-          if (sName < tName) {
-            if (sIsRequired)
-              return false;
-            continue;
-          }
-          t1 = tNamed[tIndex + 1];
-          if (sIsRequired && !t1)
-            return false;
-          t1 = sNamed[sIndex - 1];
-          if (!A._isSubtype(universe, tNamed[tIndex + 2], tEnv, t1, sEnv))
-            return false;
-          break;
-        }
-      }
-      for (; sIndex < sNamedLength;) {
-        if (sNamed[sIndex + 1])
-          return false;
-        sIndex += 3;
-      }
-      return true;
-    },
-    _isInterfaceSubtype(universe, s, sEnv, t, tEnv) {
-      var rule, recipes, $length, supertypeArgs, i, t1, t2,
-        sName = s._primary,
-        tName = t._primary;
-      for (; sName !== tName;) {
-        rule = universe.tR[sName];
-        if (rule == null)
-          return false;
-        if (typeof rule == "string") {
-          sName = rule;
-          continue;
-        }
-        recipes = rule[tName];
-        if (recipes == null)
-          return false;
-        $length = recipes.length;
-        supertypeArgs = $length > 0 ? new Array($length) : init.typeUniverse.sEA;
-        for (i = 0; i < $length; ++i)
-          supertypeArgs[i] = A._Universe_evalInEnvironment(universe, s, recipes[i]);
-        return A._areArgumentsSubtypes(universe, supertypeArgs, null, sEnv, t._rest, tEnv);
-      }
-      t1 = s._rest;
-      t2 = t._rest;
-      return A._areArgumentsSubtypes(universe, t1, null, sEnv, t2, tEnv);
-    },
-    _areArgumentsSubtypes(universe, sArgs, sVariances, sEnv, tArgs, tEnv) {
-      var i, t1, t2,
-        $length = sArgs.length;
-      for (i = 0; i < $length; ++i) {
-        t1 = sArgs[i];
-        t2 = tArgs[i];
-        if (!A._isSubtype(universe, t1, sEnv, t2, tEnv))
-          return false;
-      }
-      return true;
-    },
-    _isRecordSubtype(universe, s, sEnv, t, tEnv) {
-      var i,
-        sFields = s._rest,
-        tFields = t._rest,
-        sCount = sFields.length;
-      if (sCount !== tFields.length)
-        return false;
-      if (s._primary !== t._primary)
-        return false;
-      for (i = 0; i < sCount; ++i)
-        if (!A._isSubtype(universe, sFields[i], sEnv, tFields[i], tEnv))
-          return false;
-      return true;
-    },
-    isNullable(t) {
-      var t1,
-        kind = t._kind;
-      if (!(t === type$.Null || t === type$.JSNull))
-        if (!A.isStrongTopType(t))
-          if (kind !== 7)
-            if (!(kind === 6 && A.isNullable(t._primary)))
-              t1 = kind === 8 && A.isNullable(t._primary);
-            else
-              t1 = true;
-          else
-            t1 = true;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      return t1;
-    },
-    isTopType(t) {
-      var t1;
-      if (!A.isStrongTopType(t))
-        if (!(t === type$.legacy_Object))
-          t1 = false;
-        else
-          t1 = true;
-      else
-        t1 = true;
-      return t1;
-    },
-    isStrongTopType(t) {
-      var kind = t._kind;
-      return kind === 2 || kind === 3 || kind === 4 || kind === 5 || t === type$.nullable_Object;
-    },
-    _Utils_objectAssign(o, other) {
-      var i, key,
-        keys = Object.keys(other),
-        $length = keys.length;
-      for (i = 0; i < $length; ++i) {
-        key = keys[i];
-        o[key] = other[key];
-      }
-    },
-    _Utils_newArrayOrEmpty($length) {
-      return $length > 0 ? new Array($length) : init.typeUniverse.sEA;
-    },
-    Rti: function Rti(t0, t1) {
-      var _ = this;
-      _._as = t0;
-      _._is = t1;
-      _._cachedRuntimeType = _._specializedTestResource = _._precomputed1 = null;
-      _._kind = 0;
-      _._canonicalRecipe = _._bindCache = _._evalCache = _._rest = _._primary = null;
-    },
-    _FunctionParameters: function _FunctionParameters() {
-      this._named = this._optionalPositional = this._requiredPositional = null;
-    },
-    _Type: function _Type(t0) {
-      this._rti = t0;
-    },
-    _Error: function _Error() {
-    },
-    _TypeError: function _TypeError(t0) {
-      this.__rti$_message = t0;
-    },
-    HashSet_HashSet($E) {
-      return new A._HashSet($E._eval$1("_HashSet<0>"));
-    },
-    _HashSet__newHashTable() {
-      var table = Object.create(null);
-      table["<non-identifier-key>"] = table;
-      delete table["<non-identifier-key>"];
-      return table;
-    },
-    LinkedHashSet_LinkedHashSet($E) {
-      return new A._LinkedHashSet($E._eval$1("_LinkedHashSet<0>"));
-    },
-    LinkedHashSet_LinkedHashSet$_literal(values, $E) {
-      return $E._eval$1("LinkedHashSet<0>")._as(A.fillLiteralSet(values, new A._LinkedHashSet($E._eval$1("_LinkedHashSet<0>"))));
-    },
-    _LinkedHashSet__newHashTable() {
-      var table = Object.create(null);
-      table["<non-identifier-key>"] = table;
-      delete table["<non-identifier-key>"];
-      return table;
-    },
-    _LinkedHashSetIterator$(_set, _modifications, $E) {
-      var t1 = new A._LinkedHashSetIterator(_set, _modifications, $E._eval$1("_LinkedHashSetIterator<0>"));
-      t1._cell = _set._first;
-      return t1;
-    },
-    _HashSet: function _HashSet(t0) {
-      var _ = this;
-      _._collection$_length = 0;
-      _._elements = _._collection$_rest = _._nums = _._strings = null;
-      _.$ti = t0;
-    },
-    _HashSetIterator: function _HashSetIterator(t0, t1, t2) {
-      var _ = this;
-      _._set = t0;
-      _._elements = t1;
-      _._offset = 0;
-      _._collection$_current = null;
-      _.$ti = t2;
-    },
-    _LinkedHashSet: function _LinkedHashSet(t0) {
-      var _ = this;
-      _._collection$_length = 0;
-      _._last = _._first = _._collection$_rest = _._nums = _._strings = null;
-      _._modifications = 0;
-      _.$ti = t0;
-    },
-    _LinkedHashSetCell: function _LinkedHashSetCell(t0) {
-      this._collection$_element = t0;
-      this._previous = this._next = null;
-    },
-    _LinkedHashSetIterator: function _LinkedHashSetIterator(t0, t1, t2) {
-      var _ = this;
-      _._set = t0;
-      _._modifications = t1;
-      _._collection$_current = _._cell = null;
-      _.$ti = t2;
-    },
-    ListBase: function ListBase() {
-    },
-    SetBase: function SetBase() {
-    },
-    _SetBase: function _SetBase() {
-    },
-    List_List$filled($length, fill, growable, $E) {
-      var i,
-        result = growable ? J.JSArray_JSArray$growable($length, $E) : J.JSArray_JSArray$fixed($length, $E);
-      if ($length !== 0 && fill != null)
-        for (i = 0; i < result.length; ++i)
-          result[i] = fill;
-      return result;
-    },
-    List_List$from(elements, growable, $E) {
-      var t1,
-        list = A._setArrayType([], $E._eval$1("JSArray<0>"));
-      for (t1 = elements.get$iterator(elements); t1.moveNext$0();)
-        B.JSArray_methods.add$1(list, $E._as(t1.get$current()));
-      if (growable)
-        return list;
-      return J.JSArray_markFixedList(list, $E);
-    },
-    RegExp_RegExp(source) {
-      return new A.JSSyntaxRegExp(source, A.JSSyntaxRegExp_makeNative(source, false, true, false, false, false));
-    },
-    StringBuffer__writeAll(string, objects, separator) {
-      var iterator = J.get$iterator$ax(objects);
-      if (!iterator.moveNext$0())
-        return string;
-      if (separator.length === 0) {
-        do
-          string += A.S(iterator.get$current());
-        while (iterator.moveNext$0());
-      } else {
-        string += A.S(iterator.get$current());
-        for (; iterator.moveNext$0();)
-          string = string + separator + A.S(iterator.get$current());
-      }
-      return string;
-    },
-    Error_safeToString(object) {
-      if (typeof object == "number" || A._isBool(object) || object == null)
-        return J.toString$0$(object);
-      if (typeof object == "string")
-        return JSON.stringify(object);
-      return A.Primitives_safeToString(object);
-    },
-    AssertionError$(message) {
-      return new A.AssertionError(message);
-    },
-    ArgumentError$(message) {
-      return new A.ArgumentError(false, null, null, message);
-    },
-    ArgumentError$value(value, $name, message) {
-      return new A.ArgumentError(true, value, $name, message);
-    },
-    RangeError$range(invalidValue, minValue, maxValue, $name, message) {
-      return new A.RangeError(minValue, maxValue, true, invalidValue, $name, "Invalid value");
-    },
-    RangeError_checkValidRange(start, end, $length) {
-      if (0 > start || start > $length)
-        throw A.wrapException(A.RangeError$range(start, 0, $length, "start", null));
-      if (end != null) {
-        if (start > end || end > $length)
-          throw A.wrapException(A.RangeError$range(end, start, $length, "end", null));
-        return end;
-      }
-      return $length;
-    },
-    RangeError_checkNotNegative(value, $name) {
-      if (value < 0)
-        throw A.wrapException(A.RangeError$range(value, 0, null, $name, null));
-      return value;
-    },
-    IndexError$withLength(invalidValue, $length, indexable, message, $name) {
-      return new A.IndexError($length, true, invalidValue, $name, "Index out of range");
-    },
-    UnsupportedError$(message) {
-      return new A.UnsupportedError(message);
-    },
-    UnimplementedError$(message) {
-      return new A.UnimplementedError(message);
-    },
-    ConcurrentModificationError$(modifiedObject) {
-      return new A.ConcurrentModificationError(modifiedObject);
-    },
-    Iterable_iterableToShortString(iterable, leftDelimiter, rightDelimiter) {
-      var parts, t1;
-      if (A.isToStringVisiting(iterable)) {
-        if (leftDelimiter === "(" && rightDelimiter === ")")
-          return "(...)";
-        return leftDelimiter + "..." + rightDelimiter;
-      }
-      parts = A._setArrayType([], type$.JSArray_String);
-      B.JSArray_methods.add$1($.toStringVisiting, iterable);
-      try {
-        A._iterablePartsToStrings(iterable, parts);
-      } finally {
-        if (0 >= $.toStringVisiting.length)
-          return A.ioore($.toStringVisiting, -1);
-        $.toStringVisiting.pop();
-      }
-      t1 = A.StringBuffer__writeAll(leftDelimiter, type$.Iterable_dynamic._as(parts), ", ") + rightDelimiter;
-      return t1.charCodeAt(0) == 0 ? t1 : t1;
-    },
-    Iterable_iterableToFullString(iterable, leftDelimiter, rightDelimiter) {
-      var buffer, t1;
-      if (A.isToStringVisiting(iterable))
-        return leftDelimiter + "..." + rightDelimiter;
-      buffer = new A.StringBuffer(leftDelimiter);
-      B.JSArray_methods.add$1($.toStringVisiting, iterable);
-      try {
-        t1 = buffer;
-        t1._contents = A.StringBuffer__writeAll(t1._contents, iterable, ", ");
-      } finally {
-        if (0 >= $.toStringVisiting.length)
-          return A.ioore($.toStringVisiting, -1);
-        $.toStringVisiting.pop();
-      }
-      buffer._contents += rightDelimiter;
-      t1 = buffer._contents;
-      return t1.charCodeAt(0) == 0 ? t1 : t1;
-    },
-    _iterablePartsToStrings(iterable, parts) {
-      var next, ultimateString, penultimateString, penultimate, ultimate, ultimate0, elision,
-        it = iterable.get$iterator(iterable),
-        $length = 0, count = 0;
-      while (true) {
-        if (!($length < 80 || count < 3))
-          break;
-        if (!it.moveNext$0())
-          return;
-        next = A.S(it.get$current());
-        B.JSArray_methods.add$1(parts, next);
-        $length += next.length + 2;
-        ++count;
-      }
-      if (!it.moveNext$0()) {
-        if (count <= 5)
-          return;
-        if (0 >= parts.length)
-          return A.ioore(parts, -1);
-        ultimateString = parts.pop();
-        if (0 >= parts.length)
-          return A.ioore(parts, -1);
-        penultimateString = parts.pop();
-      } else {
-        penultimate = it.get$current();
-        ++count;
-        if (!it.moveNext$0()) {
-          if (count <= 4) {
-            B.JSArray_methods.add$1(parts, A.S(penultimate));
-            return;
-          }
-          ultimateString = A.S(penultimate);
-          if (0 >= parts.length)
-            return A.ioore(parts, -1);
-          penultimateString = parts.pop();
-          $length += ultimateString.length + 2;
-        } else {
-          ultimate = it.get$current();
-          ++count;
-          for (; it.moveNext$0(); penultimate = ultimate, ultimate = ultimate0) {
-            ultimate0 = it.get$current();
-            ++count;
-            if (count > 100) {
-              while (true) {
-                if (!($length > 75 && count > 3))
-                  break;
-                if (0 >= parts.length)
-                  return A.ioore(parts, -1);
-                $length -= parts.pop().length + 2;
-                --count;
-              }
-              B.JSArray_methods.add$1(parts, "...");
-              return;
-            }
-          }
-          penultimateString = A.S(penultimate);
-          ultimateString = A.S(ultimate);
-          $length += ultimateString.length + penultimateString.length + 4;
-        }
-      }
-      if (count > parts.length + 2) {
-        $length += 5;
-        elision = "...";
-      } else
-        elision = null;
-      while (true) {
-        if (!($length > 80 && parts.length > 3))
-          break;
-        if (0 >= parts.length)
-          return A.ioore(parts, -1);
-        $length -= parts.pop().length + 2;
-        if (elision == null) {
-          $length += 5;
-          elision = "...";
-        }
-      }
-      if (elision != null)
-        B.JSArray_methods.add$1(parts, elision);
-      B.JSArray_methods.add$1(parts, penultimateString);
-      B.JSArray_methods.add$1(parts, ultimateString);
-    },
-    print(object) {
-      A.printString(object);
-    },
-    Error: function Error() {
-    },
-    AssertionError: function AssertionError(t0) {
-      this.message = t0;
-    },
-    TypeError: function TypeError() {
-    },
-    ArgumentError: function ArgumentError(t0, t1, t2, t3) {
-      var _ = this;
-      _._hasValue = t0;
-      _.invalidValue = t1;
-      _.name = t2;
-      _.message = t3;
-    },
-    RangeError: function RangeError(t0, t1, t2, t3, t4, t5) {
-      var _ = this;
-      _.start = t0;
-      _.end = t1;
-      _._hasValue = t2;
-      _.invalidValue = t3;
-      _.name = t4;
-      _.message = t5;
-    },
-    IndexError: function IndexError(t0, t1, t2, t3, t4) {
-      var _ = this;
-      _.length = t0;
-      _._hasValue = t1;
-      _.invalidValue = t2;
-      _.name = t3;
-      _.message = t4;
-    },
-    UnsupportedError: function UnsupportedError(t0) {
-      this.message = t0;
-    },
-    UnimplementedError: function UnimplementedError(t0) {
-      this.message = t0;
-    },
-    ConcurrentModificationError: function ConcurrentModificationError(t0) {
-      this.modifiedObject = t0;
-    },
-    _Exception: function _Exception(t0) {
-      this.message = t0;
-    },
-    FormatException: function FormatException(t0, t1) {
-      this.message = t0;
-      this.source = t1;
-    },
-    Iterable: function Iterable() {
-    },
-    Null: function Null() {
-    },
-    Object: function Object() {
-    },
-    StringBuffer: function StringBuffer(t0) {
-      this._contents = t0;
-    },
-    HtmlElement: function HtmlElement() {
-    },
-    AnchorElement: function AnchorElement() {
-    },
-    AreaElement: function AreaElement() {
-    },
-    ButtonElement: function ButtonElement() {
-    },
-    CharacterData: function CharacterData() {
-    },
-    DomException: function DomException() {
-    },
-    DomTokenList: function DomTokenList() {
-    },
-    _ChildrenElementList: function _ChildrenElementList(t0, t1) {
-      this._element = t0;
-      this._childElements = t1;
-    },
-    Element: function Element() {
-    },
-    Event: function Event() {
-    },
-    EventTarget: function EventTarget() {
-    },
-    FormElement: function FormElement() {
-    },
-    HtmlCollection: function HtmlCollection() {
-    },
-    ImageElement: function ImageElement() {
-    },
-    _ChildNodeListLazy: function _ChildNodeListLazy(t0) {
-      this._this = t0;
-    },
-    Node: function Node() {
-    },
-    NodeList: function NodeList() {
-    },
-    SelectElement: function SelectElement() {
-    },
-    _NamedNodeMap: function _NamedNodeMap() {
-    },
-    _ElementCssClassSet: function _ElementCssClassSet(t0) {
-      this._element = t0;
-    },
-    ImmutableListMixin: function ImmutableListMixin() {
-    },
-    FixedSizeListIterator: function FixedSizeListIterator(t0, t1, t2) {
-      var _ = this;
-      _._array = t0;
-      _._html$_length = t1;
-      _._position = -1;
-      _._html$_current = null;
-      _.$ti = t2;
-    },
-    _HtmlCollection_JavaScriptObject_ListMixin: function _HtmlCollection_JavaScriptObject_ListMixin() {
-    },
-    _HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin: function _HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin() {
-    },
-    _NodeList_JavaScriptObject_ListMixin: function _NodeList_JavaScriptObject_ListMixin() {
-    },
-    _NodeList_JavaScriptObject_ListMixin_ImmutableListMixin: function _NodeList_JavaScriptObject_ListMixin_ImmutableListMixin() {
-    },
-    __NamedNodeMap_JavaScriptObject_ListMixin: function __NamedNodeMap_JavaScriptObject_ListMixin() {
-    },
-    __NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin: function __NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin() {
-    },
-    CssClassSetImpl: function CssClassSetImpl() {
-    },
-    CssClassSetImpl_add_closure: function CssClassSetImpl_add_closure(t0) {
-      this.value = t0;
-    },
-    FilteredElementList: function FilteredElementList(t0) {
-      this._childNodes = t0;
-    },
-    FilteredElementList__iterable_closure: function FilteredElementList__iterable_closure() {
-    },
-    FilteredElementList__iterable_closure0: function FilteredElementList__iterable_closure0() {
-    },
-    AttributeClassSet: function AttributeClassSet(t0) {
-      this._svg$_element = t0;
-    },
-    SvgElement: function SvgElement() {
-    },
-    ChessPieceFactory: function ChessPieceFactory() {
-    },
-    BlackPieceFactory: function BlackPieceFactory(t0) {
-      this.colour = t0;
-    },
-    WhitePieceFactory: function WhitePieceFactory(t0) {
-      this.colour = t0;
-    },
-    Pawn: function Pawn() {
-    },
-    Knight: function Knight() {
-    },
-    Bishop: function Bishop() {
-    },
-    Rook: function Rook() {
-    },
-    King: function King() {
-    },
-    Queen: function Queen() {
-    },
-    BlackPawn: function BlackPawn(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    BlackKnight: function BlackKnight(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    BlackBishop: function BlackBishop(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    BlackRook: function BlackRook(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    BlackKing: function BlackKing(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    BlackQueen: function BlackQueen(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    ChequeredBoard: function ChequeredBoard(t0) {
-      this.pieces = t0;
-    },
-    BoardWithPieces: function BoardWithPieces() {
-    },
-    BoardWithPawns: function BoardWithPawns(t0) {
-      this.base = t0;
-    },
-    BoardWithBishops: function BoardWithBishops(t0) {
-      this.base = t0;
-    },
-    BoardWithKnights: function BoardWithKnights(t0) {
-      this.base = t0;
-    },
-    BoardWithRooks: function BoardWithRooks(t0) {
-      this.base = t0;
-    },
-    BoardWithKings: function BoardWithKings(t0) {
-      this.base = t0;
-    },
-    BoardWithQueens: function BoardWithQueens(t0) {
-      this.base = t0;
-    },
-    ChessGame: function ChessGame(t0) {
-      var _ = this;
-      _.turnCount = 0;
-      _.board = t0;
-      _.__ChessGame_logic_A = _.__ChessGame_view_A = $;
-    },
-    ChessLogic: function ChessLogic(t0, t1) {
-      var _ = this;
-      _.game = t0;
-      _.whiteKing = _.blackKing = null;
-      _.activePiece = t1;
-    },
-    ChessPiece$(colour, $name, moveStrategy) {
-      var t1 = new A.ChessPiece(moveStrategy, colour, $name);
-      t1.__GamePiece_src_A = "./assets/chess/" + $name + "_" + colour + ".png";
-      return t1;
-    },
-    ChessKing$(colour) {
-      var t1 = new A.ChessKing(new A.KingMovement(), colour, "king");
-      t1.__GamePiece_src_A = "./assets/chess/king_" + colour + ".png";
-      return t1;
-    },
-    EmptyPiece$(iPosition, jPosition) {
-      var t1 = new A.EmptyPiece(new A.NoMovement(), "empty", "empty");
-      t1.__GamePiece_src_A = "./assets/chess/empty_empty.png";
-      t1.i = iPosition;
-      t1.j = jPosition;
-      return t1;
-    },
-    ChessPiece: function ChessPiece(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    ChessKing: function ChessKing(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    EmptyPiece: function EmptyPiece(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    NoMovement: function NoMovement() {
-    },
-    PawnMovement: function PawnMovement() {
-    },
-    KnightMovement: function KnightMovement() {
-    },
-    BishopMovement: function BishopMovement() {
-    },
-    RookMovement: function RookMovement() {
-    },
-    QueenMovement: function QueenMovement() {
-    },
-    KingMovement: function KingMovement() {
-    },
-    ChessBoardView: function ChessBoardView(t0, t1) {
-      this.container = t0;
-      this.game = t1;
-    },
-    ChessBoardView_createTile_closure: function ChessBoardView_createTile_closure(t0, t1) {
-      this.$this = t0;
-      this.piece = t1;
-    },
-    GamePiece: function GamePiece() {
-    },
-    PieceOption$($parent, piece) {
-      var t1 = new A.PieceOption($parent, piece);
-      t1.__PieceOption_element_A = t1.createElement$0(0);
-      return t1;
-    },
-    main() {
-      var t3, game, initState, switchBtn, pieceContainer,
-        t1 = document,
-        gameContainer = t1.getElementById("game-container"),
-        t2 = type$.Element;
-      if (t2._is(gameContainer)) {
-        t3 = J.JSArray_JSArray$growable(0, type$.List_GamePiece);
-        game = new A.ChessGame(new A.ChequeredBoard(t3));
-        game.__ChessGame_view_A = new A.ChessBoardView(gameContainer, game);
-        game.__ChessGame_logic_A = new A.ChessLogic(game, A.EmptyPiece$(0, 0));
-        t3 = type$.bool;
-        initState = J.JSArray_JSArray$growable(0, t3);
-        B.JSArray_methods.add$1(initState, true);
-        B.JSArray_methods.add$1(initState, true);
-        B.JSArray_methods.add$1(initState, true);
-        B.JSArray_methods.add$1(initState, true);
-        B.JSArray_methods.add$1(initState, true);
-        B.JSArray_methods.add$1(initState, true);
-        game.setupPieces$1(initState);
-        game.__ChessGame_logic_A.pairKings$0();
-        game.refreshView$0();
-        initState = J.JSArray_JSArray$growable(0, t3);
-        B.JSArray_methods.add$1(initState, false);
-        B.JSArray_methods.add$1(initState, false);
-        B.JSArray_methods.add$1(initState, false);
-        B.JSArray_methods.add$1(initState, false);
-        B.JSArray_methods.add$1(initState, false);
-        B.JSArray_methods.add$1(initState, false);
-        game.setupPieces$1(initState);
-        switchBtn = t1.querySelector(".colour-switcher");
-        pieceContainer = t1.getElementById("piece-box");
-        if (type$.ButtonElement._is(switchBtn) && t2._is(pieceContainer)) {
-          t1 = new A.PieceBox(game, pieceContainer, switchBtn, new A.BlackPieceFactory("black"));
-          t1.armButton$0();
-          t1.switchColour$0();
-        }
-      }
-    },
-    PieceBox: function PieceBox(t0, t1, t2, t3) {
-      var _ = this;
-      _.game = t0;
-      _.container = t1;
-      _.button = t2;
-      _.factory = t3;
-    },
-    PieceBox_armButton_closure: function PieceBox_armButton_closure(t0) {
-      this.$this = t0;
-    },
-    PieceOption: function PieceOption(t0, t1) {
-      this.parent = t0;
-      this.piece = t1;
-      this.__PieceOption_element_A = $;
-    },
-    PieceOption_createElement_closure: function PieceOption_createElement_closure(t0) {
-      this.$this = t0;
-    },
-    WhitePawn: function WhitePawn(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    WhiteKnight: function WhiteKnight(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    WhiteBishop: function WhiteBishop(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    WhiteRook: function WhiteRook(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    WhiteKing: function WhiteKing(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    WhiteQueen: function WhiteQueen(t0, t1, t2) {
-      var _ = this;
-      _.moveStrategy = t0;
-      _.colour = t1;
-      _.name = t2;
-      _.threatened = _.hasMoved = false;
-      _.myKing = null;
-      _.__GamePiece_src_A = $;
-      _.j = _.i = 0;
-    },
-    printString(string) {
-      if (typeof dartPrint == "function") {
-        dartPrint(string);
-        return;
-      }
-      if (typeof console == "object" && typeof console.log != "undefined") {
-        console.log(string);
-        return;
-      }
-      if (typeof print == "function") {
-        print(string);
-        return;
-      }
-      throw "Unable to print message: " + String(string);
-    },
-    throwLateFieldNI(fieldName) {
-      return A.throwExpression(A.LateError$fieldNI(fieldName));
-    },
-    throwLateFieldADI(fieldName) {
-      return A.throwExpression(new A.LateError("Field '" + fieldName + "' has been assigned during initialization."));
-    }
-  },
-  J = {
-    makeDispatchRecord(interceptor, proto, extension, indexability) {
-      return {i: interceptor, p: proto, e: extension, x: indexability};
-    },
-    getNativeInterceptor(object) {
-      var proto, objectProto, $constructor, interceptor, t1,
-        record = object[init.dispatchPropertyName];
-      if (record == null)
-        if ($.initNativeDispatchFlag == null) {
-          A.initNativeDispatch();
-          record = object[init.dispatchPropertyName];
-        }
-      if (record != null) {
-        proto = record.p;
-        if (false === proto)
-          return record.i;
-        if (true === proto)
-          return object;
-        objectProto = Object.getPrototypeOf(object);
-        if (proto === objectProto)
-          return record.i;
-        if (record.e === objectProto)
-          throw A.wrapException(A.UnimplementedError$("Return interceptor for " + A.S(proto(object, record))));
-      }
-      $constructor = object.constructor;
-      if ($constructor == null)
-        interceptor = null;
-      else {
-        t1 = $._JS_INTEROP_INTERCEPTOR_TAG;
-        if (t1 == null)
-          t1 = $._JS_INTEROP_INTERCEPTOR_TAG = init.getIsolateTag("_$dart_js");
-        interceptor = $constructor[t1];
-      }
-      if (interceptor != null)
-        return interceptor;
-      interceptor = A.lookupAndCacheInterceptor(object);
-      if (interceptor != null)
-        return interceptor;
-      if (typeof object == "function")
-        return B.JavaScriptFunction_methods;
-      proto = Object.getPrototypeOf(object);
-      if (proto == null)
-        return B.PlainJavaScriptObject_methods;
-      if (proto === Object.prototype)
-        return B.PlainJavaScriptObject_methods;
-      if (typeof $constructor == "function") {
-        t1 = $._JS_INTEROP_INTERCEPTOR_TAG;
-        if (t1 == null)
-          t1 = $._JS_INTEROP_INTERCEPTOR_TAG = init.getIsolateTag("_$dart_js");
-        Object.defineProperty($constructor, t1, {value: B.UnknownJavaScriptObject_methods, enumerable: false, writable: true, configurable: true});
-        return B.UnknownJavaScriptObject_methods;
-      }
-      return B.UnknownJavaScriptObject_methods;
-    },
-    JSArray_JSArray$fixed($length, $E) {
-      if ($length < 0 || $length > 4294967295)
-        throw A.wrapException(A.RangeError$range($length, 0, 4294967295, "length", null));
-      return J.JSArray_JSArray$markFixed(new Array($length), $E);
-    },
-    JSArray_JSArray$growable($length, $E) {
-      if ($length < 0)
-        throw A.wrapException(A.ArgumentError$("Length must be a non-negative integer: " + $length));
-      return A._setArrayType(new Array($length), $E._eval$1("JSArray<0>"));
-    },
-    JSArray_JSArray$markFixed(allocation, $E) {
-      return J.JSArray_markFixedList(A._setArrayType(allocation, $E._eval$1("JSArray<0>")), $E);
-    },
-    JSArray_markFixedList(list, $T) {
-      list.fixed$length = Array;
-      return list;
-    },
-    JSString__isWhitespace(codeUnit) {
-      if (codeUnit < 256)
-        switch (codeUnit) {
-          case 9:
-          case 10:
-          case 11:
-          case 12:
-          case 13:
-          case 32:
-          case 133:
-          case 160:
-            return true;
-          default:
-            return false;
-        }
-      switch (codeUnit) {
-        case 5760:
-        case 8192:
-        case 8193:
-        case 8194:
-        case 8195:
-        case 8196:
-        case 8197:
-        case 8198:
-        case 8199:
-        case 8200:
-        case 8201:
-        case 8202:
-        case 8232:
-        case 8233:
-        case 8239:
-        case 8287:
-        case 12288:
-        case 65279:
-          return true;
-        default:
-          return false;
-      }
-    },
-    JSString__skipLeadingWhitespace(string, index) {
-      var t1, codeUnit;
-      for (t1 = string.length; index < t1;) {
-        codeUnit = B.JSString_methods._codeUnitAt$1(string, index);
-        if (codeUnit !== 32 && codeUnit !== 13 && !J.JSString__isWhitespace(codeUnit))
-          break;
-        ++index;
-      }
-      return index;
-    },
-    JSString__skipTrailingWhitespace(string, index) {
-      var index0, codeUnit;
-      for (; index > 0; index = index0) {
-        index0 = index - 1;
-        codeUnit = B.JSString_methods.codeUnitAt$1(string, index0);
-        if (codeUnit !== 32 && codeUnit !== 13 && !J.JSString__isWhitespace(codeUnit))
-          break;
-      }
-      return index;
-    },
-    getInterceptor$(receiver) {
-      if (typeof receiver == "number") {
-        if (Math.floor(receiver) == receiver)
-          return J.JSInt.prototype;
-        return J.JSNumNotInt.prototype;
-      }
-      if (typeof receiver == "string")
-        return J.JSString.prototype;
-      if (receiver == null)
-        return J.JSNull.prototype;
-      if (typeof receiver == "boolean")
-        return J.JSBool.prototype;
-      if (receiver.constructor == Array)
-        return J.JSArray.prototype;
-      if (typeof receiver != "object") {
-        if (typeof receiver == "function")
-          return J.JavaScriptFunction.prototype;
-        return receiver;
-      }
-      if (receiver instanceof A.Object)
-        return receiver;
-      return J.getNativeInterceptor(receiver);
-    },
-    getInterceptor$asx(receiver) {
-      if (typeof receiver == "string")
-        return J.JSString.prototype;
-      if (receiver == null)
-        return receiver;
-      if (receiver.constructor == Array)
-        return J.JSArray.prototype;
-      if (typeof receiver != "object") {
-        if (typeof receiver == "function")
-          return J.JavaScriptFunction.prototype;
-        return receiver;
-      }
-      if (receiver instanceof A.Object)
-        return receiver;
-      return J.getNativeInterceptor(receiver);
-    },
-    getInterceptor$ax(receiver) {
-      if (receiver == null)
-        return receiver;
-      if (receiver.constructor == Array)
-        return J.JSArray.prototype;
-      if (typeof receiver != "object") {
-        if (typeof receiver == "function")
-          return J.JavaScriptFunction.prototype;
-        return receiver;
-      }
-      if (receiver instanceof A.Object)
-        return receiver;
-      return J.getNativeInterceptor(receiver);
-    },
-    getInterceptor$in(receiver) {
-      if (typeof receiver == "number") {
-        if (Math.floor(receiver) == receiver)
-          return J.JSInt.prototype;
-        return J.JSNumNotInt.prototype;
-      }
-      if (receiver == null)
-        return receiver;
-      if (!(receiver instanceof A.Object))
-        return J.UnknownJavaScriptObject.prototype;
-      return receiver;
-    },
-    getInterceptor$s(receiver) {
-      if (typeof receiver == "string")
-        return J.JSString.prototype;
-      if (receiver == null)
-        return receiver;
-      if (!(receiver instanceof A.Object))
-        return J.UnknownJavaScriptObject.prototype;
-      return receiver;
-    },
-    getInterceptor$x(receiver) {
-      if (receiver == null)
-        return receiver;
-      if (typeof receiver != "object") {
-        if (typeof receiver == "function")
-          return J.JavaScriptFunction.prototype;
-        return receiver;
-      }
-      if (receiver instanceof A.Object)
-        return receiver;
-      return J.getNativeInterceptor(receiver);
-    },
-    get$children$x(receiver) {
-      return J.getInterceptor$x(receiver).get$children(receiver);
-    },
-    get$classes$x(receiver) {
-      return J.getInterceptor$x(receiver).get$classes(receiver);
-    },
-    get$hashCode$(receiver) {
-      return J.getInterceptor$(receiver).get$hashCode(receiver);
-    },
-    get$iterator$ax(receiver) {
-      return J.getInterceptor$ax(receiver).get$iterator(receiver);
-    },
-    get$length$asx(receiver) {
-      return J.getInterceptor$asx(receiver).get$length(receiver);
-    },
-    get$runtimeType$(receiver) {
-      return J.getInterceptor$(receiver).get$runtimeType(receiver);
-    },
-    $eq$(receiver, a0) {
-      if (receiver == null)
-        return a0 == null;
-      if (typeof receiver != "object")
-        return a0 != null && receiver === a0;
-      return J.getInterceptor$(receiver).$eq(receiver, a0);
-    },
-    $index$ax(receiver, a0) {
-      if (typeof a0 === "number")
-        if (receiver.constructor == Array || A.isJsIndexable(receiver, receiver[init.dispatchPropertyName]))
-          if (a0 >>> 0 === a0 && a0 < receiver.length)
-            return receiver[a0];
-      return J.getInterceptor$ax(receiver).$index(receiver, a0);
-    },
-    _clearChildren$0$x(receiver) {
-      return J.getInterceptor$x(receiver)._clearChildren$0(receiver);
-    },
-    abs$0$in(receiver) {
-      if (typeof receiver === "number")
-        return Math.abs(receiver);
-      return J.getInterceptor$in(receiver).abs$0(receiver);
-    },
-    elementAt$1$ax(receiver, a0) {
-      return J.getInterceptor$ax(receiver).elementAt$1(receiver, a0);
-    },
-    toString$0$(receiver) {
-      return J.getInterceptor$(receiver).toString$0(receiver);
-    },
-    trim$0$s(receiver) {
-      return J.getInterceptor$s(receiver).trim$0(receiver);
-    },
-    Interceptor: function Interceptor() {
-    },
-    JSBool: function JSBool() {
-    },
-    JSNull: function JSNull() {
-    },
-    JavaScriptObject: function JavaScriptObject() {
-    },
-    LegacyJavaScriptObject: function LegacyJavaScriptObject() {
-    },
-    PlainJavaScriptObject: function PlainJavaScriptObject() {
-    },
-    UnknownJavaScriptObject: function UnknownJavaScriptObject() {
-    },
-    JavaScriptFunction: function JavaScriptFunction() {
-    },
-    JSArray: function JSArray(t0) {
-      this.$ti = t0;
-    },
-    JSUnmodifiableArray: function JSUnmodifiableArray(t0) {
-      this.$ti = t0;
-    },
-    ArrayIterator: function ArrayIterator(t0, t1, t2) {
-      var _ = this;
-      _._iterable = t0;
-      _._length = t1;
-      _._index = 0;
-      _._current = null;
-      _.$ti = t2;
-    },
-    JSNumber: function JSNumber() {
-    },
-    JSInt: function JSInt() {
-    },
-    JSNumNotInt: function JSNumNotInt() {
-    },
-    JSString: function JSString() {
-    }
-  },
-  B = {};
-  var holders = [A, J, B];
-  var $ = {};
-  A.JS_CONST.prototype = {};
-  J.Interceptor.prototype = {
-    $eq(receiver, other) {
-      return receiver === other;
-    },
-    get$hashCode(receiver) {
-      return A.Primitives_objectHashCode(receiver);
-    },
-    toString$0(receiver) {
-      return "Instance of '" + A.Primitives_objectTypeName(receiver) + "'";
-    },
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(A._instanceTypeFromConstructor(this));
-    }
-  };
-  J.JSBool.prototype = {
-    toString$0(receiver) {
-      return String(receiver);
-    },
-    $and(receiver, other) {
-      return other && receiver;
-    },
-    get$hashCode(receiver) {
-      return receiver ? 519018 : 218159;
-    },
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(type$.bool);
-    },
-    $isTrustedGetRuntimeType: 1,
-    $isbool: 1
-  };
-  J.JSNull.prototype = {
-    $eq(receiver, other) {
-      return null == other;
-    },
-    toString$0(receiver) {
-      return "null";
-    },
-    get$hashCode(receiver) {
-      return 0;
-    },
-    $isTrustedGetRuntimeType: 1
-  };
-  J.JavaScriptObject.prototype = {};
-  J.LegacyJavaScriptObject.prototype = {
-    get$hashCode(receiver) {
-      return 0;
-    },
-    toString$0(receiver) {
-      return String(receiver);
-    }
-  };
-  J.PlainJavaScriptObject.prototype = {};
-  J.UnknownJavaScriptObject.prototype = {};
-  J.JavaScriptFunction.prototype = {
-    toString$0(receiver) {
-      var dartClosure = receiver[$.$get$DART_CLOSURE_PROPERTY_NAME()];
-      if (dartClosure == null)
-        return this.super$LegacyJavaScriptObject$toString(receiver);
-      return "JavaScript function for " + J.toString$0$(dartClosure);
-    },
-    $isFunction: 1
-  };
-  J.JSArray.prototype = {
-    add$1(receiver, value) {
-      A._arrayInstanceType(receiver)._precomputed1._as(value);
-      if (!!receiver.fixed$length)
-        A.throwExpression(A.UnsupportedError$("add"));
-      receiver.push(value);
-    },
-    elementAt$1(receiver, index) {
-      if (!(index >= 0 && index < receiver.length))
-        return A.ioore(receiver, index);
-      return receiver[index];
-    },
-    toString$0(receiver) {
-      return A.Iterable_iterableToFullString(receiver, "[", "]");
-    },
-    get$iterator(receiver) {
-      return new J.ArrayIterator(receiver, receiver.length, A._arrayInstanceType(receiver)._eval$1("ArrayIterator<1>"));
-    },
-    get$hashCode(receiver) {
-      return A.Primitives_objectHashCode(receiver);
-    },
-    get$length(receiver) {
-      return receiver.length;
-    },
-    $indexSet(receiver, index, value) {
-      A._arrayInstanceType(receiver)._precomputed1._as(value);
-      if (!!receiver.immutable$list)
-        A.throwExpression(A.UnsupportedError$("indexed set"));
-      if (!(index >= 0 && index < receiver.length))
-        throw A.wrapException(A.diagnoseIndexError(receiver, index));
-      receiver[index] = value;
-    },
-    $isIterable: 1,
-    $isList: 1
-  };
-  J.JSUnmodifiableArray.prototype = {};
-  J.ArrayIterator.prototype = {
-    get$current() {
-      var t1 = this._current;
-      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
-    },
-    moveNext$0() {
-      var t2, _this = this,
-        t1 = _this._iterable,
-        $length = t1.length;
-      if (_this._length !== $length) {
-        t1 = A.throwConcurrentModificationError(t1);
-        throw A.wrapException(t1);
-      }
-      t2 = _this._index;
-      if (t2 >= $length) {
-        _this.set$_current(null);
-        return false;
-      }
-      _this.set$_current(t1[t2]);
-      ++_this._index;
-      return true;
-    },
-    set$_current(_current) {
-      this._current = this.$ti._eval$1("1?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  J.JSNumber.prototype = {
-    toString$0(receiver) {
-      if (receiver === 0 && 1 / receiver < 0)
-        return "-0.0";
-      else
-        return "" + receiver;
-    },
-    get$hashCode(receiver) {
-      var absolute, floorLog2, factor, scaled,
-        intValue = receiver | 0;
-      if (receiver === intValue)
-        return intValue & 536870911;
-      absolute = Math.abs(receiver);
-      floorLog2 = Math.log(absolute) / 0.6931471805599453 | 0;
-      factor = Math.pow(2, floorLog2);
-      scaled = absolute < 1 ? absolute / factor : factor / absolute;
-      return ((scaled * 9007199254740992 | 0) + (scaled * 3542243181176521 | 0)) * 599197 + floorLog2 * 1259 & 536870911;
-    },
-    $mod(receiver, other) {
-      var result = receiver % other;
-      if (result === 0)
-        return 0;
-      if (result > 0)
-        return result;
-      return result + other;
-    },
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(type$.num);
-    },
-    $isnum: 1
-  };
-  J.JSInt.prototype = {
-    abs$0(receiver) {
-      return Math.abs(receiver);
-    },
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(type$.int);
-    },
-    $isTrustedGetRuntimeType: 1,
-    $isint: 1
-  };
-  J.JSNumNotInt.prototype = {
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(type$.double);
-    },
-    $isTrustedGetRuntimeType: 1
-  };
-  J.JSString.prototype = {
-    codeUnitAt$1(receiver, index) {
-      if (index < 0)
-        throw A.wrapException(A.diagnoseIndexError(receiver, index));
-      if (index >= receiver.length)
-        A.throwExpression(A.diagnoseIndexError(receiver, index));
-      return receiver.charCodeAt(index);
-    },
-    _codeUnitAt$1(receiver, index) {
-      if (index >= receiver.length)
-        throw A.wrapException(A.diagnoseIndexError(receiver, index));
-      return receiver.charCodeAt(index);
-    },
-    $add(receiver, other) {
-      return receiver + other;
-    },
-    substring$2(receiver, start, end) {
-      return receiver.substring(start, A.RangeError_checkValidRange(start, end, receiver.length));
-    },
-    trim$0(receiver) {
-      var startIndex, t1, endIndex0,
-        result = receiver.trim(),
-        endIndex = result.length;
-      if (endIndex === 0)
-        return result;
-      if (this._codeUnitAt$1(result, 0) === 133) {
-        startIndex = J.JSString__skipLeadingWhitespace(result, 1);
-        if (startIndex === endIndex)
-          return "";
-      } else
-        startIndex = 0;
-      t1 = endIndex - 1;
-      endIndex0 = this.codeUnitAt$1(result, t1) === 133 ? J.JSString__skipTrailingWhitespace(result, t1) : endIndex;
-      if (startIndex === 0 && endIndex0 === endIndex)
-        return result;
-      return result.substring(startIndex, endIndex0);
-    },
-    toString$0(receiver) {
-      return receiver;
-    },
-    get$hashCode(receiver) {
-      var t1, hash, i;
-      for (t1 = receiver.length, hash = 0, i = 0; i < t1; ++i) {
-        hash = hash + receiver.charCodeAt(i) & 536870911;
-        hash = hash + ((hash & 524287) << 10) & 536870911;
-        hash ^= hash >> 6;
-      }
-      hash = hash + ((hash & 67108863) << 3) & 536870911;
-      hash ^= hash >> 11;
-      return hash + ((hash & 16383) << 15) & 536870911;
-    },
-    get$runtimeType(receiver) {
-      return A.createRuntimeType(type$.String);
-    },
-    get$length(receiver) {
-      return receiver.length;
-    },
-    $isTrustedGetRuntimeType: 1,
-    $isString: 1
-  };
-  A.LateError.prototype = {
-    toString$0(_) {
-      return "LateInitializationError: " + this._message;
-    }
-  };
-  A.ListIterator.prototype = {
-    get$current() {
-      var t1 = this.__internal$_current;
-      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
-    },
-    moveNext$0() {
-      var t3, _this = this,
-        t1 = _this.__internal$_iterable,
-        t2 = J.getInterceptor$asx(t1),
-        $length = t2.get$length(t1);
-      if (_this.__internal$_length !== $length)
-        throw A.wrapException(A.ConcurrentModificationError$(t1));
-      t3 = _this.__internal$_index;
-      if (t3 >= $length) {
-        _this.set$__internal$_current(null);
-        return false;
-      }
-      _this.set$__internal$_current(t2.elementAt$1(t1, t3));
-      ++_this.__internal$_index;
-      return true;
-    },
-    set$__internal$_current(_current) {
-      this.__internal$_current = this.$ti._eval$1("1?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  A.MappedIterable.prototype = {
-    get$iterator(_) {
-      var t1 = A._instanceType(this);
-      return new A.MappedIterator(J.get$iterator$ax(this.__internal$_iterable), this._f, t1._eval$1("@<1>")._bind$1(t1._rest[1])._eval$1("MappedIterator<1,2>"));
-    },
-    get$length(_) {
-      return J.get$length$asx(this.__internal$_iterable);
-    },
-    elementAt$1(_, index) {
-      return this._f.call$1(J.elementAt$1$ax(this.__internal$_iterable, index));
-    }
-  };
-  A.MappedIterator.prototype = {
-    moveNext$0() {
-      var _this = this,
-        t1 = _this._iterator;
-      if (t1.moveNext$0()) {
-        _this.set$__internal$_current(_this._f.call$1(t1.get$current()));
-        return true;
-      }
-      _this.set$__internal$_current(null);
-      return false;
-    },
-    get$current() {
-      var t1 = this.__internal$_current;
-      return t1 == null ? this.$ti._rest[1]._as(t1) : t1;
-    },
-    set$__internal$_current(_current) {
-      this.__internal$_current = this.$ti._eval$1("2?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  A.WhereIterable.prototype = {
-    get$iterator(_) {
-      return new A.WhereIterator(J.get$iterator$ax(this.__internal$_iterable), this._f, this.$ti._eval$1("WhereIterator<1>"));
-    }
-  };
-  A.WhereIterator.prototype = {
-    moveNext$0() {
-      var t1, t2;
-      for (t1 = this._iterator, t2 = this._f; t1.moveNext$0();)
-        if (A.boolConversionCheck(t2.call$1(t1.get$current())))
-          return true;
-      return false;
-    },
-    get$current() {
-      return this._iterator.get$current();
-    },
-    $isIterator: 1
-  };
-  A.Closure.prototype = {
-    toString$0(_) {
-      var $constructor = this.constructor,
-        $name = $constructor == null ? null : $constructor.name;
-      return "Closure '" + A.unminifyOrTag($name == null ? "unknown" : $name) + "'";
-    },
-    $isFunction: 1,
-    get$$call() {
-      return this;
-    },
-    "call*": "call$1",
-    $requiredArgCount: 1,
-    $defaultValues: null
-  };
-  A.Closure2Args.prototype = {"call*": "call$2", $requiredArgCount: 2};
-  A.TearOffClosure.prototype = {};
-  A.StaticClosure.prototype = {
-    toString$0(_) {
-      var $name = this.$static_name;
-      if ($name == null)
-        return "Closure of unknown static method";
-      return "Closure '" + A.unminifyOrTag($name) + "'";
-    }
-  };
-  A.BoundClosure.prototype = {
-    $eq(_, other) {
-      if (other == null)
-        return false;
-      if (this === other)
-        return true;
-      if (!(other instanceof A.BoundClosure))
-        return false;
-      return this.$_target === other.$_target && this._receiver === other._receiver;
-    },
-    get$hashCode(_) {
-      return (A.objectHashCode(this._receiver) ^ A.Primitives_objectHashCode(this.$_target)) >>> 0;
-    },
-    toString$0(_) {
-      return "Closure '" + this.$_name + "' of " + ("Instance of '" + A.Primitives_objectTypeName(this._receiver) + "'");
-    }
-  };
-  A._CyclicInitializationError.prototype = {
-    toString$0(_) {
-      return "Reading static variable '" + this.variableName + "' during its initialization";
-    }
-  };
-  A.RuntimeError.prototype = {
-    toString$0(_) {
-      return "RuntimeError: " + this.message;
-    }
-  };
-  A._AssertionError.prototype = {
-    toString$0(_) {
-      return "Assertion failed: " + A.Error_safeToString(this.message);
-    }
-  };
-  A.initHooks_closure.prototype = {
-    call$1(o) {
-      return this.getTag(o);
-    },
-    $signature: 1
-  };
-  A.initHooks_closure0.prototype = {
-    call$2(o, tag) {
-      return this.getUnknownTag(o, tag);
-    },
-    $signature: 2
-  };
-  A.initHooks_closure1.prototype = {
-    call$1(tag) {
-      return this.prototypeForTag(A._asString(tag));
-    },
-    $signature: 3
-  };
-  A.JSSyntaxRegExp.prototype = {
-    toString$0(_) {
-      return "RegExp/" + this.pattern + "/" + this._nativeRegExp.flags;
-    }
-  };
-  A.Rti.prototype = {
-    _eval$1(recipe) {
-      return A._Universe_evalInEnvironment(init.typeUniverse, this, recipe);
-    },
-    _bind$1(typeOrTuple) {
-      return A._Universe_bind(init.typeUniverse, this, typeOrTuple);
-    }
-  };
-  A._FunctionParameters.prototype = {};
-  A._Type.prototype = {
-    toString$0(_) {
-      return A._rtiToString(this._rti, null);
-    }
-  };
-  A._Error.prototype = {
-    toString$0(_) {
-      return this.__rti$_message;
-    }
-  };
-  A._TypeError.prototype = {};
-  A._HashSet.prototype = {
-    get$iterator(_) {
-      return new A._HashSetIterator(this, this._computeElements$0(), A._instanceType(this)._eval$1("_HashSetIterator<1>"));
-    },
-    get$length(_) {
-      return this._collection$_length;
-    },
-    contains$1(_, object) {
-      var strings, t1;
-      if (object !== "__proto__") {
-        strings = this._strings;
-        return strings == null ? false : strings[object] != null;
-      } else {
-        t1 = this._contains$1(object);
-        return t1;
-      }
-    },
-    _contains$1(object) {
-      var rest = this._collection$_rest;
-      if (rest == null)
-        return false;
-      return this._findBucketIndex$2(rest[this._computeHashCode$1(object)], object) >= 0;
-    },
-    add$1(_, element) {
-      var strings, t1, _this = this;
-      A._instanceType(_this)._precomputed1._as(element);
-      if (element !== "__proto__") {
-        strings = _this._strings;
-        return _this._addHashTableEntry$2(strings == null ? _this._strings = A._HashSet__newHashTable() : strings, element);
-      } else {
-        t1 = _this._add$1(element);
-        return t1;
-      }
-    },
-    _add$1(element) {
-      var rest, hash, bucket, _this = this;
-      A._instanceType(_this)._precomputed1._as(element);
-      rest = _this._collection$_rest;
-      if (rest == null)
-        rest = _this._collection$_rest = A._HashSet__newHashTable();
-      hash = _this._computeHashCode$1(element);
-      bucket = rest[hash];
-      if (bucket == null)
-        rest[hash] = [element];
-      else {
-        if (_this._findBucketIndex$2(bucket, element) >= 0)
-          return false;
-        bucket.push(element);
-      }
-      ++_this._collection$_length;
-      _this._elements = null;
-      return true;
-    },
-    _computeElements$0() {
-      var strings, names, entries, index, i, nums, rest, bucket, $length, i0, _this = this,
-        result = _this._elements;
-      if (result != null)
-        return result;
-      result = A.List_List$filled(_this._collection$_length, null, false, type$.dynamic);
-      strings = _this._strings;
-      if (strings != null) {
-        names = Object.getOwnPropertyNames(strings);
-        entries = names.length;
-        for (index = 0, i = 0; i < entries; ++i) {
-          result[index] = names[i];
-          ++index;
-        }
-      } else
-        index = 0;
-      nums = _this._nums;
-      if (nums != null) {
-        names = Object.getOwnPropertyNames(nums);
-        entries = names.length;
-        for (i = 0; i < entries; ++i) {
-          result[index] = +names[i];
-          ++index;
-        }
-      }
-      rest = _this._collection$_rest;
-      if (rest != null) {
-        names = Object.getOwnPropertyNames(rest);
-        entries = names.length;
-        for (i = 0; i < entries; ++i) {
-          bucket = rest[names[i]];
-          $length = bucket.length;
-          for (i0 = 0; i0 < $length; ++i0) {
-            result[index] = bucket[i0];
-            ++index;
-          }
-        }
-      }
-      return _this._elements = result;
-    },
-    _addHashTableEntry$2(table, element) {
-      A._instanceType(this)._precomputed1._as(element);
-      if (table[element] != null)
-        return false;
-      table[element] = 0;
-      ++this._collection$_length;
-      this._elements = null;
-      return true;
-    },
-    _computeHashCode$1(element) {
-      return B.JSString_methods.get$hashCode(element) & 1073741823;
-    },
-    _findBucketIndex$2(bucket, element) {
-      var $length, i;
-      if (bucket == null)
-        return -1;
-      $length = bucket.length;
-      for (i = 0; i < $length; ++i)
-        if (J.$eq$(bucket[i], element))
-          return i;
-      return -1;
-    }
-  };
-  A._HashSetIterator.prototype = {
-    get$current() {
-      var t1 = this._collection$_current;
-      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
-    },
-    moveNext$0() {
-      var _this = this,
-        elements = _this._elements,
-        offset = _this._offset,
-        t1 = _this._set;
-      if (elements !== t1._elements)
-        throw A.wrapException(A.ConcurrentModificationError$(t1));
-      else if (offset >= elements.length) {
-        _this.set$_collection$_current(null);
-        return false;
-      } else {
-        _this.set$_collection$_current(elements[offset]);
-        _this._offset = offset + 1;
-        return true;
-      }
-    },
-    set$_collection$_current(_current) {
-      this._collection$_current = this.$ti._eval$1("1?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  A._LinkedHashSet.prototype = {
-    get$iterator(_) {
-      var _this = this,
-        t1 = new A._LinkedHashSetIterator(_this, _this._modifications, A._instanceType(_this)._eval$1("_LinkedHashSetIterator<1>"));
-      t1._cell = _this._first;
-      return t1;
-    },
-    get$length(_) {
-      return this._collection$_length;
-    },
-    add$1(_, element) {
-      var strings, nums, _this = this;
-      A._instanceType(_this)._precomputed1._as(element);
-      if (typeof element == "string" && element !== "__proto__") {
-        strings = _this._strings;
-        return _this._addHashTableEntry$2(strings == null ? _this._strings = A._LinkedHashSet__newHashTable() : strings, element);
-      } else if (typeof element == "number" && (element & 1073741823) === element) {
-        nums = _this._nums;
-        return _this._addHashTableEntry$2(nums == null ? _this._nums = A._LinkedHashSet__newHashTable() : nums, element);
-      } else
-        return _this._add$1(element);
-    },
-    _add$1(element) {
-      var rest, hash, bucket, _this = this;
-      A._instanceType(_this)._precomputed1._as(element);
-      rest = _this._collection$_rest;
-      if (rest == null)
-        rest = _this._collection$_rest = A._LinkedHashSet__newHashTable();
-      hash = _this._computeHashCode$1(element);
-      bucket = rest[hash];
-      if (bucket == null)
-        rest[hash] = [_this._newLinkedCell$1(element)];
-      else {
-        if (_this._findBucketIndex$2(bucket, element) >= 0)
-          return false;
-        bucket.push(_this._newLinkedCell$1(element));
-      }
-      return true;
-    },
-    _addHashTableEntry$2(table, element) {
-      A._instanceType(this)._precomputed1._as(element);
-      if (type$.nullable__LinkedHashSetCell._as(table[element]) != null)
-        return false;
-      table[element] = this._newLinkedCell$1(element);
-      return true;
-    },
-    _modified$0() {
-      this._modifications = this._modifications + 1 & 1073741823;
-    },
-    _newLinkedCell$1(element) {
-      var t1, _this = this,
-        cell = new A._LinkedHashSetCell(A._instanceType(_this)._precomputed1._as(element));
-      if (_this._first == null)
-        _this._first = _this._last = cell;
-      else {
-        t1 = _this._last;
-        t1.toString;
-        cell._previous = t1;
-        _this._last = t1._next = cell;
-      }
-      ++_this._collection$_length;
-      _this._modified$0();
-      return cell;
-    },
-    _computeHashCode$1(element) {
-      return J.get$hashCode$(element) & 1073741823;
-    },
-    _findBucketIndex$2(bucket, element) {
-      var $length, i;
-      if (bucket == null)
-        return -1;
-      $length = bucket.length;
-      for (i = 0; i < $length; ++i)
-        if (J.$eq$(bucket[i]._collection$_element, element))
-          return i;
-      return -1;
-    },
-    $isLinkedHashSet: 1
-  };
-  A._LinkedHashSetCell.prototype = {};
-  A._LinkedHashSetIterator.prototype = {
-    get$current() {
-      var t1 = this._collection$_current;
-      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
-    },
-    moveNext$0() {
-      var _this = this,
-        cell = _this._cell,
-        t1 = _this._set;
-      if (_this._modifications !== t1._modifications)
-        throw A.wrapException(A.ConcurrentModificationError$(t1));
-      else if (cell == null) {
-        _this.set$_collection$_current(null);
-        return false;
-      } else {
-        _this.set$_collection$_current(_this.$ti._eval$1("1?")._as(cell._collection$_element));
-        _this._cell = cell._next;
-        return true;
-      }
-    },
-    set$_collection$_current(_current) {
-      this._collection$_current = this.$ti._eval$1("1?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  A.ListBase.prototype = {
-    get$iterator(receiver) {
-      return new A.ListIterator(receiver, this.get$length(receiver), A.instanceType(receiver)._eval$1("ListIterator<ListBase.E>"));
-    },
-    elementAt$1(receiver, index) {
-      return this.$index(receiver, index);
-    },
-    get$isEmpty(receiver) {
-      return this.get$length(receiver) === 0;
-    },
-    toList$1$growable(receiver, growable) {
-      var t1, first, result, i, _this = this;
-      if (_this.get$isEmpty(receiver)) {
-        t1 = J.JSArray_JSArray$growable(0, A.instanceType(receiver)._eval$1("ListBase.E"));
-        return t1;
-      }
-      first = _this.$index(receiver, 0);
-      result = A.List_List$filled(_this.get$length(receiver), first, true, A.instanceType(receiver)._eval$1("ListBase.E"));
-      for (i = 1; i < _this.get$length(receiver); ++i)
-        B.JSArray_methods.$indexSet(result, i, _this.$index(receiver, i));
-      return result;
-    },
-    toList$0($receiver) {
-      return this.toList$1$growable($receiver, true);
-    },
-    toString$0(receiver) {
-      return A.Iterable_iterableToFullString(receiver, "[", "]");
-    },
-    $isIterable: 1,
-    $isList: 1
-  };
-  A.SetBase.prototype = {
-    toString$0(_) {
-      return A.Iterable_iterableToFullString(this, "{", "}");
-    },
-    join$1(_, separator) {
-      var first, t1,
-        iterator = this.get$iterator(this);
-      if (!iterator.moveNext$0())
-        return "";
-      first = J.toString$0$(iterator.get$current());
-      if (!iterator.moveNext$0())
-        return first;
-      if (separator.length === 0) {
-        t1 = first;
-        do
-          t1 += A.S(iterator.get$current());
-        while (iterator.moveNext$0());
-      } else {
-        t1 = first;
-        do
-          t1 = t1 + separator + A.S(iterator.get$current());
-        while (iterator.moveNext$0());
-      }
-      return t1.charCodeAt(0) == 0 ? t1 : t1;
-    },
-    elementAt$1(_, index) {
-      var iterator, skipCount;
-      A.RangeError_checkNotNegative(index, "index");
-      iterator = this.get$iterator(this);
-      for (skipCount = index; iterator.moveNext$0();) {
-        if (skipCount === 0)
-          return iterator.get$current();
-        --skipCount;
-      }
-      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, null, "index"));
-    },
-    $isIterable: 1,
-    $isSet: 1
-  };
-  A._SetBase.prototype = {};
-  A.Error.prototype = {};
-  A.AssertionError.prototype = {
-    toString$0(_) {
-      var t1 = this.message;
-      if (t1 != null)
-        return "Assertion failed: " + A.Error_safeToString(t1);
-      return "Assertion failed";
-    }
-  };
-  A.TypeError.prototype = {};
-  A.ArgumentError.prototype = {
-    get$_errorName() {
-      return "Invalid argument" + (!this._hasValue ? "(s)" : "");
-    },
-    get$_errorExplanation() {
-      return "";
-    },
-    toString$0(_) {
-      var _this = this,
-        $name = _this.name,
-        nameString = $name == null ? "" : " (" + $name + ")",
-        message = _this.message,
-        messageString = message == null ? "" : ": " + message,
-        prefix = _this.get$_errorName() + nameString + messageString;
-      if (!_this._hasValue)
-        return prefix;
-      return prefix + _this.get$_errorExplanation() + ": " + A.Error_safeToString(_this.get$invalidValue());
-    },
-    get$invalidValue() {
-      return this.invalidValue;
-    }
-  };
-  A.RangeError.prototype = {
-    get$invalidValue() {
-      return A._asNumQ(this.invalidValue);
-    },
-    get$_errorName() {
-      return "RangeError";
-    },
-    get$_errorExplanation() {
-      var explanation,
-        start = this.start,
-        end = this.end;
-      if (start == null)
-        explanation = end != null ? ": Not less than or equal to " + A.S(end) : "";
-      else if (end == null)
-        explanation = ": Not greater than or equal to " + A.S(start);
-      else if (end > start)
-        explanation = ": Not in inclusive range " + A.S(start) + ".." + A.S(end);
-      else
-        explanation = end < start ? ": Valid value range is empty" : ": Only valid value is " + A.S(start);
-      return explanation;
-    }
-  };
-  A.IndexError.prototype = {
-    get$invalidValue() {
-      return A._asInt(this.invalidValue);
-    },
-    get$_errorName() {
-      return "RangeError";
-    },
-    get$_errorExplanation() {
-      if (A._asInt(this.invalidValue) < 0)
-        return ": index must not be negative";
-      var t1 = this.length;
-      if (t1 === 0)
-        return ": no indices are valid";
-      return ": index should be less than " + t1;
-    },
-    get$length(receiver) {
-      return this.length;
-    }
-  };
-  A.UnsupportedError.prototype = {
-    toString$0(_) {
-      return "Unsupported operation: " + this.message;
-    }
-  };
-  A.UnimplementedError.prototype = {
-    toString$0(_) {
-      var message = this.message;
-      return message != null ? "UnimplementedError: " + message : "UnimplementedError";
-    }
-  };
-  A.ConcurrentModificationError.prototype = {
-    toString$0(_) {
-      var t1 = this.modifiedObject;
-      if (t1 == null)
-        return "Concurrent modification during iteration.";
-      return "Concurrent modification during iteration: " + A.Error_safeToString(t1) + ".";
-    }
-  };
-  A._Exception.prototype = {
-    toString$0(_) {
-      return "Exception: " + this.message;
-    }
-  };
-  A.FormatException.prototype = {
-    toString$0(_) {
-      var message = this.message,
-        report = "" !== message ? "FormatException: " + message : "FormatException",
-        source = this.source;
-      if (source.length > 78)
-        source = B.JSString_methods.substring$2(source, 0, 75) + "...";
-      return report + "\n" + source;
-    }
-  };
-  A.Iterable.prototype = {
-    get$length(_) {
-      var count,
-        it = this.get$iterator(this);
-      for (count = 0; it.moveNext$0();)
-        ++count;
-      return count;
-    },
-    elementAt$1(_, index) {
-      var iterator, skipCount;
-      A.RangeError_checkNotNegative(index, "index");
-      iterator = this.get$iterator(this);
-      for (skipCount = index; iterator.moveNext$0();) {
-        if (skipCount === 0)
-          return iterator.get$current();
-        --skipCount;
-      }
-      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, null, "index"));
-    },
-    toString$0(_) {
-      return A.Iterable_iterableToShortString(this, "(", ")");
-    }
-  };
-  A.Null.prototype = {
-    get$hashCode(_) {
-      return A.Object.prototype.get$hashCode.call(this, this);
-    },
-    toString$0(_) {
-      return "null";
-    }
-  };
-  A.Object.prototype = {$isObject: 1,
-    $eq(_, other) {
-      return this === other;
-    },
-    get$hashCode(_) {
-      return A.Primitives_objectHashCode(this);
-    },
-    toString$0(_) {
-      return "Instance of '" + A.Primitives_objectTypeName(this) + "'";
-    },
-    get$runtimeType(_) {
-      return A.getRuntimeTypeOfDartObject(this);
-    },
-    toString() {
-      return this.toString$0(this);
-    }
-  };
-  A.StringBuffer.prototype = {
-    get$length(_) {
-      return this._contents.length;
-    },
-    toString$0(_) {
-      var t1 = this._contents;
-      return t1.charCodeAt(0) == 0 ? t1 : t1;
-    }
-  };
-  A.HtmlElement.prototype = {};
-  A.AnchorElement.prototype = {
-    toString$0(receiver) {
-      var t1 = String(receiver);
-      t1.toString;
-      return t1;
-    }
-  };
-  A.AreaElement.prototype = {
-    toString$0(receiver) {
-      var t1 = String(receiver);
-      t1.toString;
-      return t1;
-    }
-  };
-  A.ButtonElement.prototype = {$isButtonElement: 1};
-  A.CharacterData.prototype = {
-    get$length(receiver) {
-      return receiver.length;
-    }
-  };
-  A.DomException.prototype = {
-    toString$0(receiver) {
-      var t1 = String(receiver);
-      t1.toString;
-      return t1;
-    }
-  };
-  A.DomTokenList.prototype = {
-    get$length(receiver) {
-      var t1 = receiver.length;
-      t1.toString;
-      return t1;
-    }
-  };
-  A._ChildrenElementList.prototype = {
-    get$isEmpty(_) {
-      return this._element.firstElementChild == null;
-    },
-    get$length(_) {
-      return this._childElements.length;
-    },
-    $index(_, index) {
-      var t1 = this._childElements;
-      if (!(index >= 0 && index < t1.length))
-        return A.ioore(t1, index);
-      return type$.Element._as(t1[index]);
-    },
-    add$1(_, value) {
-      this._element.appendChild(value).toString;
-      return value;
-    },
-    get$iterator(_) {
-      var t1 = this.toList$0(this);
-      return new J.ArrayIterator(t1, t1.length, A._arrayInstanceType(t1)._eval$1("ArrayIterator<1>"));
-    },
-    clear$0(_) {
-      J._clearChildren$0$x(this._element);
-    }
-  };
-  A.Element.prototype = {
-    get$children(receiver) {
-      var t1 = receiver.children;
-      t1.toString;
-      return new A._ChildrenElementList(receiver, t1);
-    },
-    get$classes(receiver) {
-      return new A._ElementCssClassSet(receiver);
-    },
-    toString$0(receiver) {
-      var t1 = receiver.localName;
-      t1.toString;
-      return t1;
-    },
-    $isElement: 1
-  };
-  A.Event.prototype = {$isEvent: 1};
-  A.EventTarget.prototype = {
-    addEventListener$2(receiver, type, listener) {
-      this._addEventListener$3(receiver, type, type$.nullable_dynamic_Function_Event._as(listener), null);
-    },
-    _addEventListener$3(receiver, type, listener, options) {
-      return receiver.addEventListener(type, A.convertDartClosureToJS(type$.nullable_dynamic_Function_Event._as(listener), 1), options);
-    }
-  };
-  A.FormElement.prototype = {
-    get$length(receiver) {
-      return receiver.length;
-    }
-  };
-  A.HtmlCollection.prototype = {
-    get$length(receiver) {
-      var t1 = receiver.length;
-      t1.toString;
-      return t1;
-    },
-    $index(receiver, index) {
-      var t1 = receiver.length,
-        t2 = index >>> 0 !== index || index >= t1;
-      t2.toString;
-      if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
-      t1 = receiver[index];
-      t1.toString;
-      return t1;
-    },
-    elementAt$1(receiver, index) {
-      if (!(index >= 0 && index < receiver.length))
-        return A.ioore(receiver, index);
-      return receiver[index];
-    },
-    $isJavaScriptIndexingBehavior: 1,
-    $isIterable: 1,
-    $isList: 1,
-    $isHtmlCollection: 1
-  };
-  A.ImageElement.prototype = {
-    set$src(receiver, value) {
-      receiver.src = value;
-    },
-    $isImageElement: 1
-  };
-  A._ChildNodeListLazy.prototype = {
-    get$iterator(_) {
-      var t1 = this._this.childNodes;
-      return new A.FixedSizeListIterator(t1, t1.length, A.instanceType(t1)._eval$1("FixedSizeListIterator<ImmutableListMixin.E>"));
-    },
-    get$length(_) {
-      return this._this.childNodes.length;
-    },
-    $index(_, index) {
-      var t1 = this._this.childNodes;
-      if (!(index >= 0 && index < t1.length))
-        return A.ioore(t1, index);
-      return t1[index];
-    }
-  };
-  A.Node.prototype = {
-    _clearChildren$0(receiver) {
-      var t1;
-      for (; t1 = receiver.firstChild, t1 != null;)
-        receiver.removeChild(t1).toString;
-    },
-    toString$0(receiver) {
-      var value = receiver.nodeValue;
-      return value == null ? this.super$Interceptor$toString(receiver) : value;
-    },
-    $isNode: 1
-  };
-  A.NodeList.prototype = {
-    get$length(receiver) {
-      var t1 = receiver.length;
-      t1.toString;
-      return t1;
-    },
-    $index(receiver, index) {
-      var t1 = receiver.length,
-        t2 = index >>> 0 !== index || index >= t1;
-      t2.toString;
-      if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
-      t1 = receiver[index];
-      t1.toString;
-      return t1;
-    },
-    elementAt$1(receiver, index) {
-      if (!(index >= 0 && index < receiver.length))
-        return A.ioore(receiver, index);
-      return receiver[index];
-    },
-    $isJavaScriptIndexingBehavior: 1,
-    $isIterable: 1,
-    $isList: 1
-  };
-  A.SelectElement.prototype = {
-    get$length(receiver) {
-      return receiver.length;
-    }
-  };
-  A._NamedNodeMap.prototype = {
-    get$length(receiver) {
-      var t1 = receiver.length;
-      t1.toString;
-      return t1;
-    },
-    $index(receiver, index) {
-      var t1 = receiver.length,
-        t2 = index >>> 0 !== index || index >= t1;
-      t2.toString;
-      if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
-      t1 = receiver[index];
-      t1.toString;
-      return t1;
-    },
-    elementAt$1(receiver, index) {
-      if (!(index >= 0 && index < receiver.length))
-        return A.ioore(receiver, index);
-      return receiver[index];
-    },
-    $isJavaScriptIndexingBehavior: 1,
-    $isIterable: 1,
-    $isList: 1
-  };
-  A._ElementCssClassSet.prototype = {
-    readClasses$0() {
-      var t1, t2, _i, trimmed,
-        s = A.LinkedHashSet_LinkedHashSet(type$.String);
-      for (t1 = this._element.className.split(" "), t2 = t1.length, _i = 0; _i < t2; ++_i) {
-        trimmed = J.trim$0$s(t1[_i]);
-        if (trimmed.length !== 0)
-          s.add$1(0, trimmed);
-      }
-      return s;
-    },
-    writeClasses$1(s) {
-      this._element.className = type$.Set_String._as(s).join$1(0, " ");
-    },
-    get$length(_) {
-      var t1 = this._element.classList.length;
-      t1.toString;
-      return t1;
-    },
-    add$1(_, value) {
-      var list = this._element.classList,
-        t1 = list.contains(value);
-      t1.toString;
-      list.add(value);
-      return !t1;
-    }
-  };
-  A.ImmutableListMixin.prototype = {
-    get$iterator(receiver) {
-      return new A.FixedSizeListIterator(receiver, this.get$length(receiver), A.instanceType(receiver)._eval$1("FixedSizeListIterator<ImmutableListMixin.E>"));
-    }
-  };
-  A.FixedSizeListIterator.prototype = {
-    moveNext$0() {
-      var _this = this,
-        nextPosition = _this._position + 1,
-        t1 = _this._html$_length;
-      if (nextPosition < t1) {
-        _this.set$_html$_current(J.$index$ax(_this._array, nextPosition));
-        _this._position = nextPosition;
-        return true;
-      }
-      _this.set$_html$_current(null);
-      _this._position = t1;
-      return false;
-    },
-    get$current() {
-      var t1 = this._html$_current;
-      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
-    },
-    set$_html$_current(_current) {
-      this._html$_current = this.$ti._eval$1("1?")._as(_current);
-    },
-    $isIterator: 1
-  };
-  A._HtmlCollection_JavaScriptObject_ListMixin.prototype = {};
-  A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin.prototype = {};
-  A._NodeList_JavaScriptObject_ListMixin.prototype = {};
-  A._NodeList_JavaScriptObject_ListMixin_ImmutableListMixin.prototype = {};
-  A.__NamedNodeMap_JavaScriptObject_ListMixin.prototype = {};
-  A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin.prototype = {};
-  A.CssClassSetImpl.prototype = {
-    _validateToken$1(value) {
-      var t1 = $.$get$CssClassSetImpl__validTokenRE()._nativeRegExp;
-      if (t1.test(value))
-        return value;
-      throw A.wrapException(A.ArgumentError$value(value, "value", "Not a valid class token"));
-    },
-    toString$0(_) {
-      return this.readClasses$0().join$1(0, " ");
-    },
-    get$iterator(_) {
-      var t1 = this.readClasses$0();
-      return A._LinkedHashSetIterator$(t1, t1._modifications, A._instanceType(t1)._precomputed1);
-    },
-    get$length(_) {
-      return this.readClasses$0()._collection$_length;
-    },
-    add$1(_, value) {
-      var t1, s, ret;
-      this._validateToken$1(value);
-      t1 = type$.dynamic_Function_Set_String._as(new A.CssClassSetImpl_add_closure(value));
-      s = this.readClasses$0();
-      ret = t1.call$1(s);
-      this.writeClasses$1(s);
-      return A._asBool(ret == null ? false : ret);
-    },
-    elementAt$1(_, index) {
-      return this.readClasses$0().elementAt$1(0, index);
-    }
-  };
-  A.CssClassSetImpl_add_closure.prototype = {
-    call$1(s) {
-      return type$.Set_String._as(s).add$1(0, this.value);
-    },
-    $signature: 4
-  };
-  A.FilteredElementList.prototype = {
-    get$_html_common$_iterable() {
-      var t1 = this._childNodes,
-        t2 = A._instanceType(t1);
-      return new A.MappedIterable(new A.WhereIterable(t1, t2._eval$1("bool(ListBase.E)")._as(new A.FilteredElementList__iterable_closure()), t2._eval$1("WhereIterable<ListBase.E>")), t2._eval$1("Element(ListBase.E)")._as(new A.FilteredElementList__iterable_closure0()), t2._eval$1("MappedIterable<ListBase.E,Element>"));
-    },
-    add$1(_, value) {
-      this._childNodes._this.appendChild(value).toString;
-    },
-    clear$0(_) {
-      J._clearChildren$0$x(this._childNodes._this);
-    },
-    get$length(_) {
-      return J.get$length$asx(this.get$_html_common$_iterable().__internal$_iterable);
-    },
-    $index(_, index) {
-      var t1 = this.get$_html_common$_iterable();
-      return t1._f.call$1(J.elementAt$1$ax(t1.__internal$_iterable, index));
-    },
-    get$iterator(_) {
-      var t1 = A.List_List$from(this.get$_html_common$_iterable(), false, type$.Element);
-      return new J.ArrayIterator(t1, t1.length, A._arrayInstanceType(t1)._eval$1("ArrayIterator<1>"));
-    }
-  };
-  A.FilteredElementList__iterable_closure.prototype = {
-    call$1(n) {
-      return type$.Element._is(type$.Node._as(n));
-    },
-    $signature: 5
-  };
-  A.FilteredElementList__iterable_closure0.prototype = {
-    call$1(n) {
-      return type$.Element._as(type$.Node._as(n));
-    },
-    $signature: 6
-  };
-  A.AttributeClassSet.prototype = {
-    readClasses$0() {
-      var t1, t2, _i, trimmed,
-        classname = this._svg$_element.getAttribute("class"),
-        s = A.LinkedHashSet_LinkedHashSet(type$.String);
-      if (classname == null)
-        return s;
-      for (t1 = classname.split(" "), t2 = t1.length, _i = 0; _i < t2; ++_i) {
-        trimmed = J.trim$0$s(t1[_i]);
-        if (trimmed.length !== 0)
-          s.add$1(0, trimmed);
-      }
-      return s;
-    },
-    writeClasses$1(s) {
-      this._svg$_element.setAttribute("class", s.join$1(0, " "));
-    }
-  };
-  A.SvgElement.prototype = {
-    get$classes(receiver) {
-      return new A.AttributeClassSet(receiver);
-    },
-    get$children(receiver) {
-      return new A.FilteredElementList(new A._ChildNodeListLazy(receiver));
-    }
-  };
-  A.ChessPieceFactory.prototype = {};
-  A.BlackPieceFactory.prototype = {
-    createPawn$0() {
-      var t1 = new A.BlackPawn(new A.PawnMovement(), "b", "pawn");
-      t1.__GamePiece_src_A = "./assets/chess/pawn_b.png";
-      return t1;
-    },
-    createKnight$0() {
-      var t1 = new A.BlackKnight(new A.KnightMovement(), "b", "knight");
-      t1.__GamePiece_src_A = "./assets/chess/knight_b.png";
-      return t1;
-    },
-    createBishop$0() {
-      var t1 = new A.BlackBishop(new A.BishopMovement(), "b", "bishop");
-      t1.__GamePiece_src_A = "./assets/chess/bishop_b.png";
-      return t1;
-    },
-    createRook$0() {
-      var t1 = new A.BlackRook(new A.RookMovement(), "b", "rook");
-      t1.__GamePiece_src_A = "./assets/chess/rook_b.png";
-      return t1;
-    },
-    createKing$0() {
-      var t1 = new A.BlackKing(new A.KingMovement(), "b", "king");
-      t1.__GamePiece_src_A = "./assets/chess/king_b.png";
-      return t1;
-    },
-    createQueen$0() {
-      var t1 = new A.BlackQueen(new A.QueenMovement(), "b", "queen");
-      t1.__GamePiece_src_A = "./assets/chess/queen_b.png";
-      return t1;
-    }
-  };
-  A.WhitePieceFactory.prototype = {
-    createPawn$0() {
-      var t1 = new A.WhitePawn(new A.PawnMovement(), "w", "pawn");
-      t1.__GamePiece_src_A = "./assets/chess/pawn_w.png";
-      return t1;
-    },
-    createKnight$0() {
-      var t1 = new A.WhiteKnight(new A.KnightMovement(), "w", "knight");
-      t1.__GamePiece_src_A = "./assets/chess/knight_w.png";
-      return t1;
-    },
-    createBishop$0() {
-      var t1 = new A.WhiteBishop(new A.BishopMovement(), "w", "bishop");
-      t1.__GamePiece_src_A = "./assets/chess/bishop_w.png";
-      return t1;
-    },
-    createRook$0() {
-      var t1 = new A.WhiteRook(new A.RookMovement(), "w", "rook");
-      t1.__GamePiece_src_A = "./assets/chess/rook_w.png";
-      return t1;
-    },
-    createKing$0() {
-      var t1 = new A.WhiteKing(new A.KingMovement(), "w", "king");
-      t1.__GamePiece_src_A = "./assets/chess/king_w.png";
-      return t1;
-    },
-    createQueen$0() {
-      var t1 = new A.WhiteQueen(new A.QueenMovement(), "w", "queen");
-      t1.__GamePiece_src_A = "./assets/chess/queen_w.png";
-      return t1;
-    }
-  };
-  A.Pawn.prototype = {};
-  A.Knight.prototype = {};
-  A.Bishop.prototype = {};
-  A.Rook.prototype = {};
-  A.King.prototype = {};
-  A.Queen.prototype = {};
-  A.BlackPawn.prototype = {};
-  A.BlackKnight.prototype = {};
-  A.BlackBishop.prototype = {};
-  A.BlackRook.prototype = {};
-  A.BlackKing.prototype = {};
-  A.BlackQueen.prototype = {};
-  A.ChequeredBoard.prototype = {
-    setupPieces$0() {
-      var t1, t2, i, row, j;
-      for (t1 = this.pieces, t2 = type$.JSArray_ChessPiece, i = 0; i < 8; ++i) {
-        row = A._setArrayType(new Array(0), t2);
-        for (j = 0; j < 8; ++j)
-          B.JSArray_methods.add$1(row, A.EmptyPiece$(i, j));
-        B.JSArray_methods.add$1(t1, row);
-      }
-    },
-    getBoardState$0() {
-      return this.pieces;
-    },
-    placePiece$3(piece, i, j) {
-      var t1 = this.pieces;
-      if (!(i >= 0 && i < t1.length))
-        return A.ioore(t1, i);
-      B.JSArray_methods.$indexSet(t1[i], j, piece);
-      piece.i = i;
-      piece.j = j;
-    },
-    removePiece$2(i, j) {
-      var t1 = this.pieces;
-      if (!(i >= 0 && i < t1.length))
-        return A.ioore(t1, i);
-      B.JSArray_methods.$indexSet(t1[i], j, A.EmptyPiece$(i, j));
-    },
-    getPiece$2(i, j) {
-      var t1 = this.pieces;
-      if (!(i >= 0 && i < t1.length))
-        return A.ioore(t1, i);
-      t1 = t1[i];
-      if (!(j >= 0 && j < t1.length))
-        return A.ioore(t1, j);
-      return t1[j];
-    },
-    $isGameBoard: 1
-  };
-  A.BoardWithPieces.prototype = {
-    removePiece$2(i, j) {
-      this.base.removePiece$2(i, j);
-    },
-    placePiece$3(piece, i, j) {
-      this.base.placePiece$3(piece, i, j);
-    },
-    getPiece$2(i, j) {
-      return this.base.getPiece$2(i, j);
-    },
-    getBoardState$0() {
-      return this.base.getBoardState$0();
-    },
-    $isGameBoard: 1
-  };
-  A.BoardWithPawns.prototype = {
-    setupPieces$0() {
-      var j, pawn,
-        t1 = this.base;
-      t1.setupPieces$0();
-      for (j = 0; j < 8; ++j) {
-        pawn = new A.ChessPiece(new A.PawnMovement(), "b", "pawn");
-        pawn.__GamePiece_src_A = "./assets/chess/pawn_b.png";
-        t1.placePiece$3(pawn, 1, j);
-        pawn = new A.ChessPiece(new A.PawnMovement(), "w", "pawn");
-        pawn.__GamePiece_src_A = "./assets/chess/pawn_w.png";
-        t1.placePiece$3(pawn, 6, j);
-      }
-    }
-  };
-  A.BoardWithBishops.prototype = {
-    setupPieces$0() {
-      var _this = this;
-      _this.base.setupPieces$0();
-      _this.placeBishop$3("b", 0, 2);
-      _this.placeBishop$3("b", 0, 5);
-      _this.placeBishop$3("w", 7, 2);
-      _this.placeBishop$3("w", 7, 5);
-    },
-    placeBishop$3(colour, i, j) {
-      this.base.placePiece$3(A.ChessPiece$(colour, "bishop", new A.BishopMovement()), i, j);
-    }
-  };
-  A.BoardWithKnights.prototype = {
-    setupPieces$0() {
-      var _this = this;
-      _this.base.setupPieces$0();
-      _this.placeKnight$3("b", 0, 1);
-      _this.placeKnight$3("b", 0, 6);
-      _this.placeKnight$3("w", 7, 1);
-      _this.placeKnight$3("w", 7, 6);
-    },
-    placeKnight$3(colour, i, j) {
-      this.base.placePiece$3(A.ChessPiece$(colour, "knight", new A.KnightMovement()), i, j);
-    }
-  };
-  A.BoardWithRooks.prototype = {
-    setupPieces$0() {
-      var _this = this;
-      _this.base.setupPieces$0();
-      _this.placeRook$3("b", 0, 0);
-      _this.placeRook$3("b", 0, 7);
-      _this.placeRook$3("w", 7, 0);
-      _this.placeRook$3("w", 7, 7);
-    },
-    placeRook$3(colour, i, j) {
-      this.base.placePiece$3(A.ChessPiece$(colour, "rook", new A.RookMovement()), i, j);
-    }
-  };
-  A.BoardWithKings.prototype = {
-    setupPieces$0() {
-      var t1 = this.base;
-      t1.setupPieces$0();
-      t1.placePiece$3(A.ChessKing$("b"), 0, 4);
-      t1.placePiece$3(A.ChessKing$("w"), 7, 4);
-    }
-  };
-  A.BoardWithQueens.prototype = {
-    setupPieces$0() {
-      this.base.setupPieces$0();
-      this.placeQueen$3("b", 0, 3);
-      this.placeQueen$3("w", 7, 3);
-    },
-    placeQueen$3(colour, i, j) {
-      this.base.placePiece$3(A.ChessPiece$(colour, "queen", new A.QueenMovement()), i, j);
-    }
-  };
-  A.ChessGame.prototype = {
-    threatenAllWith$1(selection) {
-      var t2, t3, _i, t4, t5, t6,
-        t1 = this.__ChessGame_logic_A;
-      t1 === $ && A.throwLateFieldNI("logic");
-      t1.clearMoveOptions$0();
-      this.__ChessGame_logic_A.activePiece = selection;
-      for (t1 = this.board.getBoardState$0(), t2 = t1.length, t3 = selection.colour, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
-        for (t4 = B.JSArray_methods.get$iterator(t1[_i]); t4.moveNext$0();) {
-          t5 = t4.get$current();
-          t6 = t5.colour;
-          if (t6 !== t3)
-            t5.threatened = true;
-        }
-    },
-    refreshView$0() {
-      var t1 = this.__ChessGame_view_A;
-      t1 === $ && A.throwLateFieldNI("view");
-      t1.displayBoard$1(this.board.getBoardState$0());
-    },
-    setupPieces$1(state) {
-      var t1, t2, _this = this;
-      type$.List_bool._as(state);
-      t1 = J.JSArray_JSArray$growable(0, type$.List_GamePiece);
-      t1 = _this.board = new A.ChequeredBoard(t1);
-      t2 = state.length;
-      if (0 >= t2)
-        return A.ioore(state, 0);
-      if (state[0]) {
-        t1 = new A.BoardWithPawns(t1);
-        _this.board = t1;
-      }
-      if (1 >= t2)
-        return A.ioore(state, 1);
-      if (state[1]) {
-        t1 = new A.BoardWithBishops(t1);
-        _this.board = t1;
-      }
-      if (2 >= t2)
-        return A.ioore(state, 2);
-      if (state[2]) {
-        t1 = new A.BoardWithKnights(t1);
-        _this.board = t1;
-      }
-      if (3 >= t2)
-        return A.ioore(state, 3);
-      if (state[3]) {
-        t1 = new A.BoardWithRooks(t1);
-        _this.board = t1;
-      }
-      if (4 >= t2)
-        return A.ioore(state, 4);
-      if (state[4]) {
-        t1 = new A.BoardWithQueens(t1);
-        _this.board = t1;
-      }
-      if (5 >= t2)
-        return A.ioore(state, 5);
-      (state[5] ? _this.board = new A.BoardWithKings(t1) : t1).setupPieces$0();
-      _this.refreshView$0();
-    },
-    $isGame: 1
-  };
-  A.ChessLogic.prototype = {
-    findKings$0() {
-      var t1, t2, _i, t3, t4, t5, t6, t7;
-      for (t1 = this.game.board.getBoardState$0(), t2 = t1.length, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
-        for (t3 = B.JSArray_methods.get$iterator(t1[_i]); t3.moveNext$0();) {
-          t4 = t3.get$current();
-          if (t4 instanceof A.ChessKing) {
-            t5 = t4.colour;
-            t6 = "" + t4.i;
-            t7 = "" + t4.j;
-            if (t5 === "b") {
-              A.printString("black king found ! " + t6 + " " + t7);
-              this.blackKing = t4;
-            } else {
-              A.printString("white king found ! " + t6 + " " + t7);
-              this.whiteKing = t4;
-            }
-          }
-        }
-    },
-    pairKings$0() {
-      var t1, t2, _i, t3, t4, t5, _this = this;
-      _this.findKings$0();
-      if (_this.blackKing instanceof A.ChessKing && _this.whiteKing instanceof A.ChessKing)
-        for (t1 = _this.game.board.getBoardState$0(), t2 = t1.length, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
-          for (t3 = B.JSArray_methods.get$iterator(t1[_i]); t3.moveNext$0();) {
-            t4 = t3.get$current();
-            t5 = t4.colour;
-            if (t5 === "b")
-              t4.myKing = _this.blackKing;
-            if (t5 === "w")
-              t4.myKing = _this.whiteKing;
-          }
-    },
-    submitMove$2(i, j) {
-      var t1, t2, piece, _this = this;
-      if (_this.validMove$3(_this.activePiece, i, j)) {
-        t1 = _this.activePiece;
-        t2 = _this.game;
-        t2.board.removePiece$2(t1.i, t1.j);
-        t2.board.removePiece$2(i, j);
-        t2.board.placePiece$3(t1, i, j);
-        t1.hasMoved = true;
-        ++t2.turnCount;
-        t1 = t2.__ChessGame_logic_A;
-        t1 === $ && A.throwLateFieldNI("logic");
-        t1.activePiece = A.EmptyPiece$(0, 0);
-        t2.__ChessGame_logic_A.clearMoveOptions$0();
-        t2.refreshView$0();
-        return;
-      }
-      _this.clearMoveOptions$0();
-      _this.activePiece = A.EmptyPiece$(0, 0);
-      t1 = _this.game;
-      piece = t1.board.getPiece$2(i, j);
-      t2 = B.JSInt_methods.$mod(t1.turnCount, 2) === 0 ? "w" : "b";
-      if (piece.colour === t2) {
-        piece.moveStrategy.move$2(t1.board, piece);
-        _this.activePiece = piece;
-      }
-      t1.refreshView$0();
-    },
-    clearMoveOptions$0() {
-      var t1, i, j;
-      for (t1 = this.game, i = 0; i < 8; ++i)
-        for (j = 0; j < 8; ++j)
-          t1.board.getPiece$2(i, j).threatened = false;
-    },
-    validMove$3(piece, i, j) {
-      var t1 = this.game.board.getPiece$2(i, j).threatened;
-      return t1;
-    }
-  };
-  A.ChessPiece.prototype = {
-    canCapture$3(board, i, j) {
-      var target, t1;
-      if (this.validCoords$2(i, j)) {
-        target = board.getPiece$2(i, j);
-        t1 = this.isSafeMove$3(board, i, j);
-        if (t1)
-          if (!(target instanceof A.EmptyPiece) && target.colour !== this.colour)
-            return target.threatened = true;
-      }
-      return false;
-    },
-    canMove$3(board, i, j) {
-      var target;
-      if (this.validCoords$2(i, j)) {
-        target = board.getPiece$2(i, j);
-        if (target instanceof A.EmptyPiece && this.isSafeMove$3(board, i, j))
-          return target.threatened = true;
-      }
-      return false;
-    },
-    isSafeMove$3(board, endI, endJ) {
-      var king, verdict, t1, t2, _this = this,
-        startI = _this.i,
-        startJ = _this.j,
-        target = board.getPiece$2(endI, endJ);
-      board.removePiece$2(startI, startJ);
-      board.removePiece$2(endI, endJ);
-      board.placePiece$3(_this, endI, endJ);
-      king = _this.myKing;
-      if (king instanceof A.ChessKing && king.isTheatened$1(board)) {
-        A.print("my king is at " + king.i + " , " + king.j);
-        verdict = false;
-      } else
-        verdict = true;
-      board.placePiece$3(_this, startI, startJ);
-      board.placePiece$3(target, endI, endJ);
-      t1 = "" + endI;
-      t2 = "" + endJ;
-      if (verdict)
-        A.print("this move is safe! (" + t1 + ", " + t2 + ")");
-      else
-        A.print("this move is unsafe! (" + t1 + ", " + t2 + ")");
-      return verdict;
-    },
-    validCoords$2(i, j) {
-      var t1 = 0 <= i && i < 8;
-      return B.JSBool_methods.$and(t1, 0 <= j && j < 8);
-    }
-  };
-  A.ChessKing.prototype = {
-    isTheatened$1(board) {
-      var _this = this;
-      if (_this.attackedByPawn$1(board) || _this.attackedByKnight$1(board))
-        return true;
-      else if (_this.attackedOnDiagonals$1(board) || _this.attackedOnHorizontals$1(board))
-        return true;
-      return false;
-    },
-    attackedOnDiagonals$1(board) {
-      var t1, threats, t2, _i, _this = this,
-        diagonalThreats = A.HashSet_HashSet(type$.String);
-      diagonalThreats.add$1(0, "queen");
-      diagonalThreats.add$1(0, "bishop");
-      t1 = type$.dynamic;
-      threats = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([_this.threatAlongImpulse$3(board, -1, 1), _this.threatAlongImpulse$3(board, 1, 1), _this.threatAlongImpulse$3(board, 1, -1), _this.threatAlongImpulse$3(board, -1, -1)], t1), true, t1);
-      for (t1 = threats.length, t2 = type$.ChessPiece, _i = 0; _i < t1; ++_i)
-        if (diagonalThreats.contains$1(0, t2._as(threats[_i]).name))
-          return true;
-      return false;
-    },
-    attackedOnHorizontals$1(board) {
-      var t1, threats, t2, _i, _this = this,
-        horizontalThreats = A.HashSet_HashSet(type$.String);
-      horizontalThreats.add$1(0, "queen");
-      horizontalThreats.add$1(0, "rook");
-      t1 = type$.dynamic;
-      threats = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([_this.threatAlongImpulse$3(board, -1, 0), _this.threatAlongImpulse$3(board, 0, 1), _this.threatAlongImpulse$3(board, 1, 0), _this.threatAlongImpulse$3(board, 0, -1)], t1), true, t1);
-      for (t1 = threats.length, t2 = type$.ChessPiece, _i = 0; _i < t1; ++_i)
-        if (horizontalThreats.contains$1(0, t2._as(threats[_i]).name))
-          return true;
-      return false;
-    },
-    threatAlongImpulse$3(board, di, dj) {
-      var t1, threat, _this = this,
-        a = _this.i,
-        b = _this.j;
-      for (; true;) {
-        a += di;
-        b += dj;
-        t1 = 0 <= a && a < 8;
-        if (!B.JSBool_methods.$and(t1, 0 <= b && b < 8))
-          return A.EmptyPiece$(0, 0);
-        threat = _this.getPiece$3(board, a, b);
-        if (threat instanceof A.EmptyPiece)
-          continue;
-        if (threat.colour === _this.colour)
-          return A.EmptyPiece$(0, 0);
-        return threat;
-      }
-    },
-    attackedByPawn$1(board) {
-      var t3, t4, t5, _i, t6, t7, pawn, _this = this,
-        t1 = _this.colour,
-        t2 = _this.i,
-        row = t1 === "w" ? t2 - 1 : t2 + 1;
-      for (t2 = _this.j, t2 = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([t2 - 1, t2 + 1], type$.dynamic), true, type$.int), t3 = t2.length, t4 = 0 <= row, t5 = row < 8, _i = 0; _i < t3; ++_i) {
-        t6 = A._asInt(t2[_i]);
-        t7 = t4 && t5;
-        if (B.JSBool_methods.$and(t7, 0 <= t6 && t6 < 8)) {
-          pawn = _this.getPiece$3(board, row, t6);
-          if (pawn.colour !== t1 && pawn.name === "pawn")
-            return true;
-        }
-      }
-      return false;
-    },
-    attackedByKnight$1(board) {
-      var t1, t2, _i, a, t3, _i0, b, t4, t5, knight, _this = this,
-        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, 2, -2, -1], type$.dynamic), true, type$.int);
-      for (t1 = components.length, t2 = _this.colour, _i = 0; _i < t1; ++_i) {
-        a = components[_i];
-        for (t3 = J.getInterceptor$in(a), _i0 = 0; _i0 < t1; ++_i0) {
-          b = components[_i0];
-          if (t3.abs$0(a) + J.abs$0$in(b) === 3) {
-            t4 = _this.i;
-            if (typeof a !== "number")
-              return A.iae(a);
-            t5 = _this.j;
-            if (typeof b !== "number")
-              return A.iae(b);
-            knight = _this.getPiece$3(board, t4 + a, t5 + b);
-            if (knight.colour !== t2 && knight.name === "knight")
-              return true;
-          }
-        }
-      }
-      return false;
-    },
-    getPiece$3(board, i, j) {
-      var piece;
-      if (this.validCoords$2(i, j)) {
-        piece = board.getPiece$2(i, j);
-        return piece;
-      }
-      return A.EmptyPiece$(0, 0);
-    }
-  };
-  A.EmptyPiece.prototype = {};
-  A.NoMovement.prototype = {
-    move$2(board, piece) {
-    },
-    $isMovementStrategy: 1
-  };
-  A.PawnMovement.prototype = {
-    move$2(board, piece) {
-      if (piece.colour === "w") {
-        if (piece.canMove$3(board, piece.i - 1, piece.j))
-          if (!piece.hasMoved)
-            piece.canMove$3(board, piece.i - 2, piece.j);
-        piece.canCapture$3(board, piece.i - 1, piece.j + 1);
-        piece.canCapture$3(board, piece.i - 1, piece.j - 1);
-        return null;
-      } else {
-        if (piece.canMove$3(board, piece.i + 1, piece.j))
-          if (!piece.hasMoved)
-            piece.canMove$3(board, piece.i + 2, piece.j);
-        piece.canCapture$3(board, piece.i + 1, piece.j + 1);
-        piece.canCapture$3(board, piece.i + 1, piece.j - 1);
-        return null;
-      }
-    },
-    $isMovementStrategy: 1
-  };
-  A.KnightMovement.prototype = {
-    move$2(board, piece) {
-      var _i, a, t2, _i0, b, t3, t4,
-        t1 = type$.dynamic,
-        options = J.JSArray_JSArray$growable(0, t1),
-        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, 2, -2, -1], t1), true, type$.int);
-      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
-        a = components[_i];
-        for (t2 = J.getInterceptor$in(a), _i0 = 0; _i0 < t1; ++_i0) {
-          b = components[_i0];
-          if (t2.abs$0(a) + J.abs$0$in(b) === 3) {
-            t3 = piece.i;
-            if (typeof a !== "number")
-              return A.iae(a);
-            t4 = piece.j;
-            if (typeof b !== "number")
-              return A.iae(b);
-            piece.canMove$3(board, t3 + a, t4 + b);
-            piece.canCapture$3(board, piece.i + a, piece.j + b);
-          }
-        }
-      }
-      return options;
-    },
-    $isMovementStrategy: 1
-  };
-  A.BishopMovement.prototype = {
-    move$2(board, piece) {
-      var t1, _i, a, _i0,
-        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, -1], type$.dynamic), true, type$.int);
-      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
-        a = components[_i];
-        for (_i0 = 0; _i0 < t1; ++_i0)
-          this.exploreImpulse$4(piece, board, a, components[_i0]);
-      }
-    },
-    exploreImpulse$4(piece, board, di, dj) {
-      var i = piece.i,
-        j = piece.j;
-      for (; true;) {
-        i += di;
-        j += dj;
-        if (piece.canCapture$3(board, i, j) || !piece.canMove$3(board, i, j))
-          return;
-      }
-    },
-    $isMovementStrategy: 1
-  };
-  A.RookMovement.prototype = {
-    move$2(board, piece) {
-      var _this = this;
-      _this.exploreImpulse$4(piece, board, 0, 1);
-      _this.exploreImpulse$4(piece, board, 0, -1);
-      _this.exploreImpulse$4(piece, board, 1, 0);
-      _this.exploreImpulse$4(piece, board, -1, 0);
-    },
-    exploreImpulse$4(piece, board, di, dj) {
-      var i = piece.i,
-        j = piece.j;
-      for (; true;) {
-        i += di;
-        j += dj;
-        if (piece.canCapture$3(board, i, j) || !piece.canMove$3(board, i, j))
-          return;
-      }
-    },
-    $isMovementStrategy: 1
-  };
-  A.QueenMovement.prototype = {
-    move$2(board, piece) {
-      var t2, _i,
-        t1 = type$.dynamic,
-        pair = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([new A.RookMovement(), new A.BishopMovement()], t1), true, t1);
-      for (t1 = pair.length, t2 = type$.MovementStrategy, _i = 0; _i < t1; ++_i)
-        t2._as(pair[_i]).move$2(board, piece);
-    },
-    $isMovementStrategy: 1
-  };
-  A.KingMovement.prototype = {
-    move$2(board, piece) {
-      var t1, _i, a, _i0, b, t2, t3,
-        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([-1, 0, 1], type$.dynamic), true, type$.int);
-      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
-        a = components[_i];
-        for (_i0 = 0; _i0 < t1; ++_i0) {
-          b = components[_i0];
-          t2 = piece.i;
-          if (typeof a !== "number")
-            return A.iae(a);
-          t3 = piece.j;
-          if (typeof b !== "number")
-            return A.iae(b);
-          piece.canCapture$3(board, t2 + a, t3 + b);
-          piece.canMove$3(board, piece.i + a, piece.j + b);
-        }
-      }
-    },
-    $isMovementStrategy: 1
-  };
-  A.ChessBoardView.prototype = {
-    displayBoard$1(boardstate) {
-      var t1, t2, t3, t4, _i, rowOfPieces, t5, row, t6, t7, t8, tile, t9, img, t10, element, subtype,
-        _s9_ = "piece-img";
-      type$.List_List_GamePiece._as(boardstate);
-      t1 = this.container;
-      t2 = J.getInterceptor$x(t1);
-      t2.get$children(t1).clear$0(0);
-      for (t3 = boardstate.length, t4 = type$.ImageElement, _i = 0; _i < boardstate.length; boardstate.length === t3 || (0, A.throwConcurrentModificationError)(boardstate), ++_i) {
-        rowOfPieces = boardstate[_i];
-        t5 = document;
-        row = t5.createElement("div");
-        t6 = J.getInterceptor$x(row);
-        t6.get$classes(row).add$1(0, "board-row");
-        for (t7 = B.JSArray_methods.get$iterator(rowOfPieces); t7.moveNext$0();) {
-          t8 = t7.get$current();
-          tile = this.createTile$1(t8);
-          t9 = t8 instanceof A.EmptyPiece;
-          if (!t9) {
-            img = t5.createElement("img");
-            J.get$classes$x(img).add$1(0, _s9_);
-            if (t4._is(img)) {
-              t10 = t8.__GamePiece_src_A;
-              t10 === $ && A.throwLateFieldNI("src");
-              B.ImageElement_methods.set$src(img, t10);
-            }
-            J.get$children$x(tile).add$1(0, img);
-          }
-          if (t8.threatened) {
-            element = t5.createElement("div");
-            t10 = J.getInterceptor$x(element);
-            t10.get$classes(element).add$1(0, "marker");
-            subtype = t9 ? "dot" : "circle";
-            t10.get$classes(element).add$1(0, subtype);
-            t9 = J.getInterceptor$x(tile);
-            t9.get$children(tile).add$1(0, element);
-            if (t8 instanceof A.ChessKing) {
-              t9.get$children(tile).clear$0(0);
-              img = t5.createElement("img");
-              J.get$classes$x(img).add$1(0, _s9_);
-              if (t4._is(img))
-                B.ImageElement_methods.set$src(img, "./assets/chess/king_" + t8.colour + "_check.png");
-              t9.get$children(tile).add$1(0, img);
-            }
-          }
-          t6.get$children(row).add$1(0, tile);
-        }
-        t2.get$children(t1).add$1(0, row);
-      }
-    },
-    createTile$1(piece) {
-      var tile = document.createElement("div"),
-        t1 = J.getInterceptor$x(tile);
-      t1.get$classes(tile).add$1(0, "chess-tile");
-      if (B.JSInt_methods.$mod(piece.i + piece.j, 2) !== 0)
-        t1.get$classes(tile).add$1(0, "dark");
-      t1.addEventListener$2(tile, "click", new A.ChessBoardView_createTile_closure(this, piece));
-      return tile;
-    },
-    $isGameView: 1
-  };
-  A.ChessBoardView_createTile_closure.prototype = {
-    call$1($event) {
-      var t1, t2, t3, counter, t4;
-      type$.Event._as($event);
-      t1 = this.$this.game;
-      t2 = this.piece;
-      t3 = t2.i;
-      t2 = t2.j;
-      counter = t1.turnCount;
-      t4 = t1.__ChessGame_logic_A;
-      t4 === $ && A.throwLateFieldNI("logic");
-      t4.submitMove$2(t3, t2);
-      t4 = t1.turnCount;
-      if (counter === t4) {
-        t1.turnCount = t4 + 1;
-        t1.__ChessGame_logic_A.submitMove$2(t3, t2);
-      }
-      t1.refreshView$0();
-    },
-    $signature: 0
-  };
-  A.GamePiece.prototype = {
-    getElement$0() {
-      var t1,
-        img = document.createElement("img");
-      J.get$classes$x(img).add$1(0, "piece-img");
-      if (type$.ImageElement._is(img)) {
-        t1 = this.__GamePiece_src_A;
-        t1 === $ && A.throwLateFieldNI("src");
-        B.ImageElement_methods.set$src(img, t1);
-      }
-      return img;
-    }
-  };
-  A.PieceBox.prototype = {
-    armButton$0() {
-      B.ButtonElement_methods.addEventListener$2(this.button, "click", new A.PieceBox_armButton_closure(this));
-    },
-    switchColour$0() {
-      var t1, t2, list, _this = this,
-        current = _this.factory.colour;
-      _this.factory = current === "black" ? new A.WhitePieceFactory("white") : new A.BlackPieceFactory("black");
-      t1 = _this.button;
-      t2 = current + "-btn";
-      list = t1.classList;
-      list.contains(t2).toString;
-      list.remove(t2);
-      t2 = _this.factory.colour + "-btn";
-      list = t1.classList;
-      list.contains(t2).toString;
-      list.add(t2);
-      t1.innerText = (_this.factory.colour === "black" ? "Black" : "White") + "PieceFactory";
-      _this.presentPieces$0();
-    },
-    presentPieces$0() {
-      var options, t3, _i, option, t4, t5, _this = this,
-        t1 = _this.container,
-        t2 = J.getInterceptor$x(t1);
-      t2.get$children(t1).clear$0(0);
-      options = J.JSArray_JSArray$growable(0, type$.PieceOption);
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createPawn$0()));
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createBishop$0()));
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createKnight$0()));
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createRook$0()));
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createQueen$0()));
-      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createKing$0()));
-      t3 = options.length;
-      _i = 0;
-      for (; _i < options.length; options.length === t3 || (0, A.throwConcurrentModificationError)(options), ++_i) {
-        option = options[_i];
-        t4 = t2.get$children(t1);
-        t5 = option.__PieceOption_element_A;
-        t5 === $ && A.throwLateFieldNI("element");
-        t4.add$1(0, t5);
-      }
-    }
-  };
-  A.PieceBox_armButton_closure.prototype = {
-    call$1($event) {
-      type$.Event._as($event);
-      this.$this.switchColour$0();
-    },
-    $signature: 0
-  };
-  A.PieceOption.prototype = {
-    createElement$0(_) {
-      var element = document.createElement("div"),
-        t1 = J.getInterceptor$x(element);
-      t1.get$classes(element).add$1(0, "piece-option");
-      t1.addEventListener$2(element, "click", new A.PieceOption_createElement_closure(this));
-      t1.get$children(element).add$1(0, this.piece.getElement$0());
-      return element;
-    }
-  };
-  A.PieceOption_createElement_closure.prototype = {
-    call$1($event) {
-      var t1, t2;
-      type$.Event._as($event);
-      t1 = this.$this;
-      t2 = t1.parent;
-      t1 = t1.piece;
-      A.print("selection indicated! " + t1.name + " " + t1.colour);
-      t2.game.threatenAllWith$1(t1);
-      t2.presentPieces$0();
-    },
-    $signature: 0
-  };
-  A.WhitePawn.prototype = {};
-  A.WhiteKnight.prototype = {};
-  A.WhiteBishop.prototype = {};
-  A.WhiteRook.prototype = {};
-  A.WhiteKing.prototype = {};
-  A.WhiteQueen.prototype = {};
-  (function aliases() {
-    var _ = J.Interceptor.prototype;
-    _.super$Interceptor$toString = _.toString$0;
-    _ = J.LegacyJavaScriptObject.prototype;
-    _.super$LegacyJavaScriptObject$toString = _.toString$0;
-  })();
-  (function inheritance() {
-    var _mixin = hunkHelpers.mixin,
-      _inherit = hunkHelpers.inherit,
-      _inheritMany = hunkHelpers.inheritMany;
-    _inherit(A.Object, null);
-    _inheritMany(A.Object, [A.JS_CONST, J.Interceptor, J.ArrayIterator, A.Error, A.ListIterator, A.Iterable, A.MappedIterator, A.WhereIterator, A.Closure, A.JSSyntaxRegExp, A.Rti, A._FunctionParameters, A._Type, A.SetBase, A._HashSetIterator, A._LinkedHashSetCell, A._LinkedHashSetIterator, A.ListBase, A._Exception, A.FormatException, A.Null, A.StringBuffer, A.ImmutableListMixin, A.FixedSizeListIterator, A.ChessPieceFactory, A.GamePiece, A.ChequeredBoard, A.BoardWithPieces, A.ChessGame, A.ChessLogic, A.NoMovement, A.PawnMovement, A.KnightMovement, A.BishopMovement, A.RookMovement, A.QueenMovement, A.KingMovement, A.ChessBoardView, A.PieceBox, A.PieceOption]);
-    _inheritMany(J.Interceptor, [J.JSBool, J.JSNull, J.JavaScriptObject, J.JSNumber, J.JSString]);
-    _inheritMany(J.JavaScriptObject, [J.LegacyJavaScriptObject, J.JSArray, A.EventTarget, A.DomException, A.DomTokenList, A.Event, A._HtmlCollection_JavaScriptObject_ListMixin, A._NodeList_JavaScriptObject_ListMixin, A.__NamedNodeMap_JavaScriptObject_ListMixin]);
-    _inheritMany(J.LegacyJavaScriptObject, [J.PlainJavaScriptObject, J.UnknownJavaScriptObject, J.JavaScriptFunction]);
-    _inherit(J.JSUnmodifiableArray, J.JSArray);
-    _inheritMany(J.JSNumber, [J.JSInt, J.JSNumNotInt]);
-    _inheritMany(A.Error, [A.LateError, A._CyclicInitializationError, A.RuntimeError, A.AssertionError, A._Error, A.TypeError, A.ArgumentError, A.UnsupportedError, A.UnimplementedError, A.ConcurrentModificationError]);
-    _inheritMany(A.Iterable, [A.MappedIterable, A.WhereIterable]);
-    _inheritMany(A.Closure, [A.Closure2Args, A.TearOffClosure, A.initHooks_closure, A.initHooks_closure1, A.CssClassSetImpl_add_closure, A.FilteredElementList__iterable_closure, A.FilteredElementList__iterable_closure0, A.ChessBoardView_createTile_closure, A.PieceBox_armButton_closure, A.PieceOption_createElement_closure]);
-    _inheritMany(A.TearOffClosure, [A.StaticClosure, A.BoundClosure]);
-    _inherit(A._AssertionError, A.AssertionError);
-    _inherit(A.initHooks_closure0, A.Closure2Args);
-    _inherit(A._TypeError, A._Error);
-    _inheritMany(A.SetBase, [A._SetBase, A.CssClassSetImpl]);
-    _inheritMany(A._SetBase, [A._HashSet, A._LinkedHashSet]);
-    _inheritMany(A.ArgumentError, [A.RangeError, A.IndexError]);
-    _inherit(A.Node, A.EventTarget);
-    _inheritMany(A.Node, [A.Element, A.CharacterData]);
-    _inheritMany(A.Element, [A.HtmlElement, A.SvgElement]);
-    _inheritMany(A.HtmlElement, [A.AnchorElement, A.AreaElement, A.ButtonElement, A.FormElement, A.ImageElement, A.SelectElement]);
-    _inheritMany(A.ListBase, [A._ChildrenElementList, A._ChildNodeListLazy, A.FilteredElementList]);
-    _inherit(A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin, A._HtmlCollection_JavaScriptObject_ListMixin);
-    _inherit(A.HtmlCollection, A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin);
-    _inherit(A._NodeList_JavaScriptObject_ListMixin_ImmutableListMixin, A._NodeList_JavaScriptObject_ListMixin);
-    _inherit(A.NodeList, A._NodeList_JavaScriptObject_ListMixin_ImmutableListMixin);
-    _inherit(A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin, A.__NamedNodeMap_JavaScriptObject_ListMixin);
-    _inherit(A._NamedNodeMap, A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin);
-    _inheritMany(A.CssClassSetImpl, [A._ElementCssClassSet, A.AttributeClassSet]);
-    _inheritMany(A.ChessPieceFactory, [A.BlackPieceFactory, A.WhitePieceFactory]);
-    _inherit(A.ChessPiece, A.GamePiece);
-    _inheritMany(A.ChessPiece, [A.Pawn, A.Knight, A.Bishop, A.Rook, A.King, A.Queen, A.ChessKing, A.EmptyPiece]);
-    _inheritMany(A.Pawn, [A.BlackPawn, A.WhitePawn]);
-    _inheritMany(A.Knight, [A.BlackKnight, A.WhiteKnight]);
-    _inheritMany(A.Bishop, [A.BlackBishop, A.WhiteBishop]);
-    _inheritMany(A.Rook, [A.BlackRook, A.WhiteRook]);
-    _inheritMany(A.King, [A.BlackKing, A.WhiteKing]);
-    _inheritMany(A.Queen, [A.BlackQueen, A.WhiteQueen]);
-    _inheritMany(A.BoardWithPieces, [A.BoardWithPawns, A.BoardWithBishops, A.BoardWithKnights, A.BoardWithRooks, A.BoardWithKings, A.BoardWithQueens]);
-    _mixin(A._HtmlCollection_JavaScriptObject_ListMixin, A.ListBase);
-    _mixin(A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin, A.ImmutableListMixin);
-    _mixin(A._NodeList_JavaScriptObject_ListMixin, A.ListBase);
-    _mixin(A._NodeList_JavaScriptObject_ListMixin_ImmutableListMixin, A.ImmutableListMixin);
-    _mixin(A.__NamedNodeMap_JavaScriptObject_ListMixin, A.ListBase);
-    _mixin(A.__NamedNodeMap_JavaScriptObject_ListMixin_ImmutableListMixin, A.ImmutableListMixin);
-  })();
-  var init = {
-    typeUniverse: {eC: new Map(), tR: {}, eT: {}, tPV: {}, sEA: []},
-    mangledGlobalNames: {int: "int", double: "double", num: "num", String: "String", bool: "bool", Null: "Null", List: "List"},
-    mangledNames: {},
-    types: ["Null(Event)", "@(@)", "@(@,String)", "@(String)", "bool(Set<String>)", "bool(Node)", "Element(Node)"],
-    interceptorsByTag: null,
-    leafTags: null,
-    arrayRti: Symbol("$ti")
-  };
-  A._Universe_addRules(init.typeUniverse, JSON.parse('{"PlainJavaScriptObject":"LegacyJavaScriptObject","UnknownJavaScriptObject":"LegacyJavaScriptObject","JavaScriptFunction":"LegacyJavaScriptObject","AbortPaymentEvent":"Event","ExtendableEvent":"Event","AElement":"SvgElement","GraphicsElement":"SvgElement","AudioElement":"HtmlElement","MediaElement":"HtmlElement","HtmlDocument":"Node","Document":"Node","CDataSection":"CharacterData","Text":"CharacterData","MathMLElement":"Element","HtmlFormControlsCollection":"HtmlCollection","JSBool":{"bool":[],"TrustedGetRuntimeType":[]},"JSNull":{"TrustedGetRuntimeType":[]},"JSArray":{"List":["1"],"Iterable":["1"]},"JSUnmodifiableArray":{"JSArray":["1"],"List":["1"],"Iterable":["1"]},"ArrayIterator":{"Iterator":["1"]},"JSNumber":{"num":[]},"JSInt":{"int":[],"num":[],"TrustedGetRuntimeType":[]},"JSNumNotInt":{"num":[],"TrustedGetRuntimeType":[]},"JSString":{"String":[],"TrustedGetRuntimeType":[]},"ListIterator":{"Iterator":["1"]},"MappedIterable":{"Iterable":["2"]},"MappedIterator":{"Iterator":["2"]},"WhereIterable":{"Iterable":["1"]},"WhereIterator":{"Iterator":["1"]},"Closure":{"Function":[]},"Closure2Args":{"Function":[]},"TearOffClosure":{"Function":[]},"StaticClosure":{"Function":[]},"BoundClosure":{"Function":[]},"_HashSet":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"_HashSetIterator":{"Iterator":["1"]},"_LinkedHashSet":{"SetBase":["1"],"LinkedHashSet":["1"],"Set":["1"],"Iterable":["1"]},"_LinkedHashSetIterator":{"Iterator":["1"]},"ListBase":{"List":["1"],"Iterable":["1"]},"SetBase":{"Set":["1"],"Iterable":["1"]},"_SetBase":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"int":{"num":[]},"List":{"Iterable":["1"]},"Set":{"Iterable":["1"]},"Element":{"Node":[]},"HtmlElement":{"Element":[],"Node":[]},"AnchorElement":{"Element":[],"Node":[]},"AreaElement":{"Element":[],"Node":[]},"ButtonElement":{"Element":[],"Node":[]},"CharacterData":{"Node":[]},"_ChildrenElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"FormElement":{"Element":[],"Node":[]},"HtmlCollection":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"ImageElement":{"Element":[],"Node":[]},"_ChildNodeListLazy":{"ListBase":["Node"],"List":["Node"],"Iterable":["Node"],"ListBase.E":"Node"},"NodeList":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"SelectElement":{"Element":[],"Node":[]},"_NamedNodeMap":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"_ElementCssClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FixedSizeListIterator":{"Iterator":["1"]},"CssClassSetImpl":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FilteredElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"AttributeClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"SvgElement":{"Element":[],"Node":[]},"BlackPieceFactory":{"ChessPieceFactory":[]},"WhitePieceFactory":{"ChessPieceFactory":[]},"Pawn":{"ChessPiece":[],"GamePiece":[]},"Knight":{"ChessPiece":[],"GamePiece":[]},"Bishop":{"ChessPiece":[],"GamePiece":[]},"Rook":{"ChessPiece":[],"GamePiece":[]},"King":{"ChessPiece":[],"GamePiece":[]},"Queen":{"ChessPiece":[],"GamePiece":[]},"BlackPawn":{"ChessPiece":[],"GamePiece":[]},"BlackKnight":{"ChessPiece":[],"GamePiece":[]},"BlackBishop":{"ChessPiece":[],"GamePiece":[]},"BlackRook":{"ChessPiece":[],"GamePiece":[]},"BlackKing":{"ChessPiece":[],"GamePiece":[]},"BlackQueen":{"ChessPiece":[],"GamePiece":[]},"ChequeredBoard":{"GameBoard":[]},"BoardWithPieces":{"GameBoard":[]},"BoardWithPawns":{"GameBoard":[]},"BoardWithBishops":{"GameBoard":[]},"BoardWithKnights":{"GameBoard":[]},"BoardWithRooks":{"GameBoard":[]},"BoardWithKings":{"GameBoard":[]},"BoardWithQueens":{"GameBoard":[]},"ChessGame":{"Game":[]},"ChessPiece":{"GamePiece":[]},"ChessKing":{"ChessPiece":[],"GamePiece":[]},"EmptyPiece":{"ChessPiece":[],"GamePiece":[]},"NoMovement":{"MovementStrategy":[]},"PawnMovement":{"MovementStrategy":[]},"KnightMovement":{"MovementStrategy":[]},"BishopMovement":{"MovementStrategy":[]},"RookMovement":{"MovementStrategy":[]},"QueenMovement":{"MovementStrategy":[]},"KingMovement":{"MovementStrategy":[]},"ChessBoardView":{"GameView":[]},"WhitePawn":{"ChessPiece":[],"GamePiece":[]},"WhiteKnight":{"ChessPiece":[],"GamePiece":[]},"WhiteBishop":{"ChessPiece":[],"GamePiece":[]},"WhiteRook":{"ChessPiece":[],"GamePiece":[]},"WhiteKing":{"ChessPiece":[],"GamePiece":[]},"WhiteQueen":{"ChessPiece":[],"GamePiece":[]}}'));
-  A._Universe_addErasedTypes(init.typeUniverse, JSON.parse('{"_SetBase":1}'));
-  var type$ = (function rtii() {
-    var findType = A.findType;
-    return {
-      ButtonElement: findType("ButtonElement"),
-      ChessPiece: findType("ChessPiece"),
-      Element: findType("Element"),
-      Event: findType("Event"),
-      Function: findType("Function"),
-      ImageElement: findType("ImageElement"),
-      Iterable_dynamic: findType("Iterable<@>"),
-      JSArray_ChessPiece: findType("JSArray<ChessPiece>"),
-      JSArray_String: findType("JSArray<String>"),
-      JSArray_dynamic: findType("JSArray<@>"),
-      JSNull: findType("JSNull"),
-      JavaScriptFunction: findType("JavaScriptFunction"),
-      JavaScriptIndexingBehavior_dynamic: findType("JavaScriptIndexingBehavior<@>"),
-      List_GamePiece: findType("List<GamePiece>"),
-      List_List_GamePiece: findType("List<List<GamePiece>>"),
-      List_bool: findType("List<bool>"),
-      MovementStrategy: findType("MovementStrategy"),
-      Node: findType("Node"),
-      Null: findType("Null"),
-      Object: findType("Object"),
-      PieceOption: findType("PieceOption"),
-      Record: findType("Record"),
-      Set_String: findType("Set<String>"),
-      String: findType("String"),
-      TrustedGetRuntimeType: findType("TrustedGetRuntimeType"),
-      UnknownJavaScriptObject: findType("UnknownJavaScriptObject"),
-      bool: findType("bool"),
-      double: findType("double"),
-      dynamic: findType("@"),
-      dynamic_Function_Set_String: findType("@(Set<String>)"),
-      int: findType("int"),
-      legacy_Never: findType("0&*"),
-      legacy_Object: findType("Object*"),
-      nullable_Future_Null: findType("Future<Null>?"),
-      nullable_Object: findType("Object?"),
-      nullable__LinkedHashSetCell: findType("_LinkedHashSetCell?"),
-      nullable_dynamic_Function_Event: findType("@(Event)?"),
-      num: findType("num")
-    };
-  })();
-  (function constants() {
-    B.ButtonElement_methods = A.ButtonElement.prototype;
-    B.ImageElement_methods = A.ImageElement.prototype;
-    B.Interceptor_methods = J.Interceptor.prototype;
-    B.JSArray_methods = J.JSArray.prototype;
-    B.JSBool_methods = J.JSBool.prototype;
-    B.JSInt_methods = J.JSInt.prototype;
-    B.JSString_methods = J.JSString.prototype;
-    B.JavaScriptFunction_methods = J.JavaScriptFunction.prototype;
-    B.JavaScriptObject_methods = J.JavaScriptObject.prototype;
-    B.PlainJavaScriptObject_methods = J.PlainJavaScriptObject.prototype;
-    B.UnknownJavaScriptObject_methods = J.UnknownJavaScriptObject.prototype;
-    B.C_JS_CONST = function getTagFallback(o) {
-  var s = Object.prototype.toString.call(o);
-  return s.substring(8, s.length - 1);
-};
-    B.C_JS_CONST0 = function() {
-  var toStringFunction = Object.prototype.toString;
-  function getTag(o) {
-    var s = toStringFunction.call(o);
-    return s.substring(8, s.length - 1);
-  }
-  function getUnknownTag(object, tag) {
-    if (/^HTML[A-Z].*Element$/.test(tag)) {
-      var name = toStringFunction.call(object);
-      if (name == "[object Object]") return null;
-      return "HTMLElement";
-    }
-  }
-  function getUnknownTagGenericBrowser(object, tag) {
-    if (self.HTMLElement && object instanceof HTMLElement) return "HTMLElement";
-    return getUnknownTag(object, tag);
-  }
-  function prototypeForTag(tag) {
-    if (typeof window == "undefined") return null;
-    if (typeof window[tag] == "undefined") return null;
-    var constructor = window[tag];
-    if (typeof constructor != "function") return null;
-    return constructor.prototype;
-  }
-  function discriminator(tag) { return null; }
-  var isBrowser = typeof navigator == "object";
-  return {
-    getTag: getTag,
-    getUnknownTag: isBrowser ? getUnknownTagGenericBrowser : getUnknownTag,
-    prototypeForTag: prototypeForTag,
-    discriminator: discriminator };
-};
-    B.C_JS_CONST6 = function(getTagFallback) {
-  return function(hooks) {
-    if (typeof navigator != "object") return hooks;
-    var ua = navigator.userAgent;
-    if (ua.indexOf("DumpRenderTree") >= 0) return hooks;
-    if (ua.indexOf("Chrome") >= 0) {
-      function confirm(p) {
-        return typeof window == "object" && window[p] && window[p].name == p;
-      }
-      if (confirm("Window") && confirm("HTMLElement")) return hooks;
-    }
-    hooks.getTag = getTagFallback;
-  };
-};
-    B.C_JS_CONST1 = function(hooks) {
-  if (typeof dartExperimentalFixupGetTag != "function") return hooks;
-  hooks.getTag = dartExperimentalFixupGetTag(hooks.getTag);
-};
-    B.C_JS_CONST2 = function(hooks) {
-  var getTag = hooks.getTag;
-  var prototypeForTag = hooks.prototypeForTag;
-  function getTagFixed(o) {
-    var tag = getTag(o);
-    if (tag == "Document") {
-      if (!!o.xmlVersion) return "!Document";
-      return "!HTMLDocument";
-    }
-    return tag;
-  }
-  function prototypeForTagFixed(tag) {
-    if (tag == "Document") return null;
-    return prototypeForTag(tag);
-  }
-  hooks.getTag = getTagFixed;
-  hooks.prototypeForTag = prototypeForTagFixed;
-};
-    B.C_JS_CONST5 = function(hooks) {
-  var userAgent = typeof navigator == "object" ? navigator.userAgent : "";
-  if (userAgent.indexOf("Firefox") == -1) return hooks;
-  var getTag = hooks.getTag;
-  var quickMap = {
-    "BeforeUnloadEvent": "Event",
-    "DataTransfer": "Clipboard",
-    "GeoGeolocation": "Geolocation",
-    "Location": "!Location",
-    "WorkerMessageEvent": "MessageEvent",
-    "XMLDocument": "!Document"};
-  function getTagFirefox(o) {
-    var tag = getTag(o);
-    return quickMap[tag] || tag;
-  }
-  hooks.getTag = getTagFirefox;
-};
-    B.C_JS_CONST4 = function(hooks) {
-  var userAgent = typeof navigator == "object" ? navigator.userAgent : "";
-  if (userAgent.indexOf("Trident/") == -1) return hooks;
-  var getTag = hooks.getTag;
-  var quickMap = {
-    "BeforeUnloadEvent": "Event",
-    "DataTransfer": "Clipboard",
-    "HTMLDDElement": "HTMLElement",
-    "HTMLDTElement": "HTMLElement",
-    "HTMLPhraseElement": "HTMLElement",
-    "Position": "Geoposition"
-  };
-  function getTagIE(o) {
-    var tag = getTag(o);
-    var newTag = quickMap[tag];
-    if (newTag) return newTag;
-    if (tag == "Object") {
-      if (window.DataView && (o instanceof window.DataView)) return "DataView";
-    }
-    return tag;
-  }
-  function prototypeForTagIE(tag) {
-    var constructor = window[tag];
-    if (constructor == null) return null;
-    return constructor.prototype;
-  }
-  hooks.getTag = getTagIE;
-  hooks.prototypeForTag = prototypeForTagIE;
-};
-    B.C_JS_CONST3 = function(hooks) { return hooks; }
-;
-  })();
-  (function staticFields() {
-    $._JS_INTEROP_INTERCEPTOR_TAG = null;
-    $.toStringVisiting = A._setArrayType([], A.findType("JSArray<Object>"));
-    $.Primitives__identityHashCodeProperty = null;
-    $.BoundClosure__receiverFieldNameCache = null;
-    $.BoundClosure__interceptorFieldNameCache = null;
-    $.getTagFunction = null;
-    $.alternateTagFunction = null;
-    $.prototypeForTagFunction = null;
-    $.dispatchRecordsForInstanceTags = null;
-    $.interceptorsForUncacheableTags = null;
-    $.initNativeDispatchFlag = null;
-  })();
-  (function lazyInitializers() {
-    var _lazyFinal = hunkHelpers.lazyFinal;
-    _lazyFinal($, "DART_CLOSURE_PROPERTY_NAME", "$get$DART_CLOSURE_PROPERTY_NAME", () => A.getIsolateAffinityTag("_$dart_dartClosure"));
-    _lazyFinal($, "CssClassSetImpl__validTokenRE", "$get$CssClassSetImpl__validTokenRE", () => A.RegExp_RegExp("^\\S+$"));
-  })();
-  (function nativeSupport() {
-    !function() {
-      var intern = function(s) {
-        var o = {};
-        o[s] = 1;
-        return Object.keys(hunkHelpers.convertToFastObject(o))[0];
-      };
-      init.getIsolateTag = function(name) {
-        return intern("___dart_" + name + init.isolateTag);
-      };
-      var tableProperty = "___dart_isolate_tags_";
-      var usedProperties = Object[tableProperty] || (Object[tableProperty] = Object.create(null));
-      var rootProperty = "_ZxYxX";
-      for (var i = 0;; i++) {
-        var property = intern(rootProperty + "_" + i + "_");
-        if (!(property in usedProperties)) {
-          usedProperties[property] = 1;
-          init.isolateTag = property;
-          break;
-        }
-      }
-      init.dispatchPropertyName = init.getIsolateTag("dispatch_record");
-    }();
-    hunkHelpers.setOrUpdateInterceptorsByTag({DOMError: J.JavaScriptObject, MediaError: J.JavaScriptObject, NavigatorUserMediaError: J.JavaScriptObject, OverconstrainedError: J.JavaScriptObject, PositionError: J.JavaScriptObject, GeolocationPositionError: J.JavaScriptObject, HTMLAudioElement: A.HtmlElement, HTMLBRElement: A.HtmlElement, HTMLBaseElement: A.HtmlElement, HTMLBodyElement: A.HtmlElement, HTMLCanvasElement: A.HtmlElement, HTMLContentElement: A.HtmlElement, HTMLDListElement: A.HtmlElement, HTMLDataElement: A.HtmlElement, HTMLDataListElement: A.HtmlElement, HTMLDetailsElement: A.HtmlElement, HTMLDialogElement: A.HtmlElement, HTMLDivElement: A.HtmlElement, HTMLEmbedElement: A.HtmlElement, HTMLFieldSetElement: A.HtmlElement, HTMLHRElement: A.HtmlElement, HTMLHeadElement: A.HtmlElement, HTMLHeadingElement: A.HtmlElement, HTMLHtmlElement: A.HtmlElement, HTMLIFrameElement: A.HtmlElement, HTMLInputElement: A.HtmlElement, HTMLLIElement: A.HtmlElement, HTMLLabelElement: A.HtmlElement, HTMLLegendElement: A.HtmlElement, HTMLLinkElement: A.HtmlElement, HTMLMapElement: A.HtmlElement, HTMLMediaElement: A.HtmlElement, HTMLMenuElement: A.HtmlElement, HTMLMetaElement: A.HtmlElement, HTMLMeterElement: A.HtmlElement, HTMLModElement: A.HtmlElement, HTMLOListElement: A.HtmlElement, HTMLObjectElement: A.HtmlElement, HTMLOptGroupElement: A.HtmlElement, HTMLOptionElement: A.HtmlElement, HTMLOutputElement: A.HtmlElement, HTMLParagraphElement: A.HtmlElement, HTMLParamElement: A.HtmlElement, HTMLPictureElement: A.HtmlElement, HTMLPreElement: A.HtmlElement, HTMLProgressElement: A.HtmlElement, HTMLQuoteElement: A.HtmlElement, HTMLScriptElement: A.HtmlElement, HTMLShadowElement: A.HtmlElement, HTMLSlotElement: A.HtmlElement, HTMLSourceElement: A.HtmlElement, HTMLSpanElement: A.HtmlElement, HTMLStyleElement: A.HtmlElement, HTMLTableCaptionElement: A.HtmlElement, HTMLTableCellElement: A.HtmlElement, HTMLTableDataCellElement: A.HtmlElement, HTMLTableHeaderCellElement: A.HtmlElement, HTMLTableColElement: A.HtmlElement, HTMLTableElement: A.HtmlElement, HTMLTableRowElement: A.HtmlElement, HTMLTableSectionElement: A.HtmlElement, HTMLTemplateElement: A.HtmlElement, HTMLTextAreaElement: A.HtmlElement, HTMLTimeElement: A.HtmlElement, HTMLTitleElement: A.HtmlElement, HTMLTrackElement: A.HtmlElement, HTMLUListElement: A.HtmlElement, HTMLUnknownElement: A.HtmlElement, HTMLVideoElement: A.HtmlElement, HTMLDirectoryElement: A.HtmlElement, HTMLFontElement: A.HtmlElement, HTMLFrameElement: A.HtmlElement, HTMLFrameSetElement: A.HtmlElement, HTMLMarqueeElement: A.HtmlElement, HTMLElement: A.HtmlElement, HTMLAnchorElement: A.AnchorElement, HTMLAreaElement: A.AreaElement, HTMLButtonElement: A.ButtonElement, CDATASection: A.CharacterData, CharacterData: A.CharacterData, Comment: A.CharacterData, ProcessingInstruction: A.CharacterData, Text: A.CharacterData, DOMException: A.DomException, DOMTokenList: A.DomTokenList, MathMLElement: A.Element, Element: A.Element, AbortPaymentEvent: A.Event, AnimationEvent: A.Event, AnimationPlaybackEvent: A.Event, ApplicationCacheErrorEvent: A.Event, BackgroundFetchClickEvent: A.Event, BackgroundFetchEvent: A.Event, BackgroundFetchFailEvent: A.Event, BackgroundFetchedEvent: A.Event, BeforeInstallPromptEvent: A.Event, BeforeUnloadEvent: A.Event, BlobEvent: A.Event, CanMakePaymentEvent: A.Event, ClipboardEvent: A.Event, CloseEvent: A.Event, CompositionEvent: A.Event, CustomEvent: A.Event, DeviceMotionEvent: A.Event, DeviceOrientationEvent: A.Event, ErrorEvent: A.Event, Event: A.Event, InputEvent: A.Event, SubmitEvent: A.Event, ExtendableEvent: A.Event, ExtendableMessageEvent: A.Event, FetchEvent: A.Event, FocusEvent: A.Event, FontFaceSetLoadEvent: A.Event, ForeignFetchEvent: A.Event, GamepadEvent: A.Event, HashChangeEvent: A.Event, InstallEvent: A.Event, KeyboardEvent: A.Event, MediaEncryptedEvent: A.Event, MediaKeyMessageEvent: A.Event, MediaQueryListEvent: A.Event, MediaStreamEvent: A.Event, MediaStreamTrackEvent: A.Event, MessageEvent: A.Event, MIDIConnectionEvent: A.Event, MIDIMessageEvent: A.Event, MouseEvent: A.Event, DragEvent: A.Event, MutationEvent: A.Event, NotificationEvent: A.Event, PageTransitionEvent: A.Event, PaymentRequestEvent: A.Event, PaymentRequestUpdateEvent: A.Event, PointerEvent: A.Event, PopStateEvent: A.Event, PresentationConnectionAvailableEvent: A.Event, PresentationConnectionCloseEvent: A.Event, ProgressEvent: A.Event, PromiseRejectionEvent: A.Event, PushEvent: A.Event, RTCDataChannelEvent: A.Event, RTCDTMFToneChangeEvent: A.Event, RTCPeerConnectionIceEvent: A.Event, RTCTrackEvent: A.Event, SecurityPolicyViolationEvent: A.Event, SensorErrorEvent: A.Event, SpeechRecognitionError: A.Event, SpeechRecognitionEvent: A.Event, SpeechSynthesisEvent: A.Event, StorageEvent: A.Event, SyncEvent: A.Event, TextEvent: A.Event, TouchEvent: A.Event, TrackEvent: A.Event, TransitionEvent: A.Event, WebKitTransitionEvent: A.Event, UIEvent: A.Event, VRDeviceEvent: A.Event, VRDisplayEvent: A.Event, VRSessionEvent: A.Event, WheelEvent: A.Event, MojoInterfaceRequestEvent: A.Event, ResourceProgressEvent: A.Event, USBConnectionEvent: A.Event, IDBVersionChangeEvent: A.Event, AudioProcessingEvent: A.Event, OfflineAudioCompletionEvent: A.Event, WebGLContextEvent: A.Event, EventTarget: A.EventTarget, HTMLFormElement: A.FormElement, HTMLCollection: A.HtmlCollection, HTMLFormControlsCollection: A.HtmlCollection, HTMLOptionsCollection: A.HtmlCollection, HTMLImageElement: A.ImageElement, Document: A.Node, DocumentFragment: A.Node, HTMLDocument: A.Node, ShadowRoot: A.Node, XMLDocument: A.Node, Attr: A.Node, DocumentType: A.Node, Node: A.Node, NodeList: A.NodeList, RadioNodeList: A.NodeList, HTMLSelectElement: A.SelectElement, NamedNodeMap: A._NamedNodeMap, MozNamedAttrMap: A._NamedNodeMap, SVGAElement: A.SvgElement, SVGAnimateElement: A.SvgElement, SVGAnimateMotionElement: A.SvgElement, SVGAnimateTransformElement: A.SvgElement, SVGAnimationElement: A.SvgElement, SVGCircleElement: A.SvgElement, SVGClipPathElement: A.SvgElement, SVGDefsElement: A.SvgElement, SVGDescElement: A.SvgElement, SVGDiscardElement: A.SvgElement, SVGEllipseElement: A.SvgElement, SVGFEBlendElement: A.SvgElement, SVGFEColorMatrixElement: A.SvgElement, SVGFEComponentTransferElement: A.SvgElement, SVGFECompositeElement: A.SvgElement, SVGFEConvolveMatrixElement: A.SvgElement, SVGFEDiffuseLightingElement: A.SvgElement, SVGFEDisplacementMapElement: A.SvgElement, SVGFEDistantLightElement: A.SvgElement, SVGFEFloodElement: A.SvgElement, SVGFEFuncAElement: A.SvgElement, SVGFEFuncBElement: A.SvgElement, SVGFEFuncGElement: A.SvgElement, SVGFEFuncRElement: A.SvgElement, SVGFEGaussianBlurElement: A.SvgElement, SVGFEImageElement: A.SvgElement, SVGFEMergeElement: A.SvgElement, SVGFEMergeNodeElement: A.SvgElement, SVGFEMorphologyElement: A.SvgElement, SVGFEOffsetElement: A.SvgElement, SVGFEPointLightElement: A.SvgElement, SVGFESpecularLightingElement: A.SvgElement, SVGFESpotLightElement: A.SvgElement, SVGFETileElement: A.SvgElement, SVGFETurbulenceElement: A.SvgElement, SVGFilterElement: A.SvgElement, SVGForeignObjectElement: A.SvgElement, SVGGElement: A.SvgElement, SVGGeometryElement: A.SvgElement, SVGGraphicsElement: A.SvgElement, SVGImageElement: A.SvgElement, SVGLineElement: A.SvgElement, SVGLinearGradientElement: A.SvgElement, SVGMarkerElement: A.SvgElement, SVGMaskElement: A.SvgElement, SVGMetadataElement: A.SvgElement, SVGPathElement: A.SvgElement, SVGPatternElement: A.SvgElement, SVGPolygonElement: A.SvgElement, SVGPolylineElement: A.SvgElement, SVGRadialGradientElement: A.SvgElement, SVGRectElement: A.SvgElement, SVGScriptElement: A.SvgElement, SVGSetElement: A.SvgElement, SVGStopElement: A.SvgElement, SVGStyleElement: A.SvgElement, SVGElement: A.SvgElement, SVGSVGElement: A.SvgElement, SVGSwitchElement: A.SvgElement, SVGSymbolElement: A.SvgElement, SVGTSpanElement: A.SvgElement, SVGTextContentElement: A.SvgElement, SVGTextElement: A.SvgElement, SVGTextPathElement: A.SvgElement, SVGTextPositioningElement: A.SvgElement, SVGTitleElement: A.SvgElement, SVGUseElement: A.SvgElement, SVGViewElement: A.SvgElement, SVGGradientElement: A.SvgElement, SVGComponentTransferFunctionElement: A.SvgElement, SVGFEDropShadowElement: A.SvgElement, SVGMPathElement: A.SvgElement});
-    hunkHelpers.setOrUpdateLeafTags({DOMError: true, MediaError: true, NavigatorUserMediaError: true, OverconstrainedError: true, PositionError: true, GeolocationPositionError: true, HTMLAudioElement: true, HTMLBRElement: true, HTMLBaseElement: true, HTMLBodyElement: true, HTMLCanvasElement: true, HTMLContentElement: true, HTMLDListElement: true, HTMLDataElement: true, HTMLDataListElement: true, HTMLDetailsElement: true, HTMLDialogElement: true, HTMLDivElement: true, HTMLEmbedElement: true, HTMLFieldSetElement: true, HTMLHRElement: true, HTMLHeadElement: true, HTMLHeadingElement: true, HTMLHtmlElement: true, HTMLIFrameElement: true, HTMLInputElement: true, HTMLLIElement: true, HTMLLabelElement: true, HTMLLegendElement: true, HTMLLinkElement: true, HTMLMapElement: true, HTMLMediaElement: true, HTMLMenuElement: true, HTMLMetaElement: true, HTMLMeterElement: true, HTMLModElement: true, HTMLOListElement: true, HTMLObjectElement: true, HTMLOptGroupElement: true, HTMLOptionElement: true, HTMLOutputElement: true, HTMLParagraphElement: true, HTMLParamElement: true, HTMLPictureElement: true, HTMLPreElement: true, HTMLProgressElement: true, HTMLQuoteElement: true, HTMLScriptElement: true, HTMLShadowElement: true, HTMLSlotElement: true, HTMLSourceElement: true, HTMLSpanElement: true, HTMLStyleElement: true, HTMLTableCaptionElement: true, HTMLTableCellElement: true, HTMLTableDataCellElement: true, HTMLTableHeaderCellElement: true, HTMLTableColElement: true, HTMLTableElement: true, HTMLTableRowElement: true, HTMLTableSectionElement: true, HTMLTemplateElement: true, HTMLTextAreaElement: true, HTMLTimeElement: true, HTMLTitleElement: true, HTMLTrackElement: true, HTMLUListElement: true, HTMLUnknownElement: true, HTMLVideoElement: true, HTMLDirectoryElement: true, HTMLFontElement: true, HTMLFrameElement: true, HTMLFrameSetElement: true, HTMLMarqueeElement: true, HTMLElement: false, HTMLAnchorElement: true, HTMLAreaElement: true, HTMLButtonElement: true, CDATASection: true, CharacterData: true, Comment: true, ProcessingInstruction: true, Text: true, DOMException: true, DOMTokenList: true, MathMLElement: true, Element: false, AbortPaymentEvent: true, AnimationEvent: true, AnimationPlaybackEvent: true, ApplicationCacheErrorEvent: true, BackgroundFetchClickEvent: true, BackgroundFetchEvent: true, BackgroundFetchFailEvent: true, BackgroundFetchedEvent: true, BeforeInstallPromptEvent: true, BeforeUnloadEvent: true, BlobEvent: true, CanMakePaymentEvent: true, ClipboardEvent: true, CloseEvent: true, CompositionEvent: true, CustomEvent: true, DeviceMotionEvent: true, DeviceOrientationEvent: true, ErrorEvent: true, Event: true, InputEvent: true, SubmitEvent: true, ExtendableEvent: true, ExtendableMessageEvent: true, FetchEvent: true, FocusEvent: true, FontFaceSetLoadEvent: true, ForeignFetchEvent: true, GamepadEvent: true, HashChangeEvent: true, InstallEvent: true, KeyboardEvent: true, MediaEncryptedEvent: true, MediaKeyMessageEvent: true, MediaQueryListEvent: true, MediaStreamEvent: true, MediaStreamTrackEvent: true, MessageEvent: true, MIDIConnectionEvent: true, MIDIMessageEvent: true, MouseEvent: true, DragEvent: true, MutationEvent: true, NotificationEvent: true, PageTransitionEvent: true, PaymentRequestEvent: true, PaymentRequestUpdateEvent: true, PointerEvent: true, PopStateEvent: true, PresentationConnectionAvailableEvent: true, PresentationConnectionCloseEvent: true, ProgressEvent: true, PromiseRejectionEvent: true, PushEvent: true, RTCDataChannelEvent: true, RTCDTMFToneChangeEvent: true, RTCPeerConnectionIceEvent: true, RTCTrackEvent: true, SecurityPolicyViolationEvent: true, SensorErrorEvent: true, SpeechRecognitionError: true, SpeechRecognitionEvent: true, SpeechSynthesisEvent: true, StorageEvent: true, SyncEvent: true, TextEvent: true, TouchEvent: true, TrackEvent: true, TransitionEvent: true, WebKitTransitionEvent: true, UIEvent: true, VRDeviceEvent: true, VRDisplayEvent: true, VRSessionEvent: true, WheelEvent: true, MojoInterfaceRequestEvent: true, ResourceProgressEvent: true, USBConnectionEvent: true, IDBVersionChangeEvent: true, AudioProcessingEvent: true, OfflineAudioCompletionEvent: true, WebGLContextEvent: true, EventTarget: false, HTMLFormElement: true, HTMLCollection: true, HTMLFormControlsCollection: true, HTMLOptionsCollection: true, HTMLImageElement: true, Document: true, DocumentFragment: true, HTMLDocument: true, ShadowRoot: true, XMLDocument: true, Attr: true, DocumentType: true, Node: false, NodeList: true, RadioNodeList: true, HTMLSelectElement: true, NamedNodeMap: true, MozNamedAttrMap: true, SVGAElement: true, SVGAnimateElement: true, SVGAnimateMotionElement: true, SVGAnimateTransformElement: true, SVGAnimationElement: true, SVGCircleElement: true, SVGClipPathElement: true, SVGDefsElement: true, SVGDescElement: true, SVGDiscardElement: true, SVGEllipseElement: true, SVGFEBlendElement: true, SVGFEColorMatrixElement: true, SVGFEComponentTransferElement: true, SVGFECompositeElement: true, SVGFEConvolveMatrixElement: true, SVGFEDiffuseLightingElement: true, SVGFEDisplacementMapElement: true, SVGFEDistantLightElement: true, SVGFEFloodElement: true, SVGFEFuncAElement: true, SVGFEFuncBElement: true, SVGFEFuncGElement: true, SVGFEFuncRElement: true, SVGFEGaussianBlurElement: true, SVGFEImageElement: true, SVGFEMergeElement: true, SVGFEMergeNodeElement: true, SVGFEMorphologyElement: true, SVGFEOffsetElement: true, SVGFEPointLightElement: true, SVGFESpecularLightingElement: true, SVGFESpotLightElement: true, SVGFETileElement: true, SVGFETurbulenceElement: true, SVGFilterElement: true, SVGForeignObjectElement: true, SVGGElement: true, SVGGeometryElement: true, SVGGraphicsElement: true, SVGImageElement: true, SVGLineElement: true, SVGLinearGradientElement: true, SVGMarkerElement: true, SVGMaskElement: true, SVGMetadataElement: true, SVGPathElement: true, SVGPatternElement: true, SVGPolygonElement: true, SVGPolylineElement: true, SVGRadialGradientElement: true, SVGRectElement: true, SVGScriptElement: true, SVGSetElement: true, SVGStopElement: true, SVGStyleElement: true, SVGElement: true, SVGSVGElement: true, SVGSwitchElement: true, SVGSymbolElement: true, SVGTSpanElement: true, SVGTextContentElement: true, SVGTextElement: true, SVGTextPathElement: true, SVGTextPositioningElement: true, SVGTitleElement: true, SVGUseElement: true, SVGViewElement: true, SVGGradientElement: true, SVGComponentTransferFunctionElement: true, SVGFEDropShadowElement: true, SVGMPathElement: true});
-  })();
-  convertAllToFastObject(holders);
-  convertToFastObject($);
-  (function(callback) {
-    if (typeof document === "undefined") {
-      callback(null);
-      return;
-    }
-    if (typeof document.currentScript != "undefined") {
-      callback(document.currentScript);
-      return;
-    }
-    var scripts = document.scripts;
-    function onLoad(event) {
-      for (var i = 0; i < scripts.length; ++i)
-        scripts[i].removeEventListener("load", onLoad, false);
-      callback(event.target);
-    }
-    for (var i = 0; i < scripts.length; ++i)
-      scripts[i].addEventListener("load", onLoad, false);
-  })(function(currentScript) {
-    init.currentScript = currentScript;
-    var callMain = A.main;
-    if (typeof dartMainRunner === "function")
-      dartMainRunner(callMain, []);
-    else
-      callMain([]);
-  });
-})();
-
-//# sourceMappingURL=script.js.map
