--- docs/abstract-factory-2/style.css
@@ -1,75 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto');
-
-
-* {
-    box-sizing: border-box;
-}
-
-
-body {
-    font-family: 'Roboto', sans-serif;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100vh;
-    margin: 0;
-}
-
-#site-header {
-    background-color: #333;
-    height: 5.5vh;
-    width: 100vw;
-    justify-content: space-between;
-    color: white;
-    display: flex;
-}
-
-#site-header a {
-    text-decoration: none;
-    color: white;
-    margin: 1vw;
-}
-
-#site-header a:hover {
-    color: #ddd;
-}
-
-#site-header .left-box {
-    margin: 0 1vh;
-    display: flex;
-}
-
-#site-header .right-box {
-    margin: 0 1vh;
-    display: flex;
-}
-
-#site-header .anchor-tags {
-    margin: auto;
-    display: flex;
-    flex-direction: row;
-}
-
-#demo-area {
-    margin-top: 5vh;
-    height: 80vh;
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-    flex-direction: row;
-}
-
-.board-row {
-    display: flex;
-    flex-direction: row;
-}
-
-
-/* Game Pieces */
-
-.piece-img {
-    width: 100%;
-}
-
