--- docs/abstract-factory-2/whitepieces.dart
@@ -1,25 +1,0 @@
-import "abstractfactory.dart";
-
-class WhitePawn extends Pawn {
-  WhitePawn() : super("w") {}
-}
-
-class WhiteKnight extends Knight {
-  WhiteKnight() : super("w") {}
-}
-
-class WhiteBishop extends Bishop {
-  WhiteBishop() : super("w") {}
-}
-
-class WhiteRook extends Rook {
-  WhiteRook() : super("w") {}
-}
-
-class WhiteKing extends King {
-  WhiteKing() : super("w") {}
-}
-
-class WhiteQueen extends Queen {
-  WhiteQueen() : super("w") {}
-}
