+++ docs/abstract-factory/blackpieces.dart
@@ -1,4 +1,4 @@
-import 'chess.dart';
+import "abstractfactory.dart";
 
 class BlackPawn extends Pawn {
   BlackPawn() : super("b") {}
