+++ docs/abstract-factory/script.dart
@@ -2,7 +2,8 @@ import 'dart:html';
 
 // game-specific classes
 import 'abstractfactory.dart';
-import 'chess.dart';
+import './chess/chess_game.dart';
+import './chess/chess_pieces.dart';
 
 class PieceBox {
   ChessGame game;
@@ -28,30 +29,62 @@ class PieceBox {
 
     button.classes.remove("${current}-btn");
     button.classes.add("${factory.colour}-btn");
-    button.innerText = factory.colour;
 
+    String title = (factory.colour == "black") ? "Black" : "White";
+    button.innerText = "${title}PieceFactory";
+
+    presentPieces();
+  }
+
+  void activateSelection(ChessPiece piece) {
+    print("selection indicated! ${piece.name} ${piece.colour}");
+    selection = piece;
+    game.threatenAllWith(piece);
     presentPieces();
   }
 
   void presentPieces() {
     container.children.clear();
-
-    for (ChessPiece piece in getOptions()) {
-      container.children.add(piece.element);
+    for (PieceOption option in getOptions()) {
+      container.children.add(option.element);
     }
   }
 
-  List<ChessPiece> getOptions() {
-    List<ChessPiece> options = List.empty(growable: true);
+  List<PieceOption> getOptions() {
+    List<PieceOption> options = List.empty(growable: true);
 
-    options.add(factory.createPawn());
-    options.add(factory.createBishop());
-    options.add(factory.createKnight());
-    options.add(factory.createRook());
-    options.add(factory.createQueen());
-    options.add(factory.createKing());
+    options.add(PieceOption(this, factory.createPawn()));
+    options.add(PieceOption(this, factory.createBishop()));
+    options.add(PieceOption(this, factory.createKnight()));
+    options.add(PieceOption(this, factory.createRook()));
+    options.add(PieceOption(this, factory.createQueen()));
+    options.add(PieceOption(this, factory.createKing()));
 
     return options;
+  }
+}
+
+class PieceOption {
+  PieceBox parent;
+  ChessPiece piece;
+  late Element element;
+
+  PieceOption(this.parent, this.piece) {
+    element = createElement();
+  }
+
+  Element createElement() {
+    Element element = document.createElement("div");
+
+    element.classes.add("piece-option");
+
+    element.addEventListener("click", (event) {
+      parent.activateSelection(piece);
+    });
+
+    element.children.add(piece.getElement());
+
+    return element;
   }
 }
 
@@ -64,7 +97,16 @@ void setupBox() {
 
   if (gameContainer is Element) {
     ChessGame game = ChessGame(gameContainer);
-    game.createBoard();
+    game.startGame();
+
+    List<bool> initState = List.empty(growable: true);
+    initState.add(false);
+    initState.add(false);
+    initState.add(false);
+    initState.add(false);
+    initState.add(false);
+    initState.add(false);
+    game.setupPieces(initState);
 
     Element? switchBtn = document.querySelector(".colour-switcher");
     Element? pieceContainer = document.getElementById("piece-box");
