+++ docs/abstract-factory/script.js
@@ -1,4 +1,4 @@
-// Generated by dart2js (NullSafetyMode.sound, csp, deferred-serialization, intern-composite-values), the Dart to JavaScript compiler version: 3.0.6.
+// Generated by dart2js (NullSafetyMode.sound, csp, deferred-serialization, intern-composite-values), the Dart to JavaScript compiler version: 3.0.7.
 // The code supports the following hooks:
 // dartPrint(message):
 //    if this function is defined it is called instead of the Dart [print]
@@ -354,19 +354,25 @@
         return object.toString$0(0);
       return "Instance of '" + A.Primitives_objectTypeName(object) + "'";
     },
+    iae(argument) {
+      throw A.wrapException(A.argumentErrorValue(argument));
+    },
     ioore(receiver, index) {
       if (receiver == null)
         J.get$length$asx(receiver);
       throw A.wrapException(A.diagnoseIndexError(receiver, index));
     },
     diagnoseIndexError(indexable, index) {
-      var $length, _s5_ = "index";
+      var $length, _s5_ = "index", _null = null;
       if (!A._isInt(index))
-        return new A.ArgumentError(true, index, _s5_, null);
+        return new A.ArgumentError(true, index, _s5_, _null);
       $length = A._asInt(J.get$length$asx(indexable));
       if (index < 0 || index >= $length)
-        return A.IndexError$withLength(index, $length, indexable, _s5_);
-      return new A.RangeError(null, null, true, index, _s5_, "Value not in range");
+        return A.IndexError$withLength(index, $length, indexable, _null, _s5_);
+      return new A.RangeError(_null, _null, true, index, _s5_, "Value not in range");
+    },
+    argumentErrorValue(object) {
+      return new A.ArgumentError(true, object, null, null);
     },
     wrapException(ex) {
       var wrapper, t1;
@@ -396,6 +402,13 @@
         return J.get$hashCode$(object);
       else
         return A.Primitives_objectHashCode(object);
+    },
+    fillLiteralSet(values, result) {
+      var index,
+        $length = values.length;
+      for (index = 0; index < $length; ++index)
+        result.add$1(0, values[index]);
+      return result;
     },
     invokeClosure(closure, numberOfArguments, arg1, arg2, arg3, arg4) {
       type$.Function._as(closure);
@@ -2439,8 +2452,20 @@
     _TypeError: function _TypeError(t0) {
       this.__rti$_message = t0;
     },
+    HashSet_HashSet($E) {
+      return new A._HashSet($E._eval$1("_HashSet<0>"));
+    },
+    _HashSet__newHashTable() {
+      var table = Object.create(null);
+      table["<non-identifier-key>"] = table;
+      delete table["<non-identifier-key>"];
+      return table;
+    },
     LinkedHashSet_LinkedHashSet($E) {
       return new A._LinkedHashSet($E._eval$1("_LinkedHashSet<0>"));
+    },
+    LinkedHashSet_LinkedHashSet$_literal(values, $E) {
+      return $E._eval$1("LinkedHashSet<0>")._as(A.fillLiteralSet(values, new A._LinkedHashSet($E._eval$1("_LinkedHashSet<0>"))));
     },
     _LinkedHashSet__newHashTable() {
       var table = Object.create(null);
@@ -2452,6 +2477,20 @@
       var t1 = new A._LinkedHashSetIterator(_set, _modifications, $E._eval$1("_LinkedHashSetIterator<0>"));
       t1._cell = _set._first;
       return t1;
+    },
+    _HashSet: function _HashSet(t0) {
+      var _ = this;
+      _._collection$_length = 0;
+      _._elements = _._collection$_rest = _._nums = _._strings = null;
+      _.$ti = t0;
+    },
+    _HashSetIterator: function _HashSetIterator(t0, t1, t2) {
+      var _ = this;
+      _._set = t0;
+      _._elements = t1;
+      _._offset = 0;
+      _._collection$_current = null;
+      _.$ti = t2;
     },
     _LinkedHashSet: function _LinkedHashSet(t0) {
       var _ = this;
@@ -2479,8 +2518,8 @@
     },
     List_List$filled($length, fill, growable, $E) {
       var i,
-        result = J.JSArray_JSArray$growable($length, $E);
-      if ($length !== 0 && true)
+        result = growable ? J.JSArray_JSArray$growable($length, $E) : J.JSArray_JSArray$fixed($length, $E);
+      if ($length !== 0 && fill != null)
         for (i = 0; i < result.length; ++i)
           result[i] = fill;
       return result;
@@ -2546,7 +2585,7 @@
         throw A.wrapException(A.RangeError$range(value, 0, null, $name, null));
       return value;
     },
-    IndexError$withLength(invalidValue, $length, indexable, $name) {
+    IndexError$withLength(invalidValue, $length, indexable, message, $name) {
       return new A.IndexError($length, true, invalidValue, $name, "Index out of range");
     },
     UnsupportedError$(message) {
@@ -2675,6 +2714,9 @@
         B.JSArray_methods.add$1(parts, elision);
       B.JSArray_methods.add$1(parts, penultimateString);
       B.JSArray_methods.add$1(parts, ultimateString);
+    },
+    print(object) {
+      A.printString(object);
     },
     Error: function Error() {
     },
@@ -2823,45 +2865,6 @@
     WhitePieceFactory: function WhitePieceFactory(t0) {
       this.colour = t0;
     },
-    BlackPawn: function BlackPawn() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    BlackKnight: function BlackKnight() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    BlackBishop: function BlackBishop() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    BlackRook: function BlackRook() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    BlackKing: function BlackKing() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    BlackQueen: function BlackQueen() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
-    },
-    Game: function Game() {
-    },
-    GameBoard: function GameBoard() {
-    },
-    GamePiece: function GamePiece() {
-    },
-    ChessGame: function ChessGame(t0) {
-      this.container = t0;
-    },
-    ChessBoard: function ChessBoard(t0, t1, t2) {
-      this.board = t0;
-      this.game = t1;
-      this.container = t2;
-    },
-    ChessBoard_createTile_closure: function ChessBoard_createTile_closure(t0, t1, t2) {
-      this.$this = t0;
-      this.i = t1;
-      this.j = t2;
-    },
-    ChessPiece: function ChessPiece() {
-    },
     Pawn: function Pawn() {
     },
     Knight: function Knight() {
@@ -2874,48 +2877,292 @@
     },
     Queen: function Queen() {
     },
+    BlackPawn: function BlackPawn(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    BlackKnight: function BlackKnight(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    BlackBishop: function BlackBishop(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    BlackRook: function BlackRook(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    BlackKing: function BlackKing(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    BlackQueen: function BlackQueen(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    ChequeredBoard: function ChequeredBoard(t0) {
+      this.pieces = t0;
+    },
+    BoardWithPieces: function BoardWithPieces() {
+    },
+    BoardWithPawns: function BoardWithPawns(t0) {
+      this.base = t0;
+    },
+    BoardWithBishops: function BoardWithBishops(t0) {
+      this.base = t0;
+    },
+    BoardWithKnights: function BoardWithKnights(t0) {
+      this.base = t0;
+    },
+    BoardWithRooks: function BoardWithRooks(t0) {
+      this.base = t0;
+    },
+    BoardWithKings: function BoardWithKings(t0) {
+      this.base = t0;
+    },
+    BoardWithQueens: function BoardWithQueens(t0) {
+      this.base = t0;
+    },
+    ChessGame: function ChessGame(t0) {
+      var _ = this;
+      _.turnCount = 0;
+      _.board = t0;
+      _.__ChessGame_logic_A = _.__ChessGame_view_A = $;
+    },
+    ChessLogic: function ChessLogic(t0, t1) {
+      var _ = this;
+      _.game = t0;
+      _.whiteKing = _.blackKing = null;
+      _.activePiece = t1;
+    },
+    ChessPiece$(colour, $name, moveStrategy) {
+      var t1 = new A.ChessPiece(moveStrategy, colour, $name);
+      t1.__GamePiece_src_A = "./assets/chess/" + $name + "_" + colour + ".png";
+      return t1;
+    },
+    ChessKing$(colour) {
+      var t1 = new A.ChessKing(new A.KingMovement(), colour, "king");
+      t1.__GamePiece_src_A = "./assets/chess/king_" + colour + ".png";
+      return t1;
+    },
+    EmptyPiece$(iPosition, jPosition) {
+      var t1 = new A.EmptyPiece(new A.NoMovement(), "empty", "empty");
+      t1.__GamePiece_src_A = "./assets/chess/empty_empty.png";
+      t1.i = iPosition;
+      t1.j = jPosition;
+      return t1;
+    },
+    ChessPiece: function ChessPiece(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    ChessKing: function ChessKing(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    EmptyPiece: function EmptyPiece(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    NoMovement: function NoMovement() {
+    },
+    PawnMovement: function PawnMovement() {
+    },
+    KnightMovement: function KnightMovement() {
+    },
+    BishopMovement: function BishopMovement() {
+    },
+    RookMovement: function RookMovement() {
+    },
+    QueenMovement: function QueenMovement() {
+    },
+    KingMovement: function KingMovement() {
+    },
+    ChessBoardView: function ChessBoardView(t0, t1) {
+      this.container = t0;
+      this.game = t1;
+    },
+    ChessBoardView_createTile_closure: function ChessBoardView_createTile_closure(t0, t1) {
+      this.$this = t0;
+      this.piece = t1;
+    },
+    GamePiece: function GamePiece() {
+    },
+    PieceOption$($parent, piece) {
+      var t1 = new A.PieceOption($parent, piece);
+      t1.__PieceOption_element_A = t1.createElement$0(0);
+      return t1;
+    },
     main() {
-      var t3, switchBtn, pieceContainer,
+      var t3, game, initState, switchBtn, pieceContainer,
         t1 = document,
         gameContainer = t1.getElementById("game-container"),
         t2 = type$.Element;
       if (t2._is(gameContainer)) {
-        t3 = J.JSArray_JSArray$growable(0, type$.List_Element);
-        new A.ChessBoard(t3, new A.ChessGame(gameContainer), gameContainer).insertTiles$0();
+        t3 = J.JSArray_JSArray$growable(0, type$.List_GamePiece);
+        game = new A.ChessGame(new A.ChequeredBoard(t3));
+        game.__ChessGame_view_A = new A.ChessBoardView(gameContainer, game);
+        game.__ChessGame_logic_A = new A.ChessLogic(game, A.EmptyPiece$(0, 0));
+        t3 = type$.bool;
+        initState = J.JSArray_JSArray$growable(0, t3);
+        B.JSArray_methods.add$1(initState, true);
+        B.JSArray_methods.add$1(initState, true);
+        B.JSArray_methods.add$1(initState, true);
+        B.JSArray_methods.add$1(initState, true);
+        B.JSArray_methods.add$1(initState, true);
+        B.JSArray_methods.add$1(initState, true);
+        game.setupPieces$1(initState);
+        game.__ChessGame_logic_A.pairKings$0();
+        game.refreshView$0();
+        initState = J.JSArray_JSArray$growable(0, t3);
+        B.JSArray_methods.add$1(initState, false);
+        B.JSArray_methods.add$1(initState, false);
+        B.JSArray_methods.add$1(initState, false);
+        B.JSArray_methods.add$1(initState, false);
+        B.JSArray_methods.add$1(initState, false);
+        B.JSArray_methods.add$1(initState, false);
+        game.setupPieces$1(initState);
         switchBtn = t1.querySelector(".colour-switcher");
         pieceContainer = t1.getElementById("piece-box");
         if (type$.ButtonElement._is(switchBtn) && t2._is(pieceContainer)) {
-          t1 = new A.PieceBox(pieceContainer, switchBtn, new A.BlackPieceFactory("black"));
+          t1 = new A.PieceBox(game, pieceContainer, switchBtn, new A.BlackPieceFactory("black"));
           t1.armButton$0();
           t1.switchColour$0();
         }
       }
     },
-    PieceBox: function PieceBox(t0, t1, t2) {
-      this.container = t0;
-      this.button = t1;
-      this.factory = t2;
+    PieceBox: function PieceBox(t0, t1, t2, t3) {
+      var _ = this;
+      _.game = t0;
+      _.container = t1;
+      _.button = t2;
+      _.factory = t3;
     },
     PieceBox_armButton_closure: function PieceBox_armButton_closure(t0) {
       this.$this = t0;
     },
-    WhitePawn: function WhitePawn() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    PieceOption: function PieceOption(t0, t1) {
+      this.parent = t0;
+      this.piece = t1;
+      this.__PieceOption_element_A = $;
     },
-    WhiteKnight: function WhiteKnight() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    PieceOption_createElement_closure: function PieceOption_createElement_closure(t0) {
+      this.$this = t0;
     },
-    WhiteBishop: function WhiteBishop() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    WhitePawn: function WhitePawn(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
     },
-    WhiteRook: function WhiteRook() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    WhiteKnight: function WhiteKnight(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
     },
-    WhiteKing: function WhiteKing() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    WhiteBishop: function WhiteBishop(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
     },
-    WhiteQueen: function WhiteQueen() {
-      this.__GamePiece_element_A = this.__GamePiece_src_A = $;
+    WhiteRook: function WhiteRook(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    WhiteKing: function WhiteKing(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
+    },
+    WhiteQueen: function WhiteQueen(t0, t1, t2) {
+      var _ = this;
+      _.moveStrategy = t0;
+      _.colour = t1;
+      _.name = t2;
+      _.threatened = _.hasMoved = false;
+      _.myKing = null;
+      _.__GamePiece_src_A = $;
+      _.j = _.i = 0;
     },
     printString(string) {
       if (typeof dartPrint == "function") {
@@ -2993,10 +3240,18 @@
       }
       return B.UnknownJavaScriptObject_methods;
     },
+    JSArray_JSArray$fixed($length, $E) {
+      if ($length < 0 || $length > 4294967295)
+        throw A.wrapException(A.RangeError$range($length, 0, 4294967295, "length", null));
+      return J.JSArray_JSArray$markFixed(new Array($length), $E);
+    },
     JSArray_JSArray$growable($length, $E) {
       if ($length < 0)
         throw A.wrapException(A.ArgumentError$("Length must be a non-negative integer: " + $length));
       return A._setArrayType(new Array($length), $E._eval$1("JSArray<0>"));
+    },
+    JSArray_JSArray$markFixed(allocation, $E) {
+      return J.JSArray_markFixedList(A._setArrayType(allocation, $E._eval$1("JSArray<0>")), $E);
     },
     JSArray_markFixedList(list, $T) {
       list.fixed$length = Array;
@@ -3114,6 +3369,18 @@
         return receiver;
       return J.getNativeInterceptor(receiver);
     },
+    getInterceptor$in(receiver) {
+      if (typeof receiver == "number") {
+        if (Math.floor(receiver) == receiver)
+          return J.JSInt.prototype;
+        return J.JSNumNotInt.prototype;
+      }
+      if (receiver == null)
+        return receiver;
+      if (!(receiver instanceof A.Object))
+        return J.UnknownJavaScriptObject.prototype;
+      return receiver;
+    },
     getInterceptor$s(receiver) {
       if (typeof receiver == "string")
         return J.JSString.prototype;
@@ -3134,6 +3401,9 @@
       if (receiver instanceof A.Object)
         return receiver;
       return J.getNativeInterceptor(receiver);
+    },
+    get$children$x(receiver) {
+      return J.getInterceptor$x(receiver).get$children(receiver);
     },
     get$classes$x(receiver) {
       return J.getInterceptor$x(receiver).get$classes(receiver);
@@ -3166,6 +3436,11 @@
     },
     _clearChildren$0$x(receiver) {
       return J.getInterceptor$x(receiver)._clearChildren$0(receiver);
+    },
+    abs$0$in(receiver) {
+      if (typeof receiver === "number")
+        return Math.abs(receiver);
+      return J.getInterceptor$in(receiver).abs$0(receiver);
     },
     elementAt$1$ax(receiver, a0) {
       return J.getInterceptor$ax(receiver).elementAt$1(receiver, a0);
@@ -3237,6 +3512,9 @@
     toString$0(receiver) {
       return String(receiver);
     },
+    $and(receiver, other) {
+      return other && receiver;
+    },
     get$hashCode(receiver) {
       return receiver ? 519018 : 218159;
     },
@@ -3303,12 +3581,10 @@
       return receiver.length;
     },
     $indexSet(receiver, index, value) {
-      var t1;
       A._arrayInstanceType(receiver)._precomputed1._as(value);
       if (!!receiver.immutable$list)
         A.throwExpression(A.UnsupportedError$("indexed set"));
-      t1 = receiver.length;
-      if (index >= t1)
+      if (!(index >= 0 && index < receiver.length))
         throw A.wrapException(A.diagnoseIndexError(receiver, index));
       receiver[index] = value;
     },
@@ -3361,12 +3637,23 @@
       scaled = absolute < 1 ? absolute / factor : factor / absolute;
       return ((scaled * 9007199254740992 | 0) + (scaled * 3542243181176521 | 0)) * 599197 + floorLog2 * 1259 & 536870911;
     },
+    $mod(receiver, other) {
+      var result = receiver % other;
+      if (result === 0)
+        return 0;
+      if (result > 0)
+        return result;
+      return result + other;
+    },
     get$runtimeType(receiver) {
       return A.createRuntimeType(type$.num);
     },
     $isnum: 1
   };
   J.JSInt.prototype = {
+    abs$0(receiver) {
+      return Math.abs(receiver);
+    },
     get$runtimeType(receiver) {
       return A.createRuntimeType(type$.int);
     },
@@ -3619,6 +3906,148 @@
     }
   };
   A._TypeError.prototype = {};
+  A._HashSet.prototype = {
+    get$iterator(_) {
+      return new A._HashSetIterator(this, this._computeElements$0(), A._instanceType(this)._eval$1("_HashSetIterator<1>"));
+    },
+    get$length(_) {
+      return this._collection$_length;
+    },
+    contains$1(_, object) {
+      var strings, t1;
+      if (object !== "__proto__") {
+        strings = this._strings;
+        return strings == null ? false : strings[object] != null;
+      } else {
+        t1 = this._contains$1(object);
+        return t1;
+      }
+    },
+    _contains$1(object) {
+      var rest = this._collection$_rest;
+      if (rest == null)
+        return false;
+      return this._findBucketIndex$2(rest[this._computeHashCode$1(object)], object) >= 0;
+    },
+    add$1(_, element) {
+      var strings, t1, _this = this;
+      A._instanceType(_this)._precomputed1._as(element);
+      if (element !== "__proto__") {
+        strings = _this._strings;
+        return _this._addHashTableEntry$2(strings == null ? _this._strings = A._HashSet__newHashTable() : strings, element);
+      } else {
+        t1 = _this._add$1(element);
+        return t1;
+      }
+    },
+    _add$1(element) {
+      var rest, hash, bucket, _this = this;
+      A._instanceType(_this)._precomputed1._as(element);
+      rest = _this._collection$_rest;
+      if (rest == null)
+        rest = _this._collection$_rest = A._HashSet__newHashTable();
+      hash = _this._computeHashCode$1(element);
+      bucket = rest[hash];
+      if (bucket == null)
+        rest[hash] = [element];
+      else {
+        if (_this._findBucketIndex$2(bucket, element) >= 0)
+          return false;
+        bucket.push(element);
+      }
+      ++_this._collection$_length;
+      _this._elements = null;
+      return true;
+    },
+    _computeElements$0() {
+      var strings, names, entries, index, i, nums, rest, bucket, $length, i0, _this = this,
+        result = _this._elements;
+      if (result != null)
+        return result;
+      result = A.List_List$filled(_this._collection$_length, null, false, type$.dynamic);
+      strings = _this._strings;
+      if (strings != null) {
+        names = Object.getOwnPropertyNames(strings);
+        entries = names.length;
+        for (index = 0, i = 0; i < entries; ++i) {
+          result[index] = names[i];
+          ++index;
+        }
+      } else
+        index = 0;
+      nums = _this._nums;
+      if (nums != null) {
+        names = Object.getOwnPropertyNames(nums);
+        entries = names.length;
+        for (i = 0; i < entries; ++i) {
+          result[index] = +names[i];
+          ++index;
+        }
+      }
+      rest = _this._collection$_rest;
+      if (rest != null) {
+        names = Object.getOwnPropertyNames(rest);
+        entries = names.length;
+        for (i = 0; i < entries; ++i) {
+          bucket = rest[names[i]];
+          $length = bucket.length;
+          for (i0 = 0; i0 < $length; ++i0) {
+            result[index] = bucket[i0];
+            ++index;
+          }
+        }
+      }
+      return _this._elements = result;
+    },
+    _addHashTableEntry$2(table, element) {
+      A._instanceType(this)._precomputed1._as(element);
+      if (table[element] != null)
+        return false;
+      table[element] = 0;
+      ++this._collection$_length;
+      this._elements = null;
+      return true;
+    },
+    _computeHashCode$1(element) {
+      return B.JSString_methods.get$hashCode(element) & 1073741823;
+    },
+    _findBucketIndex$2(bucket, element) {
+      var $length, i;
+      if (bucket == null)
+        return -1;
+      $length = bucket.length;
+      for (i = 0; i < $length; ++i)
+        if (J.$eq$(bucket[i], element))
+          return i;
+      return -1;
+    }
+  };
+  A._HashSetIterator.prototype = {
+    get$current() {
+      var t1 = this._collection$_current;
+      return t1 == null ? this.$ti._precomputed1._as(t1) : t1;
+    },
+    moveNext$0() {
+      var _this = this,
+        elements = _this._elements,
+        offset = _this._offset,
+        t1 = _this._set;
+      if (elements !== t1._elements)
+        throw A.wrapException(A.ConcurrentModificationError$(t1));
+      else if (offset >= elements.length) {
+        _this.set$_collection$_current(null);
+        return false;
+      } else {
+        _this.set$_collection$_current(elements[offset]);
+        _this._offset = offset + 1;
+        return true;
+      }
+    },
+    set$_collection$_current(_current) {
+      this._collection$_current = this.$ti._eval$1("1?")._as(_current);
+    },
+    $isIterator: 1
+  };
   A._LinkedHashSet.prototype = {
     get$iterator(_) {
       var _this = this,
@@ -3695,7 +4124,8 @@
         if (J.$eq$(bucket[i]._collection$_element, element))
           return i;
       return -1;
-    }
+    },
+    $isLinkedHashSet: 1
   };
   A._LinkedHashSetCell.prototype = {};
   A._LinkedHashSetIterator.prototype = {
@@ -3733,7 +4163,7 @@
     get$isEmpty(receiver) {
       return this.get$length(receiver) === 0;
     },
-    toList$0(receiver) {
+    toList$1$growable(receiver, growable) {
       var t1, first, result, i, _this = this;
       if (_this.get$isEmpty(receiver)) {
         t1 = J.JSArray_JSArray$growable(0, A.instanceType(receiver)._eval$1("ListBase.E"));
@@ -3744,6 +4174,9 @@
       for (i = 1; i < _this.get$length(receiver); ++i)
         B.JSArray_methods.$indexSet(result, i, _this.$index(receiver, i));
       return result;
+    },
+    toList$0($receiver) {
+      return this.toList$1$growable($receiver, true);
     },
     toString$0(receiver) {
       return A.Iterable_iterableToFullString(receiver, "[", "]");
@@ -3756,44 +4189,36 @@
       return A.Iterable_iterableToFullString(this, "{", "}");
     },
     join$1(_, separator) {
-      var t1, first, t2, t3,
+      var first, t1,
         iterator = this.get$iterator(this);
       if (!iterator.moveNext$0())
         return "";
-      t1 = iterator._collection$_current;
-      first = J.toString$0$(t1 == null ? iterator.$ti._precomputed1._as(t1) : t1);
+      first = J.toString$0$(iterator.get$current());
       if (!iterator.moveNext$0())
         return first;
-      t1 = iterator.$ti._precomputed1;
       if (separator.length === 0) {
-        t2 = first;
-        do {
-          t3 = iterator._collection$_current;
-          t2 += A.S(t3 == null ? t1._as(t3) : t3);
-        } while (iterator.moveNext$0());
-        t1 = t2;
+        t1 = first;
+        do
+          t1 += A.S(iterator.get$current());
+        while (iterator.moveNext$0());
       } else {
-        t2 = first;
-        do {
-          t3 = iterator._collection$_current;
-          t2 = t2 + separator + A.S(t3 == null ? t1._as(t3) : t3);
-        } while (iterator.moveNext$0());
-        t1 = t2;
+        t1 = first;
+        do
+          t1 = t1 + separator + A.S(iterator.get$current());
+        while (iterator.moveNext$0());
       }
       return t1.charCodeAt(0) == 0 ? t1 : t1;
     },
     elementAt$1(_, index) {
-      var iterator, skipCount, t1;
+      var iterator, skipCount;
       A.RangeError_checkNotNegative(index, "index");
       iterator = this.get$iterator(this);
       for (skipCount = index; iterator.moveNext$0();) {
-        if (skipCount === 0) {
-          t1 = iterator._collection$_current;
-          return t1 == null ? iterator.$ti._precomputed1._as(t1) : t1;
-        }
+        if (skipCount === 0)
+          return iterator.get$current();
         --skipCount;
       }
-      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, "index"));
+      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, null, "index"));
     },
     $isIterable: 1,
     $isSet: 1
@@ -3879,7 +4304,8 @@
   };
   A.UnimplementedError.prototype = {
     toString$0(_) {
-      return "UnimplementedError: " + this.message;
+      var message = this.message;
+      return message != null ? "UnimplementedError: " + message : "UnimplementedError";
     }
   };
   A.ConcurrentModificationError.prototype = {
@@ -3922,7 +4348,7 @@
           return iterator.get$current();
         --skipCount;
       }
-      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, "index"));
+      throw A.wrapException(A.IndexError$withLength(index, index - skipCount, this, null, "index"));
     },
     toString$0(_) {
       return A.Iterable_iterableToShortString(this, "(", ")");
@@ -4063,7 +4489,7 @@
         t2 = index >>> 0 !== index || index >= t1;
       t2.toString;
       if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null));
+        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
       t1 = receiver[index];
       t1.toString;
       return t1;
@@ -4122,7 +4548,7 @@
         t2 = index >>> 0 !== index || index >= t1;
       t2.toString;
       if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null));
+        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
       t1 = receiver[index];
       t1.toString;
       return t1;
@@ -4152,7 +4578,7 @@
         t2 = index >>> 0 !== index || index >= t1;
       t2.toString;
       if (t2)
-        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null));
+        throw A.wrapException(A.IndexError$withLength(index, t1, receiver, null, null));
       t1 = receiver[index];
       t1.toString;
       return t1;
@@ -4328,90 +4754,698 @@
   A.ChessPieceFactory.prototype = {};
   A.BlackPieceFactory.prototype = {
     createPawn$0() {
-      var t1 = new A.BlackPawn();
+      var t1 = new A.BlackPawn(new A.PawnMovement(), "b", "pawn");
       t1.__GamePiece_src_A = "./assets/chess/pawn_b.png";
-      t1.buildElement$0();
       return t1;
     },
     createKnight$0() {
-      var t1 = new A.BlackKnight();
+      var t1 = new A.BlackKnight(new A.KnightMovement(), "b", "knight");
       t1.__GamePiece_src_A = "./assets/chess/knight_b.png";
-      t1.buildElement$0();
       return t1;
     },
     createBishop$0() {
-      var t1 = new A.BlackBishop();
+      var t1 = new A.BlackBishop(new A.BishopMovement(), "b", "bishop");
       t1.__GamePiece_src_A = "./assets/chess/bishop_b.png";
-      t1.buildElement$0();
       return t1;
     },
     createRook$0() {
-      var t1 = new A.BlackRook();
+      var t1 = new A.BlackRook(new A.RookMovement(), "b", "rook");
       t1.__GamePiece_src_A = "./assets/chess/rook_b.png";
-      t1.buildElement$0();
       return t1;
     },
     createKing$0() {
-      var t1 = new A.BlackKing();
+      var t1 = new A.BlackKing(new A.KingMovement(), "b", "king");
       t1.__GamePiece_src_A = "./assets/chess/king_b.png";
-      t1.buildElement$0();
       return t1;
     },
     createQueen$0() {
-      var t1 = new A.BlackQueen();
+      var t1 = new A.BlackQueen(new A.QueenMovement(), "b", "queen");
       t1.__GamePiece_src_A = "./assets/chess/queen_b.png";
-      t1.buildElement$0();
       return t1;
     }
   };
   A.WhitePieceFactory.prototype = {
     createPawn$0() {
-      var t1 = new A.WhitePawn();
+      var t1 = new A.WhitePawn(new A.PawnMovement(), "w", "pawn");
       t1.__GamePiece_src_A = "./assets/chess/pawn_w.png";
-      t1.buildElement$0();
       return t1;
     },
     createKnight$0() {
-      var t1 = new A.WhiteKnight();
+      var t1 = new A.WhiteKnight(new A.KnightMovement(), "w", "knight");
       t1.__GamePiece_src_A = "./assets/chess/knight_w.png";
-      t1.buildElement$0();
       return t1;
     },
     createBishop$0() {
-      var t1 = new A.WhiteBishop();
+      var t1 = new A.WhiteBishop(new A.BishopMovement(), "w", "bishop");
       t1.__GamePiece_src_A = "./assets/chess/bishop_w.png";
-      t1.buildElement$0();
       return t1;
     },
     createRook$0() {
-      var t1 = new A.WhiteRook();
+      var t1 = new A.WhiteRook(new A.RookMovement(), "w", "rook");
       t1.__GamePiece_src_A = "./assets/chess/rook_w.png";
-      t1.buildElement$0();
       return t1;
     },
     createKing$0() {
-      var t1 = new A.WhiteKing();
+      var t1 = new A.WhiteKing(new A.KingMovement(), "w", "king");
       t1.__GamePiece_src_A = "./assets/chess/king_w.png";
-      t1.buildElement$0();
       return t1;
     },
     createQueen$0() {
-      var t1 = new A.WhiteQueen();
+      var t1 = new A.WhiteQueen(new A.QueenMovement(), "w", "queen");
       t1.__GamePiece_src_A = "./assets/chess/queen_w.png";
-      t1.buildElement$0();
       return t1;
     }
   };
+  A.Pawn.prototype = {};
+  A.Knight.prototype = {};
+  A.Bishop.prototype = {};
+  A.Rook.prototype = {};
+  A.King.prototype = {};
+  A.Queen.prototype = {};
   A.BlackPawn.prototype = {};
   A.BlackKnight.prototype = {};
   A.BlackBishop.prototype = {};
   A.BlackRook.prototype = {};
   A.BlackKing.prototype = {};
   A.BlackQueen.prototype = {};
-  A.Game.prototype = {};
-  A.GameBoard.prototype = {};
+  A.ChequeredBoard.prototype = {
+    setupPieces$0() {
+      var t1, t2, i, row, j;
+      for (t1 = this.pieces, t2 = type$.JSArray_ChessPiece, i = 0; i < 8; ++i) {
+        row = A._setArrayType(new Array(0), t2);
+        for (j = 0; j < 8; ++j)
+          B.JSArray_methods.add$1(row, A.EmptyPiece$(i, j));
+        B.JSArray_methods.add$1(t1, row);
+      }
+    },
+    getBoardState$0() {
+      return this.pieces;
+    },
+    placePiece$3(piece, i, j) {
+      var t1 = this.pieces;
+      if (!(i >= 0 && i < t1.length))
+        return A.ioore(t1, i);
+      B.JSArray_methods.$indexSet(t1[i], j, piece);
+      piece.i = i;
+      piece.j = j;
+    },
+    removePiece$2(i, j) {
+      var t1 = this.pieces;
+      if (!(i >= 0 && i < t1.length))
+        return A.ioore(t1, i);
+      B.JSArray_methods.$indexSet(t1[i], j, A.EmptyPiece$(i, j));
+    },
+    getPiece$2(i, j) {
+      var t1 = this.pieces;
+      if (!(i >= 0 && i < t1.length))
+        return A.ioore(t1, i);
+      t1 = t1[i];
+      if (!(j >= 0 && j < t1.length))
+        return A.ioore(t1, j);
+      return t1[j];
+    },
+    $isGameBoard: 1
+  };
+  A.BoardWithPieces.prototype = {
+    removePiece$2(i, j) {
+      this.base.removePiece$2(i, j);
+    },
+    placePiece$3(piece, i, j) {
+      this.base.placePiece$3(piece, i, j);
+    },
+    getPiece$2(i, j) {
+      return this.base.getPiece$2(i, j);
+    },
+    getBoardState$0() {
+      return this.base.getBoardState$0();
+    },
+    $isGameBoard: 1
+  };
+  A.BoardWithPawns.prototype = {
+    setupPieces$0() {
+      var j, pawn,
+        t1 = this.base;
+      t1.setupPieces$0();
+      for (j = 0; j < 8; ++j) {
+        pawn = new A.ChessPiece(new A.PawnMovement(), "b", "pawn");
+        pawn.__GamePiece_src_A = "./assets/chess/pawn_b.png";
+        t1.placePiece$3(pawn, 1, j);
+        pawn = new A.ChessPiece(new A.PawnMovement(), "w", "pawn");
+        pawn.__GamePiece_src_A = "./assets/chess/pawn_w.png";
+        t1.placePiece$3(pawn, 6, j);
+      }
+    }
+  };
+  A.BoardWithBishops.prototype = {
+    setupPieces$0() {
+      var _this = this;
+      _this.base.setupPieces$0();
+      _this.placeBishop$3("b", 0, 2);
+      _this.placeBishop$3("b", 0, 5);
+      _this.placeBishop$3("w", 7, 2);
+      _this.placeBishop$3("w", 7, 5);
+    },
+    placeBishop$3(colour, i, j) {
+      this.base.placePiece$3(A.ChessPiece$(colour, "bishop", new A.BishopMovement()), i, j);
+    }
+  };
+  A.BoardWithKnights.prototype = {
+    setupPieces$0() {
+      var _this = this;
+      _this.base.setupPieces$0();
+      _this.placeKnight$3("b", 0, 1);
+      _this.placeKnight$3("b", 0, 6);
+      _this.placeKnight$3("w", 7, 1);
+      _this.placeKnight$3("w", 7, 6);
+    },
+    placeKnight$3(colour, i, j) {
+      this.base.placePiece$3(A.ChessPiece$(colour, "knight", new A.KnightMovement()), i, j);
+    }
+  };
+  A.BoardWithRooks.prototype = {
+    setupPieces$0() {
+      var _this = this;
+      _this.base.setupPieces$0();
+      _this.placeRook$3("b", 0, 0);
+      _this.placeRook$3("b", 0, 7);
+      _this.placeRook$3("w", 7, 0);
+      _this.placeRook$3("w", 7, 7);
+    },
+    placeRook$3(colour, i, j) {
+      this.base.placePiece$3(A.ChessPiece$(colour, "rook", new A.RookMovement()), i, j);
+    }
+  };
+  A.BoardWithKings.prototype = {
+    setupPieces$0() {
+      var t1 = this.base;
+      t1.setupPieces$0();
+      t1.placePiece$3(A.ChessKing$("b"), 0, 4);
+      t1.placePiece$3(A.ChessKing$("w"), 7, 4);
+    }
+  };
+  A.BoardWithQueens.prototype = {
+    setupPieces$0() {
+      this.base.setupPieces$0();
+      this.placeQueen$3("b", 0, 3);
+      this.placeQueen$3("w", 7, 3);
+    },
+    placeQueen$3(colour, i, j) {
+      this.base.placePiece$3(A.ChessPiece$(colour, "queen", new A.QueenMovement()), i, j);
+    }
+  };
+  A.ChessGame.prototype = {
+    threatenAllWith$1(selection) {
+      var t2, t3, _i, t4, t5, t6,
+        t1 = this.__ChessGame_logic_A;
+      t1 === $ && A.throwLateFieldNI("logic");
+      t1.clearMoveOptions$0();
+      this.__ChessGame_logic_A.activePiece = selection;
+      for (t1 = this.board.getBoardState$0(), t2 = t1.length, t3 = selection.colour, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
+        for (t4 = B.JSArray_methods.get$iterator(t1[_i]); t4.moveNext$0();) {
+          t5 = t4.get$current();
+          t6 = t5.colour;
+          if (t6 !== t3)
+            t5.threatened = true;
+        }
+    },
+    refreshView$0() {
+      var t1 = this.__ChessGame_view_A;
+      t1 === $ && A.throwLateFieldNI("view");
+      t1.displayBoard$1(this.board.getBoardState$0());
+    },
+    setupPieces$1(state) {
+      var t1, t2, _this = this;
+      type$.List_bool._as(state);
+      t1 = J.JSArray_JSArray$growable(0, type$.List_GamePiece);
+      t1 = _this.board = new A.ChequeredBoard(t1);
+      t2 = state.length;
+      if (0 >= t2)
+        return A.ioore(state, 0);
+      if (state[0]) {
+        t1 = new A.BoardWithPawns(t1);
+        _this.board = t1;
+      }
+      if (1 >= t2)
+        return A.ioore(state, 1);
+      if (state[1]) {
+        t1 = new A.BoardWithBishops(t1);
+        _this.board = t1;
+      }
+      if (2 >= t2)
+        return A.ioore(state, 2);
+      if (state[2]) {
+        t1 = new A.BoardWithKnights(t1);
+        _this.board = t1;
+      }
+      if (3 >= t2)
+        return A.ioore(state, 3);
+      if (state[3]) {
+        t1 = new A.BoardWithRooks(t1);
+        _this.board = t1;
+      }
+      if (4 >= t2)
+        return A.ioore(state, 4);
+      if (state[4]) {
+        t1 = new A.BoardWithQueens(t1);
+        _this.board = t1;
+      }
+      if (5 >= t2)
+        return A.ioore(state, 5);
+      (state[5] ? _this.board = new A.BoardWithKings(t1) : t1).setupPieces$0();
+      _this.refreshView$0();
+    },
+    $isGame: 1
+  };
+  A.ChessLogic.prototype = {
+    findKings$0() {
+      var t1, t2, _i, t3, t4, t5, t6, t7;
+      for (t1 = this.game.board.getBoardState$0(), t2 = t1.length, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
+        for (t3 = B.JSArray_methods.get$iterator(t1[_i]); t3.moveNext$0();) {
+          t4 = t3.get$current();
+          if (t4 instanceof A.ChessKing) {
+            t5 = t4.colour;
+            t6 = "" + t4.i;
+            t7 = "" + t4.j;
+            if (t5 === "b") {
+              A.printString("black king found ! " + t6 + " " + t7);
+              this.blackKing = t4;
+            } else {
+              A.printString("white king found ! " + t6 + " " + t7);
+              this.whiteKing = t4;
+            }
+          }
+        }
+    },
+    pairKings$0() {
+      var t1, t2, _i, t3, t4, t5, _this = this;
+      _this.findKings$0();
+      if (_this.blackKing instanceof A.ChessKing && _this.whiteKing instanceof A.ChessKing)
+        for (t1 = _this.game.board.getBoardState$0(), t2 = t1.length, _i = 0; _i < t1.length; t1.length === t2 || (0, A.throwConcurrentModificationError)(t1), ++_i)
+          for (t3 = B.JSArray_methods.get$iterator(t1[_i]); t3.moveNext$0();) {
+            t4 = t3.get$current();
+            t5 = t4.colour;
+            if (t5 === "b")
+              t4.myKing = _this.blackKing;
+            if (t5 === "w")
+              t4.myKing = _this.whiteKing;
+          }
+    },
+    submitMove$2(i, j) {
+      var t1, t2, piece, _this = this;
+      if (_this.validMove$3(_this.activePiece, i, j)) {
+        t1 = _this.activePiece;
+        t2 = _this.game;
+        t2.board.removePiece$2(t1.i, t1.j);
+        t2.board.removePiece$2(i, j);
+        t2.board.placePiece$3(t1, i, j);
+        t1.hasMoved = true;
+        ++t2.turnCount;
+        t1 = t2.__ChessGame_logic_A;
+        t1 === $ && A.throwLateFieldNI("logic");
+        t1.activePiece = A.EmptyPiece$(0, 0);
+        t2.__ChessGame_logic_A.clearMoveOptions$0();
+        t2.refreshView$0();
+        return;
+      }
+      _this.clearMoveOptions$0();
+      _this.activePiece = A.EmptyPiece$(0, 0);
+      t1 = _this.game;
+      piece = t1.board.getPiece$2(i, j);
+      t2 = B.JSInt_methods.$mod(t1.turnCount, 2) === 0 ? "w" : "b";
+      if (piece.colour === t2) {
+        piece.moveStrategy.move$2(t1.board, piece);
+        _this.activePiece = piece;
+      }
+      t1.refreshView$0();
+    },
+    clearMoveOptions$0() {
+      var t1, i, j;
+      for (t1 = this.game, i = 0; i < 8; ++i)
+        for (j = 0; j < 8; ++j)
+          t1.board.getPiece$2(i, j).threatened = false;
+    },
+    validMove$3(piece, i, j) {
+      var t1 = this.game.board.getPiece$2(i, j).threatened;
+      return t1;
+    }
+  };
+  A.ChessPiece.prototype = {
+    canCapture$3(board, i, j) {
+      var target, t1;
+      if (this.validCoords$2(i, j)) {
+        target = board.getPiece$2(i, j);
+        t1 = this.isSafeMove$3(board, i, j);
+        if (t1)
+          if (!(target instanceof A.EmptyPiece) && target.colour !== this.colour)
+            return target.threatened = true;
+      }
+      return false;
+    },
+    canMove$3(board, i, j) {
+      var target;
+      if (this.validCoords$2(i, j)) {
+        target = board.getPiece$2(i, j);
+        if (target instanceof A.EmptyPiece && this.isSafeMove$3(board, i, j))
+          return target.threatened = true;
+      }
+      return false;
+    },
+    isSafeMove$3(board, endI, endJ) {
+      var king, verdict, t1, t2, _this = this,
+        startI = _this.i,
+        startJ = _this.j,
+        target = board.getPiece$2(endI, endJ);
+      board.removePiece$2(startI, startJ);
+      board.removePiece$2(endI, endJ);
+      board.placePiece$3(_this, endI, endJ);
+      king = _this.myKing;
+      if (king instanceof A.ChessKing && king.isTheatened$1(board)) {
+        A.print("my king is at " + king.i + " , " + king.j);
+        verdict = false;
+      } else
+        verdict = true;
+      board.placePiece$3(_this, startI, startJ);
+      board.placePiece$3(target, endI, endJ);
+      t1 = "" + endI;
+      t2 = "" + endJ;
+      if (verdict)
+        A.print("this move is safe! (" + t1 + ", " + t2 + ")");
+      else
+        A.print("this move is unsafe! (" + t1 + ", " + t2 + ")");
+      return verdict;
+    },
+    validCoords$2(i, j) {
+      var t1 = 0 <= i && i < 8;
+      return B.JSBool_methods.$and(t1, 0 <= j && j < 8);
+    }
+  };
+  A.ChessKing.prototype = {
+    isTheatened$1(board) {
+      var _this = this;
+      if (_this.attackedByPawn$1(board) || _this.attackedByKnight$1(board))
+        return true;
+      else if (_this.attackedOnDiagonals$1(board) || _this.attackedOnHorizontals$1(board))
+        return true;
+      return false;
+    },
+    attackedOnDiagonals$1(board) {
+      var t1, threats, t2, _i, _this = this,
+        diagonalThreats = A.HashSet_HashSet(type$.String);
+      diagonalThreats.add$1(0, "queen");
+      diagonalThreats.add$1(0, "bishop");
+      t1 = type$.dynamic;
+      threats = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([_this.threatAlongImpulse$3(board, -1, 1), _this.threatAlongImpulse$3(board, 1, 1), _this.threatAlongImpulse$3(board, 1, -1), _this.threatAlongImpulse$3(board, -1, -1)], t1), true, t1);
+      for (t1 = threats.length, t2 = type$.ChessPiece, _i = 0; _i < t1; ++_i)
+        if (diagonalThreats.contains$1(0, t2._as(threats[_i]).name))
+          return true;
+      return false;
+    },
+    attackedOnHorizontals$1(board) {
+      var t1, threats, t2, _i, _this = this,
+        horizontalThreats = A.HashSet_HashSet(type$.String);
+      horizontalThreats.add$1(0, "queen");
+      horizontalThreats.add$1(0, "rook");
+      t1 = type$.dynamic;
+      threats = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([_this.threatAlongImpulse$3(board, -1, 0), _this.threatAlongImpulse$3(board, 0, 1), _this.threatAlongImpulse$3(board, 1, 0), _this.threatAlongImpulse$3(board, 0, -1)], t1), true, t1);
+      for (t1 = threats.length, t2 = type$.ChessPiece, _i = 0; _i < t1; ++_i)
+        if (horizontalThreats.contains$1(0, t2._as(threats[_i]).name))
+          return true;
+      return false;
+    },
+    threatAlongImpulse$3(board, di, dj) {
+      var t1, threat, _this = this,
+        a = _this.i,
+        b = _this.j;
+      for (; true;) {
+        a += di;
+        b += dj;
+        t1 = 0 <= a && a < 8;
+        if (!B.JSBool_methods.$and(t1, 0 <= b && b < 8))
+          return A.EmptyPiece$(0, 0);
+        threat = _this.getPiece$3(board, a, b);
+        if (threat instanceof A.EmptyPiece)
+          continue;
+        if (threat.colour === _this.colour)
+          return A.EmptyPiece$(0, 0);
+        return threat;
+      }
+    },
+    attackedByPawn$1(board) {
+      var t3, t4, t5, _i, t6, t7, pawn, _this = this,
+        t1 = _this.colour,
+        t2 = _this.i,
+        row = t1 === "w" ? t2 - 1 : t2 + 1;
+      for (t2 = _this.j, t2 = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([t2 - 1, t2 + 1], type$.dynamic), true, type$.int), t3 = t2.length, t4 = 0 <= row, t5 = row < 8, _i = 0; _i < t3; ++_i) {
+        t6 = A._asInt(t2[_i]);
+        t7 = t4 && t5;
+        if (B.JSBool_methods.$and(t7, 0 <= t6 && t6 < 8)) {
+          pawn = _this.getPiece$3(board, row, t6);
+          if (pawn.colour !== t1 && pawn.name === "pawn")
+            return true;
+        }
+      }
+      return false;
+    },
+    attackedByKnight$1(board) {
+      var t1, t2, _i, a, t3, _i0, b, t4, t5, knight, _this = this,
+        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, 2, -2, -1], type$.dynamic), true, type$.int);
+      for (t1 = components.length, t2 = _this.colour, _i = 0; _i < t1; ++_i) {
+        a = components[_i];
+        for (t3 = J.getInterceptor$in(a), _i0 = 0; _i0 < t1; ++_i0) {
+          b = components[_i0];
+          if (t3.abs$0(a) + J.abs$0$in(b) === 3) {
+            t4 = _this.i;
+            if (typeof a !== "number")
+              return A.iae(a);
+            t5 = _this.j;
+            if (typeof b !== "number")
+              return A.iae(b);
+            knight = _this.getPiece$3(board, t4 + a, t5 + b);
+            if (knight.colour !== t2 && knight.name === "knight")
+              return true;
+          }
+        }
+      }
+      return false;
+    },
+    getPiece$3(board, i, j) {
+      var piece;
+      if (this.validCoords$2(i, j)) {
+        piece = board.getPiece$2(i, j);
+        return piece;
+      }
+      return A.EmptyPiece$(0, 0);
+    }
+  };
+  A.EmptyPiece.prototype = {};
+  A.NoMovement.prototype = {
+    move$2(board, piece) {
+    },
+    $isMovementStrategy: 1
+  };
+  A.PawnMovement.prototype = {
+    move$2(board, piece) {
+      if (piece.colour === "w") {
+        if (piece.canMove$3(board, piece.i - 1, piece.j))
+          if (!piece.hasMoved)
+            piece.canMove$3(board, piece.i - 2, piece.j);
+        piece.canCapture$3(board, piece.i - 1, piece.j + 1);
+        piece.canCapture$3(board, piece.i - 1, piece.j - 1);
+        return null;
+      } else {
+        if (piece.canMove$3(board, piece.i + 1, piece.j))
+          if (!piece.hasMoved)
+            piece.canMove$3(board, piece.i + 2, piece.j);
+        piece.canCapture$3(board, piece.i + 1, piece.j + 1);
+        piece.canCapture$3(board, piece.i + 1, piece.j - 1);
+        return null;
+      }
+    },
+    $isMovementStrategy: 1
+  };
+  A.KnightMovement.prototype = {
+    move$2(board, piece) {
+      var _i, a, t2, _i0, b, t3, t4,
+        t1 = type$.dynamic,
+        options = J.JSArray_JSArray$growable(0, t1),
+        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, 2, -2, -1], t1), true, type$.int);
+      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
+        a = components[_i];
+        for (t2 = J.getInterceptor$in(a), _i0 = 0; _i0 < t1; ++_i0) {
+          b = components[_i0];
+          if (t2.abs$0(a) + J.abs$0$in(b) === 3) {
+            t3 = piece.i;
+            if (typeof a !== "number")
+              return A.iae(a);
+            t4 = piece.j;
+            if (typeof b !== "number")
+              return A.iae(b);
+            piece.canMove$3(board, t3 + a, t4 + b);
+            piece.canCapture$3(board, piece.i + a, piece.j + b);
+          }
+        }
+      }
+      return options;
+    },
+    $isMovementStrategy: 1
+  };
+  A.BishopMovement.prototype = {
+    move$2(board, piece) {
+      var t1, _i, a, _i0,
+        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([1, -1], type$.dynamic), true, type$.int);
+      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
+        a = components[_i];
+        for (_i0 = 0; _i0 < t1; ++_i0)
+          this.exploreImpulse$4(piece, board, a, components[_i0]);
+      }
+    },
+    exploreImpulse$4(piece, board, di, dj) {
+      var i = piece.i,
+        j = piece.j;
+      for (; true;) {
+        i += di;
+        j += dj;
+        if (piece.canCapture$3(board, i, j) || !piece.canMove$3(board, i, j))
+          return;
+      }
+    },
+    $isMovementStrategy: 1
+  };
+  A.RookMovement.prototype = {
+    move$2(board, piece) {
+      var _this = this;
+      _this.exploreImpulse$4(piece, board, 0, 1);
+      _this.exploreImpulse$4(piece, board, 0, -1);
+      _this.exploreImpulse$4(piece, board, 1, 0);
+      _this.exploreImpulse$4(piece, board, -1, 0);
+    },
+    exploreImpulse$4(piece, board, di, dj) {
+      var i = piece.i,
+        j = piece.j;
+      for (; true;) {
+        i += di;
+        j += dj;
+        if (piece.canCapture$3(board, i, j) || !piece.canMove$3(board, i, j))
+          return;
+      }
+    },
+    $isMovementStrategy: 1
+  };
+  A.QueenMovement.prototype = {
+    move$2(board, piece) {
+      var t2, _i,
+        t1 = type$.dynamic,
+        pair = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([new A.RookMovement(), new A.BishopMovement()], t1), true, t1);
+      for (t1 = pair.length, t2 = type$.MovementStrategy, _i = 0; _i < t1; ++_i)
+        t2._as(pair[_i]).move$2(board, piece);
+    },
+    $isMovementStrategy: 1
+  };
+  A.KingMovement.prototype = {
+    move$2(board, piece) {
+      var t1, _i, a, _i0, b, t2, t3,
+        components = A.List_List$from(A.LinkedHashSet_LinkedHashSet$_literal([-1, 0, 1], type$.dynamic), true, type$.int);
+      for (t1 = components.length, _i = 0; _i < t1; ++_i) {
+        a = components[_i];
+        for (_i0 = 0; _i0 < t1; ++_i0) {
+          b = components[_i0];
+          t2 = piece.i;
+          if (typeof a !== "number")
+            return A.iae(a);
+          t3 = piece.j;
+          if (typeof b !== "number")
+            return A.iae(b);
+          piece.canCapture$3(board, t2 + a, t3 + b);
+          piece.canMove$3(board, piece.i + a, piece.j + b);
+        }
+      }
+    },
+    $isMovementStrategy: 1
+  };
+  A.ChessBoardView.prototype = {
+    displayBoard$1(boardstate) {
+      var t1, t2, t3, t4, _i, rowOfPieces, t5, row, t6, t7, t8, tile, t9, img, t10, element, subtype,
+        _s9_ = "piece-img";
+      type$.List_List_GamePiece._as(boardstate);
+      t1 = this.container;
+      t2 = J.getInterceptor$x(t1);
+      t2.get$children(t1).clear$0(0);
+      for (t3 = boardstate.length, t4 = type$.ImageElement, _i = 0; _i < boardstate.length; boardstate.length === t3 || (0, A.throwConcurrentModificationError)(boardstate), ++_i) {
+        rowOfPieces = boardstate[_i];
+        t5 = document;
+        row = t5.createElement("div");
+        t6 = J.getInterceptor$x(row);
+        t6.get$classes(row).add$1(0, "board-row");
+        for (t7 = B.JSArray_methods.get$iterator(rowOfPieces); t7.moveNext$0();) {
+          t8 = t7.get$current();
+          tile = this.createTile$1(t8);
+          t9 = t8 instanceof A.EmptyPiece;
+          if (!t9) {
+            img = t5.createElement("img");
+            J.get$classes$x(img).add$1(0, _s9_);
+            if (t4._is(img)) {
+              t10 = t8.__GamePiece_src_A;
+              t10 === $ && A.throwLateFieldNI("src");
+              B.ImageElement_methods.set$src(img, t10);
+            }
+            J.get$children$x(tile).add$1(0, img);
+          }
+          if (t8.threatened) {
+            element = t5.createElement("div");
+            t10 = J.getInterceptor$x(element);
+            t10.get$classes(element).add$1(0, "marker");
+            subtype = t9 ? "dot" : "circle";
+            t10.get$classes(element).add$1(0, subtype);
+            t9 = J.getInterceptor$x(tile);
+            t9.get$children(tile).add$1(0, element);
+            if (t8 instanceof A.ChessKing) {
+              t9.get$children(tile).clear$0(0);
+              img = t5.createElement("img");
+              J.get$classes$x(img).add$1(0, _s9_);
+              if (t4._is(img))
+                B.ImageElement_methods.set$src(img, "./assets/chess/king_" + t8.colour + "_check.png");
+              t9.get$children(tile).add$1(0, img);
+            }
+          }
+          t6.get$children(row).add$1(0, tile);
+        }
+        t2.get$children(t1).add$1(0, row);
+      }
+    },
+    createTile$1(piece) {
+      var tile = document.createElement("div"),
+        t1 = J.getInterceptor$x(tile);
+      t1.get$classes(tile).add$1(0, "chess-tile");
+      if (B.JSInt_methods.$mod(piece.i + piece.j, 2) !== 0)
+        t1.get$classes(tile).add$1(0, "dark");
+      t1.addEventListener$2(tile, "click", new A.ChessBoardView_createTile_closure(this, piece));
+      return tile;
+    },
+    $isGameView: 1
+  };
+  A.ChessBoardView_createTile_closure.prototype = {
+    call$1($event) {
+      var t1, t2, t3, counter, t4;
+      type$.Event._as($event);
+      t1 = this.$this.game;
+      t2 = this.piece;
+      t3 = t2.i;
+      t2 = t2.j;
+      counter = t1.turnCount;
+      t4 = t1.__ChessGame_logic_A;
+      t4 === $ && A.throwLateFieldNI("logic");
+      t4.submitMove$2(t3, t2);
+      t4 = t1.turnCount;
+      if (counter === t4) {
+        t1.turnCount = t4 + 1;
+        t1.__ChessGame_logic_A.submitMove$2(t3, t2);
+      }
+      t1.refreshView$0();
+    },
+    $signature: 0
+  };
   A.GamePiece.prototype = {
-    buildElement$0() {
+    getElement$0() {
       var t1,
         img = document.createElement("img");
       J.get$classes$x(img).add$1(0, "piece-img");
@@ -4419,54 +5453,10 @@
         t1 = this.__GamePiece_src_A;
         t1 === $ && A.throwLateFieldNI("src");
         B.ImageElement_methods.set$src(img, t1);
-        this.__GamePiece_element_A = img;
       }
+      return img;
     }
   };
-  A.ChessGame.prototype = {};
-  A.ChessBoard.prototype = {
-    insertTiles$0() {
-      var t1, t2, t3, t4, dark, i, rowList, row, t5, j, tile;
-      for (t1 = this.board, t2 = this.container, t3 = J.getInterceptor$x(t2), t4 = type$.JSArray_Element, dark = false, i = 0; i < 8; ++i) {
-        dark = !dark;
-        rowList = A._setArrayType(new Array(0), t4);
-        row = document.createElement("div");
-        t5 = J.getInterceptor$x(row);
-        t5.get$classes(row).add$1(0, "board-row");
-        for (j = 0; j < 8; ++j) {
-          dark = !dark;
-          tile = this.createTile$3(i, j, dark);
-          t5.get$children(row).add$1(0, tile);
-          B.JSArray_methods.add$1(rowList, tile);
-        }
-        B.JSArray_methods.add$1(t1, rowList);
-        t3.get$children(t2).add$1(0, row);
-      }
-    },
-    createTile$3(i, j, dark) {
-      var tile = document.createElement("div"),
-        t1 = J.getInterceptor$x(tile);
-      t1.get$classes(tile).add$1(0, "chess-tile");
-      if (dark)
-        t1.get$classes(tile).add$1(0, "dark");
-      t1.addEventListener$2(tile, "click", new A.ChessBoard_createTile_closure(this, i, j));
-      return tile;
-    }
-  };
-  A.ChessBoard_createTile_closure.prototype = {
-    call$1($event) {
-      type$.Event._as($event);
-      A.printString("Chess: move was made at board[" + this.i + "][" + this.j + "]");
-    },
-    $signature: 0
-  };
-  A.ChessPiece.prototype = {};
-  A.Pawn.prototype = {};
-  A.Knight.prototype = {};
-  A.Bishop.prototype = {};
-  A.Rook.prototype = {};
-  A.King.prototype = {};
-  A.Queen.prototype = {};
   A.PieceBox.prototype = {
     armButton$0() {
       B.ButtonElement_methods.addEventListener$2(this.button, "click", new A.PieceBox_armButton_closure(this));
@@ -4484,27 +5474,27 @@
       list = t1.classList;
       list.contains(t2).toString;
       list.add(t2);
-      t1.innerText = _this.factory.colour;
+      t1.innerText = (_this.factory.colour === "black" ? "Black" : "White") + "PieceFactory";
       _this.presentPieces$0();
     },
     presentPieces$0() {
-      var options, t3, _i, piece, t4, t5, _this = this,
+      var options, t3, _i, option, t4, t5, _this = this,
         t1 = _this.container,
         t2 = J.getInterceptor$x(t1);
       t2.get$children(t1).clear$0(0);
-      options = J.JSArray_JSArray$growable(0, type$.ChessPiece);
-      B.JSArray_methods.add$1(options, _this.factory.createPawn$0());
-      B.JSArray_methods.add$1(options, _this.factory.createBishop$0());
-      B.JSArray_methods.add$1(options, _this.factory.createKnight$0());
-      B.JSArray_methods.add$1(options, _this.factory.createRook$0());
-      B.JSArray_methods.add$1(options, _this.factory.createQueen$0());
-      B.JSArray_methods.add$1(options, _this.factory.createKing$0());
+      options = J.JSArray_JSArray$growable(0, type$.PieceOption);
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createPawn$0()));
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createBishop$0()));
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createKnight$0()));
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createRook$0()));
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createQueen$0()));
+      B.JSArray_methods.add$1(options, A.PieceOption$(_this, _this.factory.createKing$0()));
       t3 = options.length;
       _i = 0;
       for (; _i < options.length; options.length === t3 || (0, A.throwConcurrentModificationError)(options), ++_i) {
-        piece = options[_i];
+        option = options[_i];
         t4 = t2.get$children(t1);
-        t5 = piece.__GamePiece_element_A;
+        t5 = option.__PieceOption_element_A;
         t5 === $ && A.throwLateFieldNI("element");
         t4.add$1(0, t5);
       }
@@ -4514,6 +5504,29 @@
     call$1($event) {
       type$.Event._as($event);
       this.$this.switchColour$0();
+    },
+    $signature: 0
+  };
+  A.PieceOption.prototype = {
+    createElement$0(_) {
+      var element = document.createElement("div"),
+        t1 = J.getInterceptor$x(element);
+      t1.get$classes(element).add$1(0, "piece-option");
+      t1.addEventListener$2(element, "click", new A.PieceOption_createElement_closure(this));
+      t1.get$children(element).add$1(0, this.piece.getElement$0());
+      return element;
+    }
+  };
+  A.PieceOption_createElement_closure.prototype = {
+    call$1($event) {
+      var t1, t2;
+      type$.Event._as($event);
+      t1 = this.$this;
+      t2 = t1.parent;
+      t1 = t1.piece;
+      A.print("selection indicated! " + t1.name + " " + t1.colour);
+      t2.game.threatenAllWith$1(t1);
+      t2.presentPieces$0();
     },
     $signature: 0
   };
@@ -4534,7 +5547,7 @@
       _inherit = hunkHelpers.inherit,
       _inheritMany = hunkHelpers.inheritMany;
     _inherit(A.Object, null);
-    _inheritMany(A.Object, [A.JS_CONST, J.Interceptor, J.ArrayIterator, A.Error, A.ListIterator, A.Iterable, A.MappedIterator, A.WhereIterator, A.Closure, A.JSSyntaxRegExp, A.Rti, A._FunctionParameters, A._Type, A.SetBase, A._LinkedHashSetCell, A._LinkedHashSetIterator, A.ListBase, A._Exception, A.FormatException, A.Null, A.StringBuffer, A.ImmutableListMixin, A.FixedSizeListIterator, A.ChessPieceFactory, A.GamePiece, A.Game, A.GameBoard, A.PieceBox]);
+    _inheritMany(A.Object, [A.JS_CONST, J.Interceptor, J.ArrayIterator, A.Error, A.ListIterator, A.Iterable, A.MappedIterator, A.WhereIterator, A.Closure, A.JSSyntaxRegExp, A.Rti, A._FunctionParameters, A._Type, A.SetBase, A._HashSetIterator, A._LinkedHashSetCell, A._LinkedHashSetIterator, A.ListBase, A._Exception, A.FormatException, A.Null, A.StringBuffer, A.ImmutableListMixin, A.FixedSizeListIterator, A.ChessPieceFactory, A.GamePiece, A.ChequeredBoard, A.BoardWithPieces, A.ChessGame, A.ChessLogic, A.NoMovement, A.PawnMovement, A.KnightMovement, A.BishopMovement, A.RookMovement, A.QueenMovement, A.KingMovement, A.ChessBoardView, A.PieceBox, A.PieceOption]);
     _inheritMany(J.Interceptor, [J.JSBool, J.JSNull, J.JavaScriptObject, J.JSNumber, J.JSString]);
     _inheritMany(J.JavaScriptObject, [J.LegacyJavaScriptObject, J.JSArray, A.EventTarget, A.DomException, A.DomTokenList, A.Event, A._HtmlCollection_JavaScriptObject_ListMixin, A._NodeList_JavaScriptObject_ListMixin, A.__NamedNodeMap_JavaScriptObject_ListMixin]);
     _inheritMany(J.LegacyJavaScriptObject, [J.PlainJavaScriptObject, J.UnknownJavaScriptObject, J.JavaScriptFunction]);
@@ -4542,13 +5555,13 @@
     _inheritMany(J.JSNumber, [J.JSInt, J.JSNumNotInt]);
     _inheritMany(A.Error, [A.LateError, A._CyclicInitializationError, A.RuntimeError, A.AssertionError, A._Error, A.TypeError, A.ArgumentError, A.UnsupportedError, A.UnimplementedError, A.ConcurrentModificationError]);
     _inheritMany(A.Iterable, [A.MappedIterable, A.WhereIterable]);
-    _inheritMany(A.Closure, [A.Closure2Args, A.TearOffClosure, A.initHooks_closure, A.initHooks_closure1, A.CssClassSetImpl_add_closure, A.FilteredElementList__iterable_closure, A.FilteredElementList__iterable_closure0, A.ChessBoard_createTile_closure, A.PieceBox_armButton_closure]);
+    _inheritMany(A.Closure, [A.Closure2Args, A.TearOffClosure, A.initHooks_closure, A.initHooks_closure1, A.CssClassSetImpl_add_closure, A.FilteredElementList__iterable_closure, A.FilteredElementList__iterable_closure0, A.ChessBoardView_createTile_closure, A.PieceBox_armButton_closure, A.PieceOption_createElement_closure]);
     _inheritMany(A.TearOffClosure, [A.StaticClosure, A.BoundClosure]);
     _inherit(A._AssertionError, A.AssertionError);
     _inherit(A.initHooks_closure0, A.Closure2Args);
     _inherit(A._TypeError, A._Error);
     _inheritMany(A.SetBase, [A._SetBase, A.CssClassSetImpl]);
-    _inherit(A._LinkedHashSet, A._SetBase);
+    _inheritMany(A._SetBase, [A._HashSet, A._LinkedHashSet]);
     _inheritMany(A.ArgumentError, [A.RangeError, A.IndexError]);
     _inherit(A.Node, A.EventTarget);
     _inheritMany(A.Node, [A.Element, A.CharacterData]);
@@ -4564,15 +5577,14 @@
     _inheritMany(A.CssClassSetImpl, [A._ElementCssClassSet, A.AttributeClassSet]);
     _inheritMany(A.ChessPieceFactory, [A.BlackPieceFactory, A.WhitePieceFactory]);
     _inherit(A.ChessPiece, A.GamePiece);
-    _inheritMany(A.ChessPiece, [A.Pawn, A.Knight, A.Bishop, A.Rook, A.King, A.Queen]);
+    _inheritMany(A.ChessPiece, [A.Pawn, A.Knight, A.Bishop, A.Rook, A.King, A.Queen, A.ChessKing, A.EmptyPiece]);
     _inheritMany(A.Pawn, [A.BlackPawn, A.WhitePawn]);
     _inheritMany(A.Knight, [A.BlackKnight, A.WhiteKnight]);
     _inheritMany(A.Bishop, [A.BlackBishop, A.WhiteBishop]);
     _inheritMany(A.Rook, [A.BlackRook, A.WhiteRook]);
     _inheritMany(A.King, [A.BlackKing, A.WhiteKing]);
     _inheritMany(A.Queen, [A.BlackQueen, A.WhiteQueen]);
-    _inherit(A.ChessGame, A.Game);
-    _inherit(A.ChessBoard, A.GameBoard);
+    _inheritMany(A.BoardWithPieces, [A.BoardWithPawns, A.BoardWithBishops, A.BoardWithKnights, A.BoardWithRooks, A.BoardWithKings, A.BoardWithQueens]);
     _mixin(A._HtmlCollection_JavaScriptObject_ListMixin, A.ListBase);
     _mixin(A._HtmlCollection_JavaScriptObject_ListMixin_ImmutableListMixin, A.ImmutableListMixin);
     _mixin(A._NodeList_JavaScriptObject_ListMixin, A.ListBase);
@@ -4589,7 +5601,7 @@
     leafTags: null,
     arrayRti: Symbol("$ti")
   };
-  A._Universe_addRules(init.typeUniverse, JSON.parse('{"PlainJavaScriptObject":"LegacyJavaScriptObject","UnknownJavaScriptObject":"LegacyJavaScriptObject","JavaScriptFunction":"LegacyJavaScriptObject","AbortPaymentEvent":"Event","ExtendableEvent":"Event","AElement":"SvgElement","GraphicsElement":"SvgElement","AudioElement":"HtmlElement","MediaElement":"HtmlElement","HtmlDocument":"Node","Document":"Node","CDataSection":"CharacterData","Text":"CharacterData","MathMLElement":"Element","HtmlFormControlsCollection":"HtmlCollection","JSBool":{"bool":[],"TrustedGetRuntimeType":[]},"JSNull":{"TrustedGetRuntimeType":[]},"JSArray":{"List":["1"],"Iterable":["1"]},"JSUnmodifiableArray":{"JSArray":["1"],"List":["1"],"Iterable":["1"]},"ArrayIterator":{"Iterator":["1"]},"JSNumber":{"num":[]},"JSInt":{"int":[],"num":[],"TrustedGetRuntimeType":[]},"JSNumNotInt":{"num":[],"TrustedGetRuntimeType":[]},"JSString":{"String":[],"TrustedGetRuntimeType":[]},"ListIterator":{"Iterator":["1"]},"MappedIterable":{"Iterable":["2"]},"MappedIterator":{"Iterator":["2"]},"WhereIterable":{"Iterable":["1"]},"WhereIterator":{"Iterator":["1"]},"Closure":{"Function":[]},"Closure2Args":{"Function":[]},"TearOffClosure":{"Function":[]},"StaticClosure":{"Function":[]},"BoundClosure":{"Function":[]},"_LinkedHashSet":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"_LinkedHashSetIterator":{"Iterator":["1"]},"ListBase":{"List":["1"],"Iterable":["1"]},"SetBase":{"Set":["1"],"Iterable":["1"]},"_SetBase":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"List":{"Iterable":["1"]},"Set":{"Iterable":["1"]},"Element":{"Node":[]},"HtmlElement":{"Element":[],"Node":[]},"AnchorElement":{"Element":[],"Node":[]},"AreaElement":{"Element":[],"Node":[]},"ButtonElement":{"Element":[],"Node":[]},"CharacterData":{"Node":[]},"_ChildrenElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"FormElement":{"Element":[],"Node":[]},"HtmlCollection":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"ImageElement":{"Element":[],"Node":[]},"_ChildNodeListLazy":{"ListBase":["Node"],"List":["Node"],"Iterable":["Node"],"ListBase.E":"Node"},"NodeList":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"SelectElement":{"Element":[],"Node":[]},"_NamedNodeMap":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"_ElementCssClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FixedSizeListIterator":{"Iterator":["1"]},"CssClassSetImpl":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FilteredElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"AttributeClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"SvgElement":{"Element":[],"Node":[]},"BlackPieceFactory":{"ChessPieceFactory":[]},"WhitePieceFactory":{"ChessPieceFactory":[]},"BlackPawn":{"ChessPiece":[]},"BlackKnight":{"ChessPiece":[]},"BlackBishop":{"ChessPiece":[]},"BlackRook":{"ChessPiece":[]},"BlackKing":{"ChessPiece":[]},"BlackQueen":{"ChessPiece":[]},"ChessGame":{"Game":[]},"ChessBoard":{"GameBoard":[]},"Pawn":{"ChessPiece":[]},"Knight":{"ChessPiece":[]},"Bishop":{"ChessPiece":[]},"Rook":{"ChessPiece":[]},"King":{"ChessPiece":[]},"Queen":{"ChessPiece":[]},"WhitePawn":{"ChessPiece":[]},"WhiteKnight":{"ChessPiece":[]},"WhiteBishop":{"ChessPiece":[]},"WhiteRook":{"ChessPiece":[]},"WhiteKing":{"ChessPiece":[]},"WhiteQueen":{"ChessPiece":[]}}'));
+  A._Universe_addRules(init.typeUniverse, JSON.parse('{"PlainJavaScriptObject":"LegacyJavaScriptObject","UnknownJavaScriptObject":"LegacyJavaScriptObject","JavaScriptFunction":"LegacyJavaScriptObject","AbortPaymentEvent":"Event","ExtendableEvent":"Event","AElement":"SvgElement","GraphicsElement":"SvgElement","AudioElement":"HtmlElement","MediaElement":"HtmlElement","HtmlDocument":"Node","Document":"Node","CDataSection":"CharacterData","Text":"CharacterData","MathMLElement":"Element","HtmlFormControlsCollection":"HtmlCollection","JSBool":{"bool":[],"TrustedGetRuntimeType":[]},"JSNull":{"TrustedGetRuntimeType":[]},"JSArray":{"List":["1"],"Iterable":["1"]},"JSUnmodifiableArray":{"JSArray":["1"],"List":["1"],"Iterable":["1"]},"ArrayIterator":{"Iterator":["1"]},"JSNumber":{"num":[]},"JSInt":{"int":[],"num":[],"TrustedGetRuntimeType":[]},"JSNumNotInt":{"num":[],"TrustedGetRuntimeType":[]},"JSString":{"String":[],"TrustedGetRuntimeType":[]},"ListIterator":{"Iterator":["1"]},"MappedIterable":{"Iterable":["2"]},"MappedIterator":{"Iterator":["2"]},"WhereIterable":{"Iterable":["1"]},"WhereIterator":{"Iterator":["1"]},"Closure":{"Function":[]},"Closure2Args":{"Function":[]},"TearOffClosure":{"Function":[]},"StaticClosure":{"Function":[]},"BoundClosure":{"Function":[]},"_HashSet":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"_HashSetIterator":{"Iterator":["1"]},"_LinkedHashSet":{"SetBase":["1"],"LinkedHashSet":["1"],"Set":["1"],"Iterable":["1"]},"_LinkedHashSetIterator":{"Iterator":["1"]},"ListBase":{"List":["1"],"Iterable":["1"]},"SetBase":{"Set":["1"],"Iterable":["1"]},"_SetBase":{"SetBase":["1"],"Set":["1"],"Iterable":["1"]},"int":{"num":[]},"List":{"Iterable":["1"]},"Set":{"Iterable":["1"]},"Element":{"Node":[]},"HtmlElement":{"Element":[],"Node":[]},"AnchorElement":{"Element":[],"Node":[]},"AreaElement":{"Element":[],"Node":[]},"ButtonElement":{"Element":[],"Node":[]},"CharacterData":{"Node":[]},"_ChildrenElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"FormElement":{"Element":[],"Node":[]},"HtmlCollection":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"ImageElement":{"Element":[],"Node":[]},"_ChildNodeListLazy":{"ListBase":["Node"],"List":["Node"],"Iterable":["Node"],"ListBase.E":"Node"},"NodeList":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"SelectElement":{"Element":[],"Node":[]},"_NamedNodeMap":{"ListBase":["Node"],"ImmutableListMixin":["Node"],"List":["Node"],"JavaScriptIndexingBehavior":["Node"],"Iterable":["Node"],"ListBase.E":"Node","ImmutableListMixin.E":"Node"},"_ElementCssClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FixedSizeListIterator":{"Iterator":["1"]},"CssClassSetImpl":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"FilteredElementList":{"ListBase":["Element"],"List":["Element"],"Iterable":["Element"],"ListBase.E":"Element"},"AttributeClassSet":{"SetBase":["String"],"Set":["String"],"Iterable":["String"]},"SvgElement":{"Element":[],"Node":[]},"BlackPieceFactory":{"ChessPieceFactory":[]},"WhitePieceFactory":{"ChessPieceFactory":[]},"Pawn":{"ChessPiece":[],"GamePiece":[]},"Knight":{"ChessPiece":[],"GamePiece":[]},"Bishop":{"ChessPiece":[],"GamePiece":[]},"Rook":{"ChessPiece":[],"GamePiece":[]},"King":{"ChessPiece":[],"GamePiece":[]},"Queen":{"ChessPiece":[],"GamePiece":[]},"BlackPawn":{"ChessPiece":[],"GamePiece":[]},"BlackKnight":{"ChessPiece":[],"GamePiece":[]},"BlackBishop":{"ChessPiece":[],"GamePiece":[]},"BlackRook":{"ChessPiece":[],"GamePiece":[]},"BlackKing":{"ChessPiece":[],"GamePiece":[]},"BlackQueen":{"ChessPiece":[],"GamePiece":[]},"ChequeredBoard":{"GameBoard":[]},"BoardWithPieces":{"GameBoard":[]},"BoardWithPawns":{"GameBoard":[]},"BoardWithBishops":{"GameBoard":[]},"BoardWithKnights":{"GameBoard":[]},"BoardWithRooks":{"GameBoard":[]},"BoardWithKings":{"GameBoard":[]},"BoardWithQueens":{"GameBoard":[]},"ChessGame":{"Game":[]},"ChessPiece":{"GamePiece":[]},"ChessKing":{"ChessPiece":[],"GamePiece":[]},"EmptyPiece":{"ChessPiece":[],"GamePiece":[]},"NoMovement":{"MovementStrategy":[]},"PawnMovement":{"MovementStrategy":[]},"KnightMovement":{"MovementStrategy":[]},"BishopMovement":{"MovementStrategy":[]},"RookMovement":{"MovementStrategy":[]},"QueenMovement":{"MovementStrategy":[]},"KingMovement":{"MovementStrategy":[]},"ChessBoardView":{"GameView":[]},"WhitePawn":{"ChessPiece":[],"GamePiece":[]},"WhiteKnight":{"ChessPiece":[],"GamePiece":[]},"WhiteBishop":{"ChessPiece":[],"GamePiece":[]},"WhiteRook":{"ChessPiece":[],"GamePiece":[]},"WhiteKing":{"ChessPiece":[],"GamePiece":[]},"WhiteQueen":{"ChessPiece":[],"GamePiece":[]}}'));
   A._Universe_addErasedTypes(init.typeUniverse, JSON.parse('{"_SetBase":1}'));
   var type$ = (function rtii() {
     var findType = A.findType;
@@ -4601,16 +5613,20 @@
       Function: findType("Function"),
       ImageElement: findType("ImageElement"),
       Iterable_dynamic: findType("Iterable<@>"),
-      JSArray_Element: findType("JSArray<Element>"),
+      JSArray_ChessPiece: findType("JSArray<ChessPiece>"),
       JSArray_String: findType("JSArray<String>"),
       JSArray_dynamic: findType("JSArray<@>"),
       JSNull: findType("JSNull"),
       JavaScriptFunction: findType("JavaScriptFunction"),
       JavaScriptIndexingBehavior_dynamic: findType("JavaScriptIndexingBehavior<@>"),
-      List_Element: findType("List<Element>"),
+      List_GamePiece: findType("List<GamePiece>"),
+      List_List_GamePiece: findType("List<List<GamePiece>>"),
+      List_bool: findType("List<bool>"),
+      MovementStrategy: findType("MovementStrategy"),
       Node: findType("Node"),
       Null: findType("Null"),
       Object: findType("Object"),
+      PieceOption: findType("PieceOption"),
       Record: findType("Record"),
       Set_String: findType("Set<String>"),
       String: findType("String"),
@@ -4618,6 +5634,7 @@
       UnknownJavaScriptObject: findType("UnknownJavaScriptObject"),
       bool: findType("bool"),
       double: findType("double"),
+      dynamic: findType("@"),
       dynamic_Function_Set_String: findType("@(Set<String>)"),
       int: findType("int"),
       legacy_Never: findType("0&*"),
@@ -4634,6 +5651,8 @@
     B.ImageElement_methods = A.ImageElement.prototype;
     B.Interceptor_methods = J.Interceptor.prototype;
     B.JSArray_methods = J.JSArray.prototype;
+    B.JSBool_methods = J.JSBool.prototype;
+    B.JSInt_methods = J.JSInt.prototype;
     B.JSString_methods = J.JSString.prototype;
     B.JavaScriptFunction_methods = J.JavaScriptFunction.prototype;
     B.JavaScriptObject_methods = J.JavaScriptObject.prototype;
