+++ docs/abstract-factory/script.js.deps
@@ -1,11 +1,15 @@
 file:///C:/tools/dart-sdk/lib/_internal/dart2js_platform.dill
 file:///C:/tools/dart-sdk/lib/libraries.json
-file:///D:/GitHub/patterns/docs/abstract-factory/abstractfactory.dart
-file:///D:/GitHub/patterns/docs/abstract-factory/blackpieces.dart
-file:///D:/GitHub/patterns/docs/abstract-factory/boardgames.dart
-file:///D:/GitHub/patterns/docs/abstract-factory/chess.dart
-file:///D:/GitHub/patterns/docs/abstract-factory/script.dart
-file:///D:/GitHub/patterns/docs/abstract-factory/whitepieces.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/abstractfactory.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/blackpieces.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/chess/chess_board.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/chess/chess_game.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/chess/chess_logic.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/chess/chess_pieces.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/chess/chess_view.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/game.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/script.dart
+file:///D:/GitHub/patterns/docs/abstract-factory-2/whitepieces.dart
 org-dartlang-sdk:///lib/_http/crypto.dart
 org-dartlang-sdk:///lib/_http/embedder_config.dart
 org-dartlang-sdk:///lib/_http/http.dart
