+++ docs/abstract-factory/style.css
@@ -1,32 +1,64 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto');
 
+
 * {
     box-sizing: border-box;
 }
 
+
 body {
     font-family: 'Roboto', sans-serif;
-    display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     height: 100vh;
-    overflow: hidden;
     margin: 0;
 }
 
-.hidden {
-    display: none;
-}  
+#site-header {
+    background-color: #333;
+    height: 5.5vh;
+    width: 100vw;
+    justify-content: space-between;
+    color: white;
+    display: flex;
+}
 
-#choices-container {
+#site-header a {
+    text-decoration: none;
+    color: white;
+    margin: 1vw;
+}
+
+#site-header a:hover {
+    color: #ddd;
+}
+
+#site-header .left-box {
+    margin: 0 1vh;
+    display: flex;
+}
+
+#site-header .right-box {
+    margin: 0 1vh;
+    display: flex;
+}
+
+#site-header .anchor-tags {
+    margin: auto;
     display: flex;
     flex-direction: row;
 }
 
-#game-container {
+#demo-area {
+    margin-top: 5vh;
+    height: 80vh;
+    width: 100%;
     display: flex;
-    flex-direction: column;
+    flex-wrap: wrap;
+    align-items: center;
+    justify-content: center;
+    flex-direction: row;
 }
 
 .board-row {
@@ -35,62 +67,9 @@ body {
 }
 
 
+/* Game Pieces */
+
 .piece-img {
     width: 100%;
 }
 
-.colour-switcher {
-    padding: 10px 10px;
-    margin: 10px 0px;
-    font-size: 1rem;
-    border-style: none;
-    border-radius: 8px;
-    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-}
-
-.black-btn {
-    background-color: #333;
-    color: white;
-}
-
-.white-btn {
-    background-color: white;
-    color: #111;
-}
-
-.colour-switcher:hover {
-    cursor: pointer;
-}
-
-
-.bookshelf {
-    display: flex;
-    flex-direction: row;
-}
-
-.totempole {
-    display: flex;
-    flex-direction: column;
-}
-
-#piece-box {
-    padding: 10px 24px;
-}
-
-#side-tray {
-    background-color: #eee;
-    padding: 24px;
-    width: 200px;
-    margin: 0px 10px;
-}
-
-.chess-tile {
-    background-color: #e0c46e;
-    margin: 0px;
-    width: 100px;
-    height: 100px;
-}
-
-.chess-tile.dark {
-    background-color: #735e47;
-}
