+++ docs/abstract-factory/whitepieces.dart
@@ -1,4 +1,4 @@
-import 'chess.dart';
+import "abstractfactory.dart";
 
 class WhitePawn extends Pawn {
   WhitePawn() : super("w") {}
